% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{admin_delete_announcements}
\alias{admin_delete_announcements}
\title{Delete an announcement}
\usage{
admin_delete_announcements(id)
}
\arguments{
\item{id}{integer required. The ID of this announcement}
}
\value{
An empty HTTP response
}
\description{
Delete an announcement
}
