% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{admin_get_announcements}
\alias{admin_get_announcements}
\title{Get a particular announcement}
\usage{
admin_get_announcements(id)
}
\arguments{
\item{id}{integer required. The ID of this announcement}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this announcement}
\item{subject}{string, The subject of this announcement.}
\item{body}{string, The body of this announcement.}
\item{releasedAt}{string, The date and time this announcement was released.}
\item{createdAt}{string, }
\item{updatedAt}{string, }
}
\description{
Get a particular announcement
}
