% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{admin_list_organizations}
\alias{admin_list_organizations}
\title{List organizations}
\usage{
admin_list_organizations(status = NULL, org_type = NULL)
}
\arguments{
\item{status}{array optional. The status of the organization (active/trial/inactive).}

\item{org_type}{array optional. The organization type (platform/ads/survey_vendor/other).}
}
\value{
An array containing the following fields:
\item{id}{integer, The ID of this organization.}
\item{name}{string, The name of this organization.}
\item{slug}{string, The slug of this organization.}
\item{accountManagerId}{integer, The user ID of the Account Manager.}
\item{csSpecialistId}{integer, The user ID of the Client Success Specialist.}
\item{status}{string, The status of the organization (active/trial/inactive).}
\item{orgType}{string, The organization type (platform/ads/survey_vendor/other).}
\item{customBranding}{string, The custom branding settings.}
\item{contractSize}{integer, The monthly contract size.}
\item{maxAnalystUsers}{integer, The max number of full platform users for the org.}
\item{maxReportUsers}{integer, The max number of report-only platform users for the org.}
\item{vertical}{string, The business vertical that the organization belongs to.}
\item{csMetadata}{string, Additional metadata about the organization in JSON format.}
\item{removeFooterInEmails}{boolean, If true, emails sent by platform will not include Civis text.}
\item{salesforceAccountId}{string, The SalesForce Account ID for this organization.}
\item{advancedSettings}{list, A list containing the following elements: 
\itemize{
\item dedicatedDjPoolEnabled boolean, If true, the Organization has a dedicated delayed jobs pool. Defaults to false.
}}
\item{tableauRefreshHistory}{array, The number of tableau refreshes used this month.}
}
\description{
List organizations
}
