% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{admin_patch_announcements}
\alias{admin_patch_announcements}
\title{Edit an announcement}
\usage{
admin_patch_announcements(id, subject = NULL, body = NULL, released_at = NULL)
}
\arguments{
\item{id}{integer required. The ID of this announcement}

\item{subject}{string optional. The subject of this announcement.}

\item{body}{string optional. The body of this announcement.}

\item{released_at}{string optional. The date and time this announcement was released.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this announcement}
\item{subject}{string, The subject of this announcement.}
\item{body}{string, The body of this announcement.}
\item{releasedAt}{string, The date and time this announcement was released.}
\item{createdAt}{string, }
\item{updatedAt}{string, }
}
\description{
Edit an announcement
}
