% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{admin_patch_themes}
\alias{admin_patch_themes}
\title{Edit a theme}
\usage{
admin_patch_themes(
  id,
  name = NULL,
  organization_ids = NULL,
  settings_json = NULL,
  logo_file_id = NULL
)
}
\arguments{
\item{id}{integer required. The ID of this theme.}

\item{name}{string optional. The name of this theme.}

\item{organization_ids}{array optional. List of organization ID's allowed to see this theme.}

\item{settings_json}{string optional. The JSON-encoded theme configuration.}

\item{logo_file_id}{string optional. The ID of the logo image file.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this theme.}
\item{name}{string, The name of this theme.}
\item{organizationIds}{array, List of organization ID's allowed to use this theme.}
\item{settings}{string, The theme configuration object.}
\item{logoFile}{list, A list containing the following elements: 
\itemize{
\item id integer, The ID of the logo image file.
\item downloadUrl string, The URL of the logo image file.
}}
\item{createdAt}{string, }
\item{updatedAt}{string, }
}
\description{
Edit a theme
}
