% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{admin_post_dump_memory}
\alias{admin_post_dump_memory}
\title{Dump memory of a web server into a file in large_tmp}
\usage{
admin_post_dump_memory(host_name = NULL)
}
\arguments{
\item{host_name}{string optional. If set, prevents the heap dump from happening unless the request is being handled by that host.}
}
\value{
A list containing the following elements:
\item{hostName}{string, The host that responded to the request.}
\item{filename}{string, The heap dump file.}
}
\description{
Dump memory of a web server into a file in large_tmp
}
