% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{aliases_delete}
\alias{aliases_delete}
\title{Delete an alias}
\usage{
aliases_delete(id)
}
\arguments{
\item{id}{integer required. The id of the Alias object.}
}
\value{
An empty HTTP response
}
\description{
Delete an alias
}
