% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{aliases_get}
\alias{aliases_get}
\title{Get an Alias}
\usage{
aliases_get(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
A list containing the following elements:
\item{id}{integer, The id of the Alias object.}
\item{objectId}{integer, The id of the object}
\item{objectType}{string, The type of the object. Valid types include: model, cass_ncoa, container_script, gdoc_export, geocode, media_optimizer, python_script, r_script, salesforce_export, javascript_script, sql_script, project, notebook, workflow, template_script, template_report, service, report, tableau and service_report.}
\item{alias}{string, The alias of the object}
\item{userId}{integer, The id of the user who created the alias}
\item{displayName}{string, The display name of the Alias object. Defaults to object name if not provided.}
}
\description{
Get an Alias
}
