% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{aliases_get_object_type}
\alias{aliases_get_object_type}
\title{Get details about an alias within an FCO type}
\usage{
aliases_get_object_type(object_type, alias)
}
\arguments{
\item{object_type}{string required. The type of the object. Valid types include: cass_ncoa, container_script, geocode, identity_resolution, dbt_script, python_script, r_script, javascript_script, sql_script, project, notebook, workflow, template_script, template_report, service, report, tableau and service_report.}

\item{alias}{string required. The alias of the object}
}
\value{
A list containing the following elements:
\item{id}{integer, The id of the Alias object.}
\item{objectId}{integer, The id of the object}
\item{objectType}{string, The type of the object. Valid types include: cass_ncoa, container_script, geocode, identity_resolution, dbt_script, python_script, r_script, javascript_script, sql_script, project, notebook, workflow, template_script, template_report, service, report, tableau and service_report.}
\item{alias}{string, The alias of the object}
\item{userId}{integer, The id of the user who created the alias}
\item{displayName}{string, The display name of the Alias object. Defaults to object name if not provided.}
}
\description{
Get details about an alias within an FCO type
}
