% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{aliases_put}
\alias{aliases_put}
\title{Replace all attributes of this Alias}
\usage{
aliases_put(id, object_id, object_type, alias, display_name = NULL)
}
\arguments{
\item{id}{integer required. The id of the Alias object.}

\item{object_id}{integer required. The id of the object}

\item{object_type}{string required. The type of the object. Valid types include: bsd_cons_export, model, cass_ncoa, container_script, gdoc_export, geocode, match, media_optimizer, python_script, ratecard_import, r_script, salesforce_export, javascript_script, sql_script, table_archival, table_unarchival, van_list_import, van_list_export, van_person_export, van_score_export, van_sync, project, notebook, workflow, template_script, template_report, service, report, tableau and service_report.}

\item{alias}{string required. The alias of the object}

\item{display_name}{string optional. The display name of the Alias object. Defaults to object name if not provided.}
}
\value{
A list containing the following elements:
\item{id}{integer, The id of the Alias object.}
\item{objectId}{integer, The id of the object}
\item{objectType}{string, The type of the object. Valid types include: bsd_cons_export, model, cass_ncoa, container_script, gdoc_export, geocode, match, media_optimizer, python_script, ratecard_import, r_script, salesforce_export, javascript_script, sql_script, table_archival, table_unarchival, van_list_import, van_list_export, van_person_export, van_score_export, van_sync, project, notebook, workflow, template_script, template_report, service, report, tableau and service_report.}
\item{alias}{string, The alias of the object}
\item{userId}{integer, The id of the user who created the alias}
\item{displayName}{string, The display name of the Alias object. Defaults to object name if not provided.}
}
\description{
Replace all attributes of this Alias
}
