% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{announcements_list}
\alias{announcements_list}
\title{List announcements}
\usage{
announcements_list(
  limit = NULL,
  page_num = NULL,
  order = NULL,
  order_dir = NULL
)
}
\arguments{
\item{limit}{integer optional. Number of results to return. Defaults to 10. Maximum allowed is 50.}

\item{page_num}{integer optional. Page number of the results to return. Defaults to the first page, 1.}

\item{order}{string optional. The field on which to order the result set. Defaults to released_at. Must be one of: released_at.}

\item{order_dir}{string optional. Direction in which to sort, either asc (ascending) or desc (descending) defaulting to desc.}
}
\value{
An array containing the following fields:
\item{id}{integer, The ID of this announcement}
\item{subject}{string, The subject of this announcement.}
\item{body}{string, The body of this announcement.}
\item{releasedAt}{string, The date and time this announcement was released.}
\item{createdAt}{string, }
\item{updatedAt}{string, }
}
\description{
List announcements
}
