% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{apps_delete_instances_shares_groups}
\alias{apps_delete_instances_shares_groups}
\title{Revoke the permissions a group has on this object}
\usage{
apps_delete_instances_shares_groups(id, group_id, slug)
}
\arguments{
\item{id}{integer required. ID of the resource to be revoked}

\item{group_id}{integer required. ID of the group}

\item{slug}{string required. The slug for the application.}
}
\value{
An empty HTTP response
}
\description{
Revoke the permissions a group has on this object
}
