% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{apps_delete_releases_shares_groups}
\alias{apps_delete_releases_shares_groups}
\title{Revoke the permissions a group has on this object}
\usage{
apps_delete_releases_shares_groups(slug, id, group_id)
}
\arguments{
\item{slug}{string required. The slug for the application.}

\item{id}{integer required. The ID of the resource that is shared.}

\item{group_id}{integer required. The ID of the group.}
}
\value{
An empty HTTP response
}
\description{
Revoke the permissions a group has on this object
}
