% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{apps_delete_releases_shares_users}
\alias{apps_delete_releases_shares_users}
\title{Revoke the permissions a user has on this object}
\usage{
apps_delete_releases_shares_users(slug, id, user_id)
}
\arguments{
\item{slug}{string required. The slug for the application.}

\item{id}{integer required. The ID of the resource that is shared.}

\item{user_id}{integer required. The ID of the user.}
}
\value{
An empty HTTP response
}
\description{
Revoke the permissions a user has on this object
}
