% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{apps_get}
\alias{apps_get}
\title{List details of a Decision Application}
\usage{
apps_get(slug)
}
\arguments{
\item{slug}{string required. The slug for the application.}
}
\value{
A list containing the following elements:
\item{slug}{string, The slug for the application.}
\item{id}{integer, The unique id of the application.}
\item{instanceName}{string, A word that describes an instance of this app.}
\item{name}{string, The name of the application.}
\item{currentRelease}{list, A list containing the following elements: 
\itemize{
\item id integer, The unique id of the release.
\item appId integer, The id of the app the release belongs to.
\item reportTemplateId integer, ID of the report template for this release.
\item resources object, A hash of resources associated with this release.
}}
\item{features}{list, App features.}
}
\description{
List details of a Decision Application
}
