% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{apps_get_releases}
\alias{apps_get_releases}
\title{Return a given app release}
\usage{
apps_get_releases(id, slug)
}
\arguments{
\item{id}{integer required. The unique id of the release.}

\item{slug}{string required. The slug for the application.}
}
\value{
A list containing the following elements:
\item{id}{integer, The unique id of the release.}
\item{appId}{integer, The id of the app the release belongs to.}
\item{reportTemplateId}{integer, ID of the report template for this release.}
\item{resources}{list, A hash of resources associated with this release.}
\item{archived}{string, The archival status of the requested item(s).}
}
\description{
Return a given app release
}
