% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{apps_list}
\alias{apps_list}
\title{List apps}
\usage{
apps_list()
}
\value{
An array containing the following fields:
\item{slug}{string, The slug for the application.}
\item{id}{integer, The unique id of the application.}
\item{instanceName}{string, A word that describes an instance of this app.}
\item{name}{string, The name of the application.}
}
\description{
List apps
}
