% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{apps_list_releases_shares}
\alias{apps_list_releases_shares}
\title{List users and groups permissioned on this object}
\usage{
apps_list_releases_shares(id, slug)
}
\arguments{
\item{id}{integer required. The unique id of the release.}

\item{slug}{string required. The slug for the application.}
}
\value{
An array containing the following fields:
\item{readers}{object, A list containing the following elements: 
\itemize{
\item users array, 
\item groups array, 
}}
\item{writers}{object, A list containing the following elements: 
\itemize{
\item users array, 
\item groups array, 
}}
\item{owners}{object, A list containing the following elements: 
\itemize{
\item users array, 
\item groups array, 
}}
\item{totalUserShares}{integer, For owners, the number of total users shared. For writers and readers, the number of visible users shared.}
\item{totalGroupShares}{integer, For owners, the number of total groups shared. For writers and readers, the number of visible groups shared.}
}
\description{
List users and groups permissioned on this object
}
