% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{apps_patch_releases}
\alias{apps_patch_releases}
\title{Update an existing Decision Application release}
\usage{
apps_patch_releases(slug, id, report_template_id = NULL, resources = NULL)
}
\arguments{
\item{slug}{string required. The slug for the application.}

\item{id}{integer required. The unique id of the release.}

\item{report_template_id}{integer optional. ID of the report template for this release.}

\item{resources}{list optional. A hash of resources associated with this release.}
}
\value{
A list containing the following elements:
\item{id}{integer, The unique id of the release.}
\item{appId}{integer, The id of the app the release belongs to.}
\item{reportTemplateId}{integer, ID of the report template for this release.}
\item{resources}{list, A hash of resources associated with this release.}
\item{archived}{string, The archival status of the requested item(s).}
}
\description{
Update an existing Decision Application release
}
