% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{apps_post_releases}
\alias{apps_post_releases}
\title{Create a new Decision Application release}
\usage{
apps_post_releases(slug, report_template_id, resources)
}
\arguments{
\item{slug}{string required. The slug for the application.}

\item{report_template_id}{integer required. ID of the report template for this release.}

\item{resources}{object required. A hash of resources associated with this release.}
}
\value{
A list containing the following elements:
\item{id}{integer, The unique id of the release.}
\item{appId}{integer, The id of the app the release belongs to.}
\item{reportTemplateId}{integer, ID of the report template for this release.}
\item{resources}{object, A hash of resources associated with this release.}
}
\description{
Create a new Decision Application release
}
