% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{apps_put_features}
\alias{apps_put_features}
\title{Update the Decision Application features for a given organization}
\usage{
apps_put_features(slug, org, features)
}
\arguments{
\item{slug}{string required. The slug for the application.}

\item{org}{string required. Organization.}

\item{features}{object required. App features.}
}
\value{
A list containing the following elements:
\item{slug}{string, The slug for the application.}
\item{id}{integer, The unique id of the application.}
\item{instanceName}{string, A word that describes an instance of this app.}
\item{name}{string, The name of the application.}
\item{currentRelease}{object, A list containing the following elements: 
\itemize{
\item id integer, The unique id of the release.
\item appId integer, The id of the app the release belongs to.
\item reportTemplateId integer, ID of the report template for this release.
\item resources object, A hash of resources associated with this release.
}}
\item{features}{object, App features.}
}
\description{
Update the Decision Application features for a given organization
}
