% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{apps_put_releases_archive}
\alias{apps_put_releases_archive}
\title{Update the archive status of this object}
\usage{
apps_put_releases_archive(id, slug, status)
}
\arguments{
\item{id}{integer required. The ID of the object.}

\item{slug}{string required. The slug for the application.}

\item{status}{boolean required. The desired archived status of the object.}
}
\value{
A list containing the following elements:
\item{id}{integer, The unique id of the release.}
\item{appId}{integer, The id of the app the release belongs to.}
\item{reportTemplateId}{integer, ID of the report template for this release.}
\item{resources}{object, A hash of resources associated with this release.}
}
\description{
Update the archive status of this object
}
