% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{bocce_clusters_get}
\alias{bocce_clusters_get}
\title{Describe a Bocce Cluster}
\usage{
bocce_clusters_get(id)
}
\arguments{
\item{id}{integer required. The ID of this cluster.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this cluster.}
\item{instanceType}{string, The EC2 instance types in this cluster.}
\item{minInstances}{integer, The minimum number of instances in this cluster.}
\item{maxInstances}{integer, The maximum number of instances in this cluster.}
\item{instances}{integer, The number of instances currently in this cluster.}
\item{instanceMaxMemory}{integer, The amount of memory available to a single instance.}
\item{instanceMaxCpu}{integer, The number of processor shares available to a single instance.}
\item{instanceMaxDiskSpace}{number, The amount of memory available to a single instance.}
\item{region}{string, The AWS region that this cluster is in.}
\item{activeJobsCount}{integer, The number of jobs currently being run in the cluster.}
\item{queuedJobsCount}{integer, The number of jobs currently waiting to be run on the cluster.}
\item{organizationIds}{array, The ids of this cluster's organizations.}
\item{organizationSlugs}{array, The slugs of this cluster's organizations.}
\item{singleOrg}{boolean, Whether this cluster is allowed to be used by multiple orgs.}
\item{desiredInstances}{integer, The number of instances this cluster should attempt to have.}
\item{remoteDesiredInstances}{integer, The number of instances this cluster should attempt to have. Returns -1 if AWS query fails}
\item{asgName}{string, The name of the AutoScaling Group for the Bocce cluster.}
\item{queueUrl}{string, The SQS queue URL where jobs are posted to.}
\item{createdAt}{string, The date and time of cluster creation.}
\item{updatedAt}{string, The date and time of the cluster's last update.}
}
\description{
Describe a Bocce Cluster
}
