% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{bocce_clusters_list_active_jobs}
\alias{bocce_clusters_list_active_jobs}
\title{List Active Jobs for a Bocce Cluster}
\usage{
bocce_clusters_list_active_jobs(id)
}
\arguments{
\item{id}{integer required. The ID of this cluster.}
}
\value{
A list containing the following elements:
\item{id}{integer, }
\item{name}{string, }
\item{type}{string, }
\item{state}{string, Whether the job is idle, queued, running, cancelled, or failed.}
\item{createdAt}{string, }
\item{updatedAt}{string, }
\item{runs}{array, An array containing the following fields: 
\itemize{
\item id integer, 
\item state string, 
\item createdAt string, The time that the run was queued.
\item startedAt string, The time that the run started.
\item finishedAt string, The time that the run completed.
\item error string, The error message for this run, if present.
}}
\item{lastRun}{object, A list containing the following elements: 
\itemize{
\item id integer, 
\item state string, 
\item createdAt string, The time that the run was queued.
\item startedAt string, The time that the run started.
\item finishedAt string, The time that the run completed.
\item error string, The error message for this run, if present.
}}
\item{hidden}{boolean, The hidden status of the object.}
\item{requiredCpu}{integer, The CPU shares required by the script.}
\item{requiredDiskSpace}{integer, The disk space in GB required by the script.}
\item{requiredMemory}{integer, The memory in MB required by the script.}
\item{author}{object, A list containing the following elements: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
\item{runningAs}{object, A list containing the following elements: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
}
\description{
List Active Jobs for a Bocce Cluster
}
