% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_future.R
\name{cancel}
\alias{cancel}
\title{Cancel the evaluation of a CivisFuture.}
\usage{
cancel(future, ...)
}
\arguments{
\item{future}{CivisFuture object.}

\item{...}{unused for CivisFuture.}
}
\description{
Cancel the evaluation of a CivisFuture.
}
