% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{channels_list}
\alias{channels_list}
\title{List available channels and provide keys for accessing them}
\usage{
channels_list()
}
\value{
An array containing the following fields:
\item{origin}{string, The domain for Civis's notifications.}
\item{subscribeKey}{string, The identification key for Civis's notifications.}
\item{cipherKey}{string, The cipher key for Civis's notifications.}
\item{authKey}{string, The authorization key for accessing channels. This is specific to you and should be kept secret.}
\item{channels}{array, An array containing the following fields: 
\itemize{
\item name string, The identifier of the channel.
}}
}
\description{
List available channels and provide keys for accessing them
}
