% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_ml.R
\name{civis_file}
\alias{civis_file}
\title{A file in the Civis Platform}
\usage{
civis_file(file_id)
}
\arguments{
\item{file_id}{The id of a Civis file.}
}
\value{
A \code{civis_file}.
}
\description{
Use \code{civis_file} to use a file in the Civis Platform with
  \link{civis_ml}.
}
