% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_ml.R
\name{civis_file_manifest}
\alias{civis_file_manifest}
\title{A manifest file in the Civis Platform}
\usage{
civis_file_manifest(file_id)
}
\arguments{
\item{file_id}{The id of a Civis file.}
}
\value{
A \code{civis_file_manifest}
}
\description{
Use \code{civis_file_manifest} to use a manifest file in the Civis Platform
with \link{civis_ml}.
}
