% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_ml_utils.R
\docType{data}
\name{CIVIS_ML_CLASSIFIERS}
\alias{CIVIS_ML_CLASSIFIERS}
\title{List of classification models.}
\format{An object of class \code{character} of length 4.}
\usage{
CIVIS_ML_CLASSIFIERS
}
\description{
List of classification models.
}
\keyword{datasets}
