% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_ml_utils.R
\docType{data}
\name{CIVIS_ML_REGRESSORS}
\alias{CIVIS_ML_REGRESSORS}
\title{List of civis_ml regression models.}
\format{An object of class \code{character} of length 7.}
\usage{
CIVIS_ML_REGRESSORS
}
\description{
List of civis_ml regression models.
}
\keyword{datasets}
