% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_future.R
\name{civis_platform}
\alias{civis_platform}
\title{Evaluate an expression in Civis Platform}
\usage{
civis_platform(...)
}
\arguments{
\item{...}{Arguments to \code{\link{CivisFuture}}}
}
\value{
The result of evaluating \code{expr}.
}
\description{
Evaluate an expression in Civis Platform
}
\examples{
\dontrun{
 eval_civis(quote(sqrt(10)))
}

}
