% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{civis_script}
\alias{civis_script}
\title{Civis Script}
\usage{
civis_script(id, run_id = NULL)
}
\arguments{
\item{id}{integer, job id of the script.}

\item{run_id}{integer, id of the run. If \code{NULL} (default), the most recent run will be used.}
}
\value{
a \code{civis_script} object.
}
\description{
Create a civis_script object from a job and run id.
}
\details{
A \code{civis_script} can be any custom, container, R, Python, SQL, or Javacript platform script.
}
\seealso{
\code{\link{read_civis}}

Other script_utils: 
\code{\link{fetch_output_file_ids}()},
\code{\link{run_civis}()},
\code{\link{run_template}()}
}
\concept{script_utils}
