% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_ml.R
\name{civis_table}
\alias{civis_table}
\title{A table in the Civis Platform}
\usage{
civis_table(table_name, database_name = NULL, sql_where = NULL,
  sql_limit = NULL)
}
\arguments{
\item{table_name}{The table name, as \code{"schema.table"}.}

\item{database_name}{The name of the database holding \code{table_name}.}

\item{sql_where}{A \code{SQL WHERE} clause used to scope the rows of the
training or prediction table. Note, the \code{WHERE} keyword is added at
runtime.}

\item{sql_limit}{A \code{SQL_LIMIT} clause for querying the training or
prediction set. Note, the \code{LIMIT} keyword is added at runtime.}
}
\value{
A \code{civis_table}.
}
\description{
Use \code{civis_table} to use a table in the Civis Platform with
\link{civis_ml}.
}
