% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{clusters_delete_kubernetes_partitions}
\alias{clusters_delete_kubernetes_partitions}
\title{Delete a Cluster Partition}
\usage{
clusters_delete_kubernetes_partitions(id, cluster_partition_id)
}
\arguments{
\item{id}{integer required. The ID of the cluster which this partition belongs to.}

\item{cluster_partition_id}{integer required. The ID of this cluster partition.}
}
\value{
An empty HTTP response
}
\description{
Delete a Cluster Partition
}
