% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{clusters_get_workers}
\alias{clusters_get_workers}
\title{Describe a Worker Cluster}
\usage{
clusters_get_workers(id)
}
\arguments{
\item{id}{integer required. The ID of this cluster.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this cluster.}
\item{instanceType}{string, The EC2 instance types in this cluster.}
\item{minInstances}{integer, The minimum number of instances in this cluster.}
\item{maxInstances}{integer, The maximum number of instances in this cluster.}
\item{instances}{integer, The number of instances currently in this cluster.}
\item{instanceMaxMemory}{integer, The amount of memory available to a single instance.}
\item{instanceMaxCpu}{integer, The number of processor shares available to a single instance.}
\item{instanceMaxDiskSpace}{number, The amount of memory available to a single instance.}
\item{region}{string, The AWS region that this cluster is in.}
\item{activeJobsCount}{integer, The number of jobs currently being run in the cluster.}
\item{queuedJobsCount}{integer, The number of jobs currently waiting to be run on the cluster.}
}
\description{
Describe a Worker Cluster
}
