% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{clusters_list_kubernetes_compute_hours}
\alias{clusters_list_kubernetes_compute_hours}
\title{List compute hours for a Kubernetes Cluster}
\usage{
clusters_list_kubernetes_compute_hours(id, include_usage_stats = NULL)
}
\arguments{
\item{id}{integer required.}

\item{include_usage_stats}{boolean optional. When true, usage stats are returned in instance config objects. Defaults to false.}
}
\value{
An array containing the following fields:
\item{totalNormalizedHours}{integer, The total number of normalized hours used by this cluster.}
\item{normalizedHoursByInstanceType}{string, Denotes the instance type the normalized hours are attributed to.}
\item{updatedAt}{string, }
\item{monthAndYear}{string, The month and year the normalized hours are attributed to.}
}
\description{
List compute hours for a Kubernetes Cluster
}
