% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{clusters_list_kubernetes_compute_metrics}
\alias{clusters_list_kubernetes_compute_metrics}
\title{Get historical compute metrics for a Kubernetes Cluster}
\usage{
clusters_list_kubernetes_compute_metrics(
  id,
  start_date = NULL,
  end_date = NULL,
  type = NULL
)
}
\arguments{
\item{id}{integer required. The ID of the Kubernetes cluster.}

\item{start_date}{string optional. UTC start date in YYYY-MM-DD format (inclusive).}

\item{end_date}{string optional. UTC end date in YYYY-MM-DD format (inclusive). Must be within 31 days of start_date.}

\item{type}{array optional. Compute metric types to retrieve. Allowed values are: ["Jobs"]. Defaults to returning metrics for all available types. The default may change as support for additional types is added.}
}
\value{
A list containing the following elements:
\item{url}{string, Presigned URL to download the compute metrics CSV file from S3.}
}
\description{
Get historical compute metrics for a Kubernetes Cluster
}
