% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{clusters_list_kubernetes_deployment_stats}
\alias{clusters_list_kubernetes_deployment_stats}
\title{Get stats about deployments associated with a Kubernetes Cluster}
\usage{
clusters_list_kubernetes_deployment_stats(id)
}
\arguments{
\item{id}{integer required. The ID of this cluster.}
}
\value{
An array containing the following fields:
\item{baseType}{string, The base type of this deployment}
\item{state}{string, State of the deployment}
\item{count}{integer, Number of deployments of base type and state}
\item{totalCpu}{integer, Total amount of CPU in millicores for deployments of base type and state}
\item{totalMemory}{integer, Total amount of Memory in megabytes for deployments of base type and state}
}
\description{
Get stats about deployments associated with a Kubernetes Cluster
}
