% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{clusters_list_kubernetes_instance_configs_historical_graphs}
\alias{clusters_list_kubernetes_instance_configs_historical_graphs}
\title{Get graphs of historical resource usage in an Instance Config}
\usage{
clusters_list_kubernetes_instance_configs_historical_graphs(
  instance_config_id,
  timeframe = NULL
)
}
\arguments{
\item{instance_config_id}{integer required. The ID of this instance config.}

\item{timeframe}{string optional. The span of time that the graphs cover. Must be one of 1_day, 1_week.}
}
\value{
A list containing the following elements:
\item{cpuGraphUrl}{string, URL for the graph of historical CPU usage in this instance config.}
\item{memGraphUrl}{string, URL for the graph of historical memory usage in this instance config.}
}
\description{
Get graphs of historical resource usage in an Instance Config
}
