% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{clusters_list_workers}
\alias{clusters_list_workers}
\title{List Worker Clusters}
\usage{
clusters_list_workers(limit = NULL, page_num = NULL, order = NULL,
  order_dir = NULL)
}
\arguments{
\item{limit}{integer optional. Number of results to return. Defaults to its maximum of 50.}

\item{page_num}{integer optional. Page number of the results to return. Defaults to the first page, 1.}

\item{order}{string optional. The field on which to order the result set. Defaults to id. Must be one of: id, created_at.}

\item{order_dir}{string optional. Direction in which to sort, either asc (ascending) or desc (descending) defaulting to asc.}
}
\value{
An array containing the following fields:
\item{id}{integer, The ID of this cluster.}
\item{instanceType}{string, The EC2 instance types in this cluster.}
\item{minInstances}{integer, The minimum number of instances in this cluster.}
\item{maxInstances}{integer, The maximum number of instances in this cluster.}
\item{region}{string, The AWS region that this cluster is in.}
\item{activeJobsCount}{integer, The number of jobs currently being run in the cluster.}
\item{queuedJobsCount}{integer, The number of jobs currently waiting to be run on the cluster.}
}
\description{
List Worker Clusters
}
