% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{clusters_patch_kubernetes}
\alias{clusters_patch_kubernetes}
\title{Update a Kubernetes Cluster}
\usage{
clusters_patch_kubernetes(id, instance_type = NULL, min_instances = NULL,
  max_instances = NULL, organization_id = NULL, organization_slug = NULL,
  security_group_id = NULL)
}
\arguments{
\item{id}{integer required. The ID of this cluster.}

\item{instance_type}{string optional. The EC2 instance types in this cluster.}

\item{min_instances}{integer optional. The minimum number of instances in this cluster.}

\item{max_instances}{integer optional. The maximum number of instances in this cluster.}

\item{organization_id}{string optional. The id of this cluster's organization.}

\item{organization_slug}{string optional. The slug of this cluster's organization.}

\item{security_group_id}{string optional. The security group to be added to the nodes of this cluster}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this cluster.}
\item{instanceType}{string, The EC2 instance types in this cluster.}
\item{minInstances}{integer, The minimum number of instances in this cluster.}
\item{maxInstances}{integer, The maximum number of instances in this cluster.}
\item{instanceMaxMemory}{integer, The amount of memory available to a single instance.}
\item{instanceMaxCpu}{integer, The number of processor shares available to a single instance.}
\item{organizationId}{string, The id of this cluster's organization.}
\item{organizationSlug}{string, The slug of this cluster's organization.}
\item{securityGroupId}{string, The security group to be added to the nodes of this cluster}
}
\description{
Update a Kubernetes Cluster
}
