% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{codes_delete}
\alias{codes_delete}
\title{Delete a code}
\usage{
codes_delete(id)
}
\arguments{
\item{id}{integer required. The ID for this code.}
}
\value{
An empty HTTP response
}
\description{
Delete a code
}
