% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{codes_get}
\alias{codes_get}
\title{Show basic code info}
\usage{
codes_get(id)
}
\arguments{
\item{id}{integer required. The ID for this code.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID for this code.}
\item{name}{string, Name of code.}
\item{type}{string, The code's type; e.g., Code::FrontEnd.}
\item{body}{string, Actual code contents; e.g., HTML, SQL, etc.}
\item{readme}{string, User specified information about this code. Markdown format accepted.}
\item{score}{integer, Internal Civis Use Only.}
\item{authS3Url}{string, Authorized_s3_link to file.}
\item{createdAt}{string, The creation time for this code.}
\item{updatedAt}{string, The last modification time for this code.}
}
\description{
Show basic code info
}
