% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{codes_list}
\alias{codes_list}
\title{List codes}
\usage{
codes_list(limit = NULL, type = NULL, featured = NULL)
}
\arguments{
\item{limit}{integer optional. The maximum number of codes to return.}

\item{type}{string optional. The code's type; e.g., Code::FrontEnd.}

\item{featured}{boolean optional. If true, return featured codes.}
}
\value{
An array containing the following fields:
\item{id}{integer, The ID for this code.}
\item{name}{string, Name of code.}
\item{type}{string, The code's type; e.g., Code::FrontEnd.}
\item{createdAt}{string, The creation time for this code.}
\item{updatedAt}{string, The last modification time for this code.}
}
\description{
List codes
}
