% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{codes_patch}
\alias{codes_patch}
\title{Update a code}
\usage{
codes_patch(id, name = NULL, type = NULL, body = NULL,
  readme = NULL, score = NULL, auth_s3_url = NULL)
}
\arguments{
\item{id}{integer required. The ID for this code.}

\item{name}{string optional. Name of code.}

\item{type}{string optional. The code's type; e.g., Code::FrontEnd.}

\item{body}{string optional. Actual code contents; e.g., HTML, SQL, etc.}

\item{readme}{string optional. User specified information about this code. Markdown format accepted.}

\item{score}{integer optional. Internal Civis Use Only.}

\item{auth_s3_url}{string optional. Authorized_s3_link to file.}
}
\value{
An empty HTTP response
}
\description{
Update a code
}
