% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{codes_put}
\alias{codes_put}
\title{Update a code}
\usage{
codes_put(id, name, type, body, readme, score = NULL, auth_s3_url = NULL)
}
\arguments{
\item{id}{integer required. The ID for this code.}

\item{name}{string required. Name of code.}

\item{type}{string required. The code's type; e.g., Code::FrontEnd.}

\item{body}{string required. Actual code contents; e.g., HTML, SQL, etc.}

\item{readme}{string required. User specified information about this code. Markdown format accepted.}

\item{score}{integer optional. Internal Civis Use Only.}

\item{auth_s3_url}{string optional. Authorized_s3_link to file.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID for this code.}
\item{name}{string, Name of code.}
\item{type}{string, The code's type; e.g., Code::FrontEnd.}
\item{body}{string, Actual code contents; e.g., HTML, SQL, etc.}
\item{readme}{string, User specified information about this code. Markdown format accepted.}
\item{score}{integer, Internal Civis Use Only.}
\item{authS3Url}{string, Authorized_s3_link to file.}
\item{createdAt}{string, The creation time for this code.}
\item{updatedAt}{string, The last modification time for this code.}
}
\description{
Update a code
}
