% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_ml_utils.R
\name{coef.civis_ml}
\alias{coef.civis_ml}
\title{Get civis_ml model coefficients}
\usage{
\method{coef}{civis_ml}(object, complete = TRUE, ...)
}
\arguments{
\item{object}{civis_ml_model}

\item{complete}{see documentation for generic \code{coef()}}

\item{...}{other arguments}
}
\value{
a matrix of coefficients, or `NULL` if none available from CivisML
}
\description{
Get civis_ml model coefficients
}
\details{
Outputs coefficients with names in the style of `stats::coef`
}
