% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_ml_utils.R
\name{coef}
\alias{coef}
\title{coef}
\usage{
coef(object, ...)
}
\arguments{
\item{model}{civis_ml_classifier}
}
\value{
coefs a vector of coefficients or `NULL` if the classifier does not have coefficients
}
\description{
coef
}
\details{
Outputs coefficients with names in the style of `stats::coef`
}
