% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{credentials_delete}
\alias{credentials_delete}
\title{Delete a credential}
\usage{
credentials_delete(id)
}
\arguments{
\item{id}{integer required. The ID of the credential.}
}
\value{
An empty HTTP response
}
\description{
Delete a credential
}
