% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{credentials_list_types}
\alias{credentials_list_types}
\title{Get list of Credential Types}
\usage{
credentials_list_types()
}
\value{
A list containing the following elements:
\item{types}{array, list of acceptable credential types}
}
\description{
Get list of Credential Types
}
