% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{credentials_post_temporary}
\alias{credentials_post_temporary}
\title{Generate a temporary credential for accessing S3}
\usage{
credentials_post_temporary(id, duration = NULL)
}
\arguments{
\item{id}{integer required. The ID of the credential.}

\item{duration}{integer optional. The number of seconds the temporary credential should be valid. Defaults to 15 minutes. Must not be less than 15 minutes or greater than 36 hours.}
}
\value{
A list containing the following elements:
\item{accessKey}{string, The identifier of the credential.}
\item{secretAccessKey}{string, The secret part of the credential.}
\item{sessionToken}{string, The session token identifier.}
}
\description{
Generate a temporary credential for accessing S3
}
