% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{databases_delete_whitelist_ips}
\alias{databases_delete_whitelist_ips}
\title{Remove a whitelisted IP address}
\usage{
databases_delete_whitelist_ips(id, whitelisted_ip_id)
}
\arguments{
\item{id}{integer required. The ID of the database this rule is applied to.}

\item{whitelisted_ip_id}{integer required. The ID of this whitelisted IP address.}
}
\value{
An empty HTTP response
}
\description{
Remove a whitelisted IP address
}
