% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{databases_get_schema_privileges}
\alias{databases_get_schema_privileges}
\title{Show schema privileges}
\usage{
databases_get_schema_privileges(id, schema_name, credential_id = NULL)
}
\arguments{
\item{id}{integer required. The ID of the database}

\item{schema_name}{string required. The name of the schema}

\item{credential_id}{integer optional. If provided, schemas will be filtered based on the given credential.}
}
\value{
A list containing the following elements:
\item{grantee}{string, Name of the granted user or group}
\item{granteeType}{string, User or group}
\item{privileges}{array, Privileges that the grantee has on this resource}
\item{grantablePrivileges}{array, Privileges that the grantee can grant to others for this resource}
}
\description{
Show schema privileges
}
