% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{databases_get_status_graphs_timeframe}
\alias{databases_get_status_graphs_timeframe}
\title{Get the status graphs for this database}
\usage{
databases_get_status_graphs_timeframe(id, timeframe)
}
\arguments{
\item{id}{integer required. The ID of the database.}

\item{timeframe}{string required. The span of time that the graphs cover. Must be one of 1_hour, 4_hours, 1_day, 2_days, 1_week.}
}
\value{
A list containing the following elements:
\item{cpuGraphUrl}{string, URL for the aws redshift cpu utliization graph.}
\item{diskGraphUrl}{string, URL for the aws redshift disk usage graph.}
\item{queueLengthGraphUrl}{string, URL for the aws redshift queue length graph.}
\item{statusGraphUrl}{string, URL for the aws redshift status graph.}
\item{maintenanceGraphUrl}{string, URL for the aws redshift maintenance graph.}
\item{queryDurationGraphUrl}{string, URL for the aws redshift table count graph.}
}
\description{
Get the status graphs for this database
}
