% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{databases_get_whitelist_ips}
\alias{databases_get_whitelist_ips}
\title{View details about a whitelisted IP}
\usage{
databases_get_whitelist_ips(id, whitelisted_ip_id)
}
\arguments{
\item{id}{integer required. The ID of the database this rule is applied to.}

\item{whitelisted_ip_id}{integer required. The ID of this whitelisted IP address.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this whitelisted IP address.}
\item{remoteHostId}{integer, The ID of the database this rule is applied to.}
\item{securityGroupId}{string, The ID of the security group this rule is applied to.}
\item{subnetMask}{string, The subnet mask that is allowed by this rule.}
\item{authorizedBy}{string, The user who authorized this rule.}
\item{isActive}{boolean, True if the rule is applied, false if it has been revoked.}
\item{createdAt}{string, The time this rule was created.}
\item{updatedAt}{string, The time this rule was last updated.}
}
\description{
View details about a whitelisted IP
}
