% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{databases_list}
\alias{databases_list}
\title{List databases}
\usage{
databases_list()
}
\value{
An array containing the following fields:
\item{id}{integer, The ID for the database.}
\item{name}{string, The name of the database in Platform.}
\item{adapter}{string, The type of the database.}
\item{clusterIdentifier}{string, The cluster identifier of the database.}
\item{host}{string, The host of the database server.}
\item{port}{integer, The port of the database.}
\item{databaseName}{string, The internal name of the database.}
\item{managed}{boolean, True if the database is Civis-managed. False otherwise.}
}
\description{
List databases
}
