% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{databases_list_advanced_settings}
\alias{databases_list_advanced_settings}
\title{Get the advanced settings for this database}
\usage{
databases_list_advanced_settings(id)
}
\arguments{
\item{id}{integer required. The ID of the database this advanced settings object belongs to.}
}
\value{
A list containing the following elements:
\item{exportCachingEnabled}{boolean, Whether or not caching is enabled for export jobs run on this database server.}
}
\description{
Get the advanced settings for this database
}
