% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{databases_list_groups}
\alias{databases_list_groups}
\title{List groups in the specified database}
\usage{
databases_list_groups(id)
}
\arguments{
\item{id}{integer required. The ID for the database.}
}
\value{
An array containing the following fields:
\item{groupName}{string, The name of the group.}
\item{members}{array, The members of the group.}
}
\description{
List groups in the specified database
}
