% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{databases_list_schemas}
\alias{databases_list_schemas}
\title{List schemas in this database}
\usage{
databases_list_schemas(id)
}
\arguments{
\item{id}{integer required. The ID of the database.}
}
\value{
An array containing the following fields:
\item{schema}{string, The name of a schema.}
}
\description{
List schemas in this database
}
