% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{databases_list_schemas_tables}
\alias{databases_list_schemas_tables}
\title{List tables in this schema}
\usage{
databases_list_schemas_tables(id, schema_name, credential_id = NULL)
}
\arguments{
\item{id}{integer required. The ID of the database}

\item{schema_name}{string required. The name of the schema}

\item{credential_id}{integer optional. If provided, schemas will be filtered based on the given credential.}
}
\value{
An array containing the following fields:
\item{name}{string, The name of the table.}
\item{schema}{string, The name of the schema containing the table.}
\item{isView}{boolean, True if this table represents a view. False if it represents a regular table.}
\item{databaseId}{integer, The ID of the database server.}
}
\description{
List tables in this schema
}
