% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{databases_list_users}
\alias{databases_list_users}
\title{Show list of database users}
\usage{
databases_list_users(id, active = NULL)
}
\arguments{
\item{id}{integer required. The ID of the database.}

\item{active}{boolean optional. If true returns active users. If false returns deactivated users. Defaults to true.}
}
\value{
A list containing the following elements:
\item{username}{string, Username}
\item{active}{boolean, Whether the user is active or deactivated}
}
\description{
Show list of database users
}
