% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{databases_list_whitelist_ips}
\alias{databases_list_whitelist_ips}
\title{List whitelisted IPs for the specified database}
\usage{
databases_list_whitelist_ips(id)
}
\arguments{
\item{id}{integer required. The ID for the database.}
}
\value{
An array containing the following fields:
\item{id}{integer, The ID of this whitelisted IP address.}
\item{remoteHostId}{integer, The ID of the database this rule is applied to.}
\item{securityGroupId}{string, The ID of the security group this rule is applied to.}
\item{subnetMask}{string, The subnet mask that is allowed by this rule.}
\item{createdAt}{string, The time this rule was created.}
\item{updatedAt}{string, The time this rule was last updated.}
}
\description{
List whitelisted IPs for the specified database
}
