% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{databases_patch_advanced_settings}
\alias{databases_patch_advanced_settings}
\title{Update the advanced settings for this database}
\usage{
databases_patch_advanced_settings(id, export_caching_enabled = NULL)
}
\arguments{
\item{id}{integer required. The ID of the database this advanced settings object belongs to.}

\item{export_caching_enabled}{boolean optional. Whether or not caching is enabled for export jobs run on this database server.}
}
\value{
A list containing the following elements:
\item{exportCachingEnabled}{boolean, Whether or not caching is enabled for export jobs run on this database server.}
}
\description{
Update the advanced settings for this database
}
