% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{databases_put_schemas_tables_projects}
\alias{databases_put_schemas_tables_projects}
\title{Add a Table to a project}
\usage{
databases_put_schemas_tables_projects(id, project_id, schema_name, table_name)
}
\arguments{
\item{id}{integer required. The ID of the database}

\item{project_id}{integer required. The ID of the project.}

\item{schema_name}{string required. The name of the schema}

\item{table_name}{string required. The name of the table}
}
\value{
An empty HTTP response
}
\description{
Add a Table to a project
}
