% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{db_query_fields.CivisConnection}
\alias{db_query_fields.CivisConnection}
\title{Query field names.}
\usage{
\method{db_query_fields}{CivisConnection}(con, sql, ...)
}
\description{
This function is used to determine the fields/columns of a query.
}
\keyword{internal}
