% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-connection.R
\name{dbIsReadOnly}
\alias{dbIsReadOnly}
\title{Is this DBMS object in read-only mode}
\usage{
dbIsReadOnly(dbObj, ...)
}
\arguments{
\item{dbObj}{An object inheriting from \linkS4class{DBIObject},
i.e. \linkS4class{DBIDriver}, \linkS4class{DBIConnection},
or a \linkS4class{DBIResult}}

\item{...}{Other arguments to methods.}
}
\description{
This generic tests whether a database object is in read-only mode
}
