% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_credential}
\alias{default_credential}
\title{The current user's default credential.}
\usage{
default_credential()
}
\value{
credential_id, id of the default credential for the current user.
}
\description{
The current user's default credential.
}
