% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{enhancements_delete_cass_ncoa_projects}
\alias{enhancements_delete_cass_ncoa_projects}
\title{Remove a CASS/NCOA Enhancement from a project}
\usage{
enhancements_delete_cass_ncoa_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the CASS/NCOA Enhancement.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Remove a CASS/NCOA Enhancement from a project
}
