% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{enhancements_delete_civis_data_match_runs}
\alias{enhancements_delete_civis_data_match_runs}
\title{Cancel a run}
\usage{
enhancements_delete_civis_data_match_runs(id, run_id)
}
\arguments{
\item{id}{integer required. The ID of the Civis Data Match job.}

\item{run_id}{integer required. The ID of the run.}
}
\value{
An empty HTTP response
}
\description{
Cancel a run
}
