% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{enhancements_delete_data_unification_projects}
\alias{enhancements_delete_data_unification_projects}
\title{Remove a container docker from a project}
\usage{
enhancements_delete_data_unification_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the resource.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Remove a container docker from a project
}
