% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{enhancements_delete_geocode_projects}
\alias{enhancements_delete_geocode_projects}
\title{Remove a Geocode Enhancement from a project}
\usage{
enhancements_delete_geocode_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the Geocode Enhancement.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Remove a Geocode Enhancement from a project
}
