% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{enhancements_delete_identity_resolution_projects}
\alias{enhancements_delete_identity_resolution_projects}
\title{Remove an Identity Resolution Enhancement from a project}
\usage{
enhancements_delete_identity_resolution_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the Identity Resolution Enhancement.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Remove an Identity Resolution Enhancement from a project
}
