% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{enhancements_get_civis_data_match_runs}
\alias{enhancements_get_civis_data_match_runs}
\title{Check status of a run}
\usage{
enhancements_get_civis_data_match_runs(id, run_id)
}
\arguments{
\item{id}{integer required. The ID of the Civis Data Match job.}

\item{run_id}{integer required. The ID of the run.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the run.}
\item{civisDataMatchId}{integer, The ID of the Civis Data Match job.}
\item{state}{string, The state of the run, one of 'queued' 'running' 'succeeded' 'failed' or 'cancelled'.}
\item{isCancelRequested}{boolean, True if run cancel requested, else false.}
\item{createdAt}{string, The time the run was created.}
\item{startedAt}{string, The time the run started at.}
\item{finishedAt}{string, The time the run completed.}
\item{error}{string, The error, if any, returned by the run.}
}
\description{
Check status of a run
}
