% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{enhancements_list_field_mapping}
\alias{enhancements_list_field_mapping}
\title{List the fields in a field mapping for Civis Data Match, Data Unification, and Table Deduplication jobs}
\usage{
enhancements_list_field_mapping()
}
\value{
An array containing the following fields:
\item{field}{string, The name of the field.}
\item{description}{string, The description of the field.}
}
\description{
List the fields in a field mapping for Civis Data Match, Data Unification, and Table Deduplication jobs
}
