% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{enhancements_list_types}
\alias{enhancements_list_types}
\title{List available enhancement types}
\usage{
enhancements_list_types()
}
\value{
An array containing the following fields:
\item{name}{string, The name of the type.}
}
\description{
List available enhancement types
}
