% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{enhancements_put_civis_data_match_projects}
\alias{enhancements_put_civis_data_match_projects}
\title{Add a Civis Data Match Enhancement to a project}
\usage{
enhancements_put_civis_data_match_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the Civis Data Match Enhancement.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Add a Civis Data Match Enhancement to a project
}
