% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{enhancements_put_geocode_projects}
\alias{enhancements_put_geocode_projects}
\title{Add a Geocode Enhancement to a project}
\usage{
enhancements_put_geocode_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the Geocode Enhancement.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Add a Geocode Enhancement to a project
}
