% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{enhancements_put_identity_resolution_projects}
\alias{enhancements_put_identity_resolution_projects}
\title{Add an Identity Resolution Enhancement to a project}
\usage{
enhancements_put_identity_resolution_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the Identity Resolution Enhancement.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Add an Identity Resolution Enhancement to a project
}
