% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{enhancements_put_person_matching_projects}
\alias{enhancements_put_person_matching_projects}
\title{Add a container docker to a project}
\usage{
enhancements_put_person_matching_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the resource.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Add a container docker to a project
}
