% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{exports_get_files_csv_runs}
\alias{exports_get_files_csv_runs}
\title{Check status of a run}
\usage{
exports_get_files_csv_runs(id, run_id)
}
\arguments{
\item{id}{integer required. The ID of the csv_export.}

\item{run_id}{integer required. The ID of the run.}
}
\value{
A list containing the following elements:
\item{id}{integer, }
\item{state}{string, }
\item{createdAt}{string, The time that the run was queued.}
\item{startedAt}{string, The time that the run started.}
\item{finishedAt}{string, The time that the run completed.}
\item{error}{string, The error message for this run, if present.}
\item{outputCachedOn}{string, The time that the output was originally exported, if a cache entry was used by the run.}
}
\description{
Check status of a run
}
