% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{exports_patch_files_csv}
\alias{exports_patch_files_csv}
\title{Update some attributes of this CSV Export}
\usage{
exports_patch_files_csv(
  id,
  name = NULL,
  source = NULL,
  destination = NULL,
  include_header = NULL,
  compression = NULL,
  column_delimiter = NULL,
  hidden = NULL,
  force_multifile = NULL,
  max_file_size = NULL
)
}
\arguments{
\item{id}{integer required. The ID of this Csv Export job.}

\item{name}{string optional. The name of this Csv Export job.}

\item{source}{list optional. A list containing the following elements: 
\itemize{
\item sql string, The SQL query for this Csv Export job
\item remoteHostId integer, The ID of the destination database host.
\item credentialId integer, The ID of the credentials for the destination database.
}}

\item{destination}{list optional. A list containing the following elements: 
\itemize{
\item filenamePrefix string, The prefix of the name of the file returned to the user.
\item storagePath list . A list containing the following elements: 
\itemize{
\item filePath string, The path within the bucket where the exported file will be saved. E.g. the file_path for "s3://mybucket/files/all/" would be "/files/all/"
\item storageHostId integer, The ID of the destination storage host.
\item credentialId integer, The ID of the credentials for the destination storage host.
\item existingFiles string, Notifies the job of what to do in the case that the exported file already exists at the provided path.One of: fail, append, overwrite. Default: fail. If "append" is specified,the new file will always be added to the provided path. If "overwrite" is specifiedall existing files at the provided path will be deleted and the new file will be added.By default, or if "fail" is specified, the export will fail if a file exists at the provided path.
}
}}

\item{include_header}{boolean optional. A boolean value indicating whether or not the header should be included. Defaults to true.}

\item{compression}{string optional. The compression of the output file. Valid arguments are "gzip" and "none". Defaults to "gzip".}

\item{column_delimiter}{string optional. The column delimiter for the output file. Valid arguments are "comma", "tab", and "pipe". Defaults to "comma".}

\item{hidden}{boolean optional. A boolean value indicating whether or not this request should be hidden. Defaults to false.}

\item{force_multifile}{boolean optional. Whether or not the csv should be split into multiple files. Default: false}

\item{max_file_size}{integer optional. The max file size, in MB, created files will be. Only available when force_multifile is true.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this Csv Export job.}
\item{name}{string, The name of this Csv Export job.}
\item{source}{list, A list containing the following elements: 
\itemize{
\item sql string, The SQL query for this Csv Export job
\item remoteHostId integer, The ID of the destination database host.
\item credentialId integer, The ID of the credentials for the destination database.
}}
\item{destination}{list, A list containing the following elements: 
\itemize{
\item filenamePrefix string, The prefix of the name of the file returned to the user.
\item storagePath list . A list containing the following elements: 
\itemize{
\item filePath string, The path within the bucket where the exported file will be saved. E.g. the file_path for "s3://mybucket/files/all/" would be "/files/all/"
\item storageHostId integer, The ID of the destination storage host.
\item credentialId integer, The ID of the credentials for the destination storage host.
\item existingFiles string, Notifies the job of what to do in the case that the exported file already exists at the provided path.One of: fail, append, overwrite. Default: fail. If "append" is specified,the new file will always be added to the provided path. If "overwrite" is specifiedall existing files at the provided path will be deleted and the new file will be added.By default, or if "fail" is specified, the export will fail if a file exists at the provided path.
}
}}
\item{includeHeader}{boolean, A boolean value indicating whether or not the header should be included. Defaults to true.}
\item{compression}{string, The compression of the output file. Valid arguments are "gzip" and "none". Defaults to "gzip".}
\item{columnDelimiter}{string, The column delimiter for the output file. Valid arguments are "comma", "tab", and "pipe". Defaults to "comma".}
\item{hidden}{boolean, A boolean value indicating whether or not this request should be hidden. Defaults to false.}
\item{forceMultifile}{boolean, Whether or not the csv should be split into multiple files. Default: false}
\item{maxFileSize}{integer, The max file size, in MB, created files will be. Only available when force_multifile is true. }
}
\description{
Update some attributes of this CSV Export
}
