% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{exports_put_files_csv_archive}
\alias{exports_put_files_csv_archive}
\title{Update the archive status of this object}
\usage{
exports_put_files_csv_archive(id, status)
}
\arguments{
\item{id}{integer required. The ID of the object.}

\item{status}{boolean required. The desired archived status of the object.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this Csv Export job.}
\item{name}{string, The name of this Csv Export job.}
\item{source}{list, A list containing the following elements: 
\itemize{
\item sql string, The SQL query for this Csv Export job
\item remoteHostId integer, The ID of the destination database host.
\item credentialId integer, The ID of the credentials for the destination database.
}}
\item{destination}{list, A list containing the following elements: 
\itemize{
\item filenamePrefix string, The prefix of the name of the file returned to the user.
\item storagePath object, 
}}
\item{includeHeader}{boolean, A boolean value indicating whether or not the header should be included. Defaults to true.}
\item{compression}{string, The compression of the output file. Valid arguments are "gzip" and "none". Defaults to "gzip".}
\item{columnDelimiter}{string, The column delimiter for the output file. Valid arguments are "comma", "tab", and "pipe". Defaults to "comma".}
\item{hidden}{boolean, A boolean value indicating whether or not this request should be hidden. Defaults to false.}
\item{forceMultifile}{boolean, Whether or not the csv should be split into multiple files. Default: false}
\item{maxFileSize}{integer, The max file size, in MB, created files will be. Only available when force_multifile is true. }
}
\description{
Update the archive status of this object
}
