% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{feature_flags_delete_groups}
\alias{feature_flags_delete_groups}
\title{Deactivate a feature for a group}
\usage{
feature_flags_delete_groups(flag_name, group_id)
}
\arguments{
\item{flag_name}{string required. The feature flag name.}

\item{group_id}{integer required. Group ID.}
}
\value{
An empty HTTP response
}
\description{
Deactivate a feature for a group
}
