% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{feature_flags_delete_organizations}
\alias{feature_flags_delete_organizations}
\title{Deactivate a feature for a organization}
\usage{
feature_flags_delete_organizations(flag_name, organization_id)
}
\arguments{
\item{flag_name}{string required. The feature flag name.}

\item{organization_id}{integer required. Organization ID.}
}
\value{
An empty HTTP response
}
\description{
Deactivate a feature for a organization
}
