% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{feature_flags_delete_users}
\alias{feature_flags_delete_users}
\title{Deactivate a feature for a user}
\usage{
feature_flags_delete_users(flag_name, user_id)
}
\arguments{
\item{flag_name}{string required. The feature flag name.}

\item{user_id}{integer required. The user ID.}
}
\value{
An empty HTTP response
}
\description{
Deactivate a feature for a user
}
