% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{feature_flags_get}
\alias{feature_flags_get}
\title{Show a feature flag}
\usage{
feature_flags_get(name)
}
\arguments{
\item{name}{string required. The name of the feature flag.}
}
\value{
A list containing the following elements:
\item{name}{string, The name of the feature flag}
\item{organizations}{array, An array containing the following fields: 
\itemize{
\item id integer, Organization ID
\item name string, Organization name
\item slug string, Organization slug
}}
\item{groups}{array, An array containing the following fields: 
\itemize{
\item id integer, Group ID
\item name string, Group name
\item slug string, Group slug
\item users array, Users within the group
}}
\item{users}{array, An array containing the following fields: 
\itemize{
\item id integer, User ID
\item username string, Username
}}
\item{percentage}{integer, The target percentage of users who should have this feature flag enabled.}
}
\description{
Show a feature flag
}
