% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{feature_flags_list}
\alias{feature_flags_list}
\title{List feature flags}
\usage{
feature_flags_list()
}
\value{
An array containing the following fields:
\item{name}{string, The name of the feature.}
\item{description}{string, }
\item{activeForMe}{boolean, Whether the feature is active for the current user.}
\item{userCount}{integer, The number of users with this feature flag enabled.}
\item{team}{string, }
\item{jira}{string, }
\item{added}{string, }
\item{groupCount}{integer, }
\item{organizationCount}{integer, }
\item{percentage}{integer, The target percentage of users who should have this feature flag enabled.}
}
\description{
List feature flags
}
