% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{feature_flags_put_organizations}
\alias{feature_flags_put_organizations}
\title{Activate a feature for a organization}
\usage{
feature_flags_put_organizations(flag_name, organization_id)
}
\arguments{
\item{flag_name}{string required. The feature flag name.}

\item{organization_id}{integer required. Organization ID.}
}
\value{
A list containing the following elements:
\item{name}{string, The name of the feature.}
\item{description}{string, }
\item{activeForMe}{boolean, Whether the feature is active for the current user.}
\item{userCount}{integer, The number of users with this feature flag enabled.}
\item{team}{string, }
\item{jira}{string, }
\item{added}{string, }
\item{groupCount}{integer, }
\item{organizationCount}{integer, }
\item{percentage}{integer, The target percentage of users who should have this feature flag enabled.}
}
\description{
Activate a feature for a organization
}
