% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagination_helpers.R
\name{fetch_all}
\alias{fetch_all}
\title{Retrieve all results from a paginated endpoint}
\usage{
fetch_all(fn, ...)
}
\arguments{
\item{fn}{The API function to be called.}

\item{...}{Arguments passed to \code{fn}.}
}
\value{
A \code{list} with the concatenated results of each page of \code{fn}.
}
\description{
Retrieve all results from a paginated endpoint
}
\examples{
\dontrun{
columns <- fetch_all(tables_list_columns, id = 123)
column_names <- columns \%>\% purrr::map_chr("name")
}
}
\seealso{
Other pagination: 
\code{\link{fetch_until}()}
}
\concept{pagination}
