% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_ml.R
\name{fetch_oos_scores}
\alias{fetch_oos_scores}
\title{Retrieve out of sample/fold predictions}
\usage{
fetch_oos_scores(model, ...)
}
\arguments{
\item{model}{A \code{civis_ml} model.}

\item{\dots}{Parameters passed to \code{read.csv}.}
}
\value{
A \code{data.frame} with out of sample/fold predictions for each
 row of the training data, and containing an additional column with
 a primary key.
}
\description{
Retrieve out of sample/fold predictions
}
\details{
Returns the out of sample (or out of fold) predictions for each
  observation. For a multiclass model, a data frame is returned with one
  column of predictions for each class.
}
\seealso{
civis_ml
}
