% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{files_delete_preprocess_csv}
\alias{files_delete_preprocess_csv}
\title{Archive a Preprocess CSV (deprecated, use archiving endpoints instead)}
\usage{
files_delete_preprocess_csv(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
An empty HTTP response
}
\description{
Archive a Preprocess CSV (deprecated, use archiving endpoints instead)
}
