% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{files_delete_projects}
\alias{files_delete_projects}
\title{Remove an item from a project}
\usage{
files_delete_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the resource.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Remove an item from a project
}
