% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{files_post_multipart}
\alias{files_post_multipart}
\title{Initiate a multipart upload}
\usage{
files_post_multipart(name, num_parts, expires_at = NULL)
}
\arguments{
\item{name}{string required. The file name.}

\item{num_parts}{integer required. The number of parts in which the file will be uploaded. This parameter determines the number of presigned URLs that are returned.}

\item{expires_at}{string optional. The date and time the file will expire. If not specified, the file will expire in 30 days. To keep a file indefinitely, specify null.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the file object.}
\item{name}{string, The file name.}
\item{createdAt}{string, The date and time the file was created.}
\item{fileSize}{integer, The file size.}
\item{expiresAt}{string, The date and time the file will expire. If not specified, the file will expire in 30 days. To keep a file indefinitely, specify null.}
\item{uploadUrls}{array, An array of URLs that may be used to upload file parts. Use separate PUT requests to complete the part uploads. Links expire after 12 hours.}
}
\description{
Initiate a multipart upload
}
