% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{files_put_preprocess_csv}
\alias{files_put_preprocess_csv}
\title{Replace all attributes of this Preprocess CSV}
\usage{
files_put_preprocess_csv(
  id,
  file_id,
  in_place = NULL,
  detect_table_columns = NULL,
  force_character_set_conversion = NULL,
  include_header = NULL,
  column_delimiter = NULL
)
}
\arguments{
\item{id}{integer required. The ID of the job created.}

\item{file_id}{integer required. The ID of the file.}

\item{in_place}{boolean optional. If true, the file is cleaned in place. If false, a new file ID is created. Defaults to true.}

\item{detect_table_columns}{boolean optional. If true, detect the table columns in the file including the sql types. If false, skip table column detection.Defaults to false.}

\item{force_character_set_conversion}{boolean optional. If true, the file will always be converted to UTF-8 and any character that cannot be converted will be discarded. If false, the character set conversion will only run if the detected character set is not compatible with UTF-8 (e.g., UTF-8, ASCII).}

\item{include_header}{boolean optional. A boolean value indicating whether or not the first row of the file is a header row. If not provided, will attempt to auto-detect whether a header row is present.}

\item{column_delimiter}{string optional. The column delimiter for the file. One of "comma", "tab", or "pipe". If not provided, the column delimiter will be auto-detected.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the job created.}
\item{fileId}{integer, The ID of the file.}
\item{inPlace}{boolean, If true, the file is cleaned in place. If false, a new file ID is created. Defaults to true.}
\item{detectTableColumns}{boolean, If true, detect the table columns in the file including the sql types. If false, skip table column detection.Defaults to false.}
\item{forceCharacterSetConversion}{boolean, If true, the file will always be converted to UTF-8 and any character that cannot be converted will be discarded. If false, the character set conversion will only run if the detected character set is not compatible with UTF-8 (e.g., UTF-8, ASCII).}
\item{includeHeader}{boolean, A boolean value indicating whether or not the first row of the file is a header row. If not provided, will attempt to auto-detect whether a header row is present.}
\item{columnDelimiter}{string, The column delimiter for the file. One of "comma", "tab", or "pipe". If not provided, the column delimiter will be auto-detected.}
\item{hidden}{boolean, The hidden status of the item.}
}
\description{
Replace all attributes of this Preprocess CSV
}
