% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagination_helpers.R
\name{find}
\alias{find}
\title{Retrieve items from a paginated endpoint matching keys}
\usage{
find(fn, keys, ...)
}
\arguments{
\item{fn}{function, the API function to be called (usually a list endpoint)}

\item{keys}{list of key value pairs.}

\item{...}{Arguments passed to \code{fn}.}
}
\value{
Items from \code{fn} that match all key value pairs in \code{keys}.
}
\description{
Retrieve items from a paginated endpoint matching keys
}
\examples{
\dontrun{
find(credentials_list, list(username = 'username', name = 'cred_name')))
find(templates_list_scripts, list(name = 'template name'))

}
}
