% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_client.R
\name{generate_client}
\alias{generate_client}
\title{Generate a cilent}
\usage{
generate_client(spec, IGNORE)
}
\arguments{
\item{spec}{usually from \code{get_spec}}

\item{IGNORE}{endpoints to ignore (e.g. "apps")}
}
\value{
A string containing one documented function for each verb at each endpoint.
}
\description{
Generate a cilent
}
