% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_database_id}
\alias{get_database_id}
\title{Return the database id for a given database name}
\usage{
get_database_id(database_name)
}
\arguments{
\item{database_name}{string, Name of database.}
}
\value{
database_id integer, id of the matching database.
}
\description{
Return the database id for a given database name
}
\concept{id}
