% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_ml_utils.R
\name{get_feature_importance}
\alias{get_feature_importance}
\title{Get model feature importance}
\usage{
get_feature_importance(model)
}
\arguments{
\item{model}{Object from \code{\link{civis_ml}}}
}
\value{
A matrix of features and their importance, ranked in descending order of importance
}
\description{
Get model feature importance
}
