% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_ml_utils.R
\name{get_job_type_version}
\alias{get_job_type_version}
\title{Derive the job type and version from the given alias.}
\usage{
get_job_type_version(alias)
}
\arguments{
\item{alias}{A one-length character vector of the CivisML alias}
}
\value{
A list containing the job_type and version
}
\description{
Derive the job type and version from the given alias.
}
