% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_ml_utils.R
\name{get_metric}
\alias{get_metric}
\title{Get model metrics}
\usage{
get_metric(model, name = NULL)
}
\arguments{
\item{model}{Object from civis_ml}

\item{name}{Name of the metric.}
}
\value{
The metric given by \code{metric_name}. If \code{NULL}, all metrics are returned.
}
\description{
Get model metrics
}
