% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{get_script_fun}
\alias{get_script_fun}
\title{Get a script function matching a job type.}
\usage{
get_script_fun(job, fun_type = c("outputs", "logs"))
}
\arguments{
\item{job}{output of \code{jobs_get}}

\item{fun_type}{one of \code{"logs"} or \code{"outputs"}}
}
\value{
The correct output or log fetching function
(e.g. \code{\link{scripts_list_containers_runs_logs}}) based on the job type.
}
\description{
Get a script function matching a job type.
}
\details{
container and custom scripts both have the same job type, but can be distinguished
by a non-null \code{fromTemplateId}.
}
