% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/await.R
\name{get_status}
\alias{get_status}
\title{Get the status from results of \code{await}}
\usage{
get_status(response)
}
\arguments{
\item{response}{the results from \code{await}}
}
\description{
Get the status from results of \code{await}
}
\seealso{
await await_all
}
