% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_table_id}
\alias{get_table_id}
\title{Return the table id for a given table}
\usage{
get_table_id(table_name, database = NULL)
}
\arguments{
\item{table_name}{string, Name of table and schema \code{"schema.tablename"}}

\item{database}{string, Name of database where data frame is to be uploaded. If no database is specified,
uses \code{options(civis.default_db)}}
}
\value{
table_id integer, id of the matching table.
}
\description{
Return the table id for a given table
}
\seealso{
Other tables: \code{\link{refresh_table}},
  \code{\link{transfer_table}}
}
