% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_ml_utils.R
\name{get_template_ids_all_versions}
\alias{get_template_ids_all_versions}
\title{Get template IDs for all accessible CivisML versions.}
\usage{
get_template_ids_all_versions()
}
\value{
A data frame containing the template id, CivisML version, and job type.
}
\description{
Get template IDs for all accessible CivisML versions.
}
