% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_ml_utils.R
\name{get_train_template_id}
\alias{get_train_template_id}
\title{Get training template IDs for the specified CivisML version}
\usage{
get_train_template_id(civisml_version = "prod")
}
\arguments{
\item{civisml_version}{A one-length character vector of the CivisML version}
}
\value{
The template ID for training
}
\description{
Get training template IDs for the specified CivisML version
}
