% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{git_repos_delete}
\alias{git_repos_delete}
\title{Remove the bookmark on a git repository}
\usage{
git_repos_delete(id)
}
\arguments{
\item{id}{integer required. The ID for this git repository.}
}
\value{
An empty HTTP response
}
\description{
Remove the bookmark on a git repository
}
