% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{git_repos_get}
\alias{git_repos_get}
\title{Get a bookmarked git repository}
\usage{
git_repos_get(id)
}
\arguments{
\item{id}{integer required. The ID for this git repository.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID for this git repository.}
\item{repoUrl}{string, The URL for this git repository.}
\item{createdAt}{string, }
\item{updatedAt}{string, }
}
\description{
Get a bookmarked git repository
}
