% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{git_repos_list}
\alias{git_repos_list}
\title{List bookmarked git repositories}
\usage{
git_repos_list(limit = NULL, page_num = NULL, order = NULL,
  order_dir = NULL)
}
\arguments{
\item{limit}{integer optional. Number of results to return. Defaults to 20. Maximum allowed is 100.}

\item{page_num}{integer optional. Page number of the results to return. Defaults to the first page, 1.}

\item{order}{string optional. The field on which to order the result set. Defaults to repo_url. Must be one of: repo_url, created_at.}

\item{order_dir}{string optional. Direction in which to sort, either asc (ascending) or desc (descending) defaulting to asc.}
}
\value{
An array containing the following fields:
\item{id}{integer, The ID for this git repository.}
\item{repoUrl}{string, The URL for this git repository.}
\item{createdAt}{string, }
\item{updatedAt}{string, }
}
\description{
List bookmarked git repositories
}
