% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{git_repos_list_refs}
\alias{git_repos_list_refs}
\title{Get all branches and tags of a bookmarked git repository}
\usage{
git_repos_list_refs(id)
}
\arguments{
\item{id}{integer required. The ID for this git repository.}
}
\value{
A list containing the following elements:
\item{branches}{array, List of branch names of this git repository.}
\item{tags}{array, List of tag names of this git repository.}
}
\description{
Get all branches and tags of a bookmarked git repository
}
