% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{git_repos_post}
\alias{git_repos_post}
\title{Bookmark a git repository}
\usage{
git_repos_post(repo_url)
}
\arguments{
\item{repo_url}{string required. The URL for this git repository.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID for this git repository.}
\item{repoUrl}{string, The URL for this git repository.}
\item{createdAt}{string, }
\item{updatedAt}{string, }
}
\description{
Bookmark a git repository
}
