% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{groups_delete}
\alias{groups_delete}
\title{Delete a Group (deprecated)}
\usage{
groups_delete(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
An empty HTTP response
}
\description{
Delete a Group (deprecated)
}
