% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{groups_delete_members}
\alias{groups_delete_members}
\title{Remove a user from a group}
\usage{
groups_delete_members(id, user_id)
}
\arguments{
\item{id}{integer required. The ID of the group.}

\item{user_id}{integer required. The ID of the user.}
}
\value{
An empty HTTP response
}
\description{
Remove a user from a group
}
