% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{groups_list_child_groups}
\alias{groups_list_child_groups}
\title{Get child groups of this group}
\usage{
groups_list_child_groups(id)
}
\arguments{
\item{id}{integer required. The ID of this group.}
}
\value{
A list containing the following elements:
\item{manageable}{array, An array containing the following fields: 
\itemize{
\item id integer, 
\item name string, 
}}
\item{writeable}{array, An array containing the following fields: 
\itemize{
\item id integer, 
\item name string, 
}}
\item{readable}{array, An array containing the following fields: 
\itemize{
\item id integer, 
\item name string, 
}}
}
\description{
Get child groups of this group
}
