% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{groups_post}
\alias{groups_post}
\title{Create a Group}
\usage{
groups_post(name, description = NULL)
}
\arguments{
\item{name}{string required. This group's name.}

\item{description}{string optional. The description of the group.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this group.}
\item{name}{string, This group's name.}
\item{createdAt}{string, The date and time when this group was created.}
\item{description}{string, The description of the group.}
\item{slug}{string, The slug for this group.}
\item{organizationId}{integer, The ID of the organization this group belongs to.}
\item{organizationName}{string, The name of the organization this group belongs to.}
\item{memberCount}{integer, The total number of members in this group.}
\item{members}{array, An array containing the following fields: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
}
\description{
Create a Group
}
