% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{groups_put_members}
\alias{groups_put_members}
\title{Add a user to a group}
\usage{
groups_put_members(id, user_id)
}
\arguments{
\item{id}{integer required. The ID of the group.}

\item{user_id}{integer required. The ID of the user.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this group.}
\item{name}{string, This group's name.}
\item{createdAt}{string, The date and time when this group was created.}
\item{description}{string, The description of the group.}
\item{slug}{string, The slug for this group.}
\item{organizationId}{integer, The ID of the organization this group belongs to.}
\item{organizationName}{string, The name of the organization this group belongs to.}
\item{memberCount}{integer, The total number of members in this group.}
\item{mustAgreeToEula}{boolean, Whether or not members of this group must sign the EULA.}
\item{defaultOtpRequiredForLogin}{boolean, The two factor authentication requirement for this group.}
\item{roleIds}{array, An array of ids of all the roles this group has.}
\item{defaultTimeZone}{string, The default time zone of this group.}
\item{defaultJobsLabel}{string, The default partition label for jobs of this group. Only available if custom_partitions feature flag is set. Do not use this attribute as it may break in the future.}
\item{defaultNotebooksLabel}{string, The default partition label for notebooks of this group. Only available if custom_partitions feature flag is set. Do not use this attribute as it may break in the future.}
\item{defaultServicesLabel}{string, The default partition label for services of this group. Only available if custom_partitions feature flag is set. Do not use this attribute as it may break in the future.}
\item{members}{array, An array containing the following fields: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
}
\description{
Add a user to a group
}
