% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_ml_plot.R
\name{hist.civis_ml}
\alias{hist.civis_ml}
\title{Plot a histogram of the predicted scores or probabilities for a CivisML object}
\usage{
\method{hist}{civis_ml}(x, name = NULL, ...)
}
\arguments{
\item{x}{civis_ml object}

\item{name}{name of the class (for multiclass output)}

\item{...}{unused.}
}
\value{
A \code{ggplot2} plot object is returned invisibly.
}
\description{
Plot a histogram of the predicted scores or probabilities for a CivisML object
}
