% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{imports_delete_files_csv}
\alias{imports_delete_files_csv}
\title{Archive a CSV Import (deprecated, use archiving endpoints instead)}
\usage{
imports_delete_files_csv(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
An empty HTTP response
}
\description{
Archive a CSV Import (deprecated, use archiving endpoints instead)
}
