% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{imports_delete_files_runs}
\alias{imports_delete_files_runs}
\title{Cancel a run}
\usage{
imports_delete_files_runs(id, run_id)
}
\arguments{
\item{id}{integer required. The ID of the import.}

\item{run_id}{integer required. The ID of the run.}
}
\value{
An empty HTTP response
}
\description{
Cancel a run
}
