% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{imports_delete_syncs}
\alias{imports_delete_syncs}
\title{Archive a sync (deprecated, use the /archive endpoint instead)}
\usage{
imports_delete_syncs(id, sync_id)
}
\arguments{
\item{id}{integer required. The ID of the import to fetch.}

\item{sync_id}{integer required. The ID of the sync to fetch.}
}
\value{
An empty HTTP response
}
\description{
Archive a sync (deprecated, use the /archive endpoint instead)
}
