% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{imports_get_batches}
\alias{imports_get_batches}
\title{Get details about a batch import}
\usage{
imports_get_batches(id)
}
\arguments{
\item{id}{integer required. The ID for the import.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID for the import.}
\item{schema}{string, The destination schema name. This schema must already exist in Redshift.}
\item{table}{string, The destination table name, without the schema prefix. This table must already exist in Redshift.}
\item{remoteHostId}{integer, The ID of the destination database host.}
\item{state}{string, The state of the run; one of "queued", "running", "succeeded", "failed", or "cancelled".}
\item{startedAt}{string, The time the last run started at.}
\item{finishedAt}{string, The time the last run completed.}
\item{error}{string, The error returned by the run, if any.}
\item{hidden}{boolean, The hidden status of the object.}
}
\description{
Get details about a batch import
}
