% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{imports_list_runs}
\alias{imports_list_runs}
\title{Get the run history of this import}
\usage{
imports_list_runs(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
An array containing the following fields:
\item{id}{integer, }
\item{state}{string, }
\item{createdAt}{string, The time that the run was queued.}
\item{startedAt}{string, The time that the run started.}
\item{finishedAt}{string, The time that the run completed.}
\item{error}{string, The error message for this run, if present.}
}
\description{
Get the run history of this import
}
