% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{imports_post_cancel}
\alias{imports_post_cancel}
\title{Cancel a run}
\usage{
imports_post_cancel(id)
}
\arguments{
\item{id}{integer required. The ID of the job.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the run.}
\item{state}{string, The state of the run, one of 'queued', 'running' or 'cancelled'.}
\item{isCancelRequested}{boolean, True if run cancel requested, else false.}
}
\description{
Cancel a run
}
