% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{imports_post_runs}
\alias{imports_post_runs}
\title{Run an import}
\usage{
imports_post_runs(id)
}
\arguments{
\item{id}{integer required. The ID of the import to run.}
}
\value{
A list containing the following elements:
\item{runId}{integer, The ID of the new run triggered.}
}
\description{
Run an import
}
