% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{jobs_delete_runs}
\alias{jobs_delete_runs}
\title{Cancel a run}
\usage{
jobs_delete_runs(id, run_id)
}
\arguments{
\item{id}{integer required. The ID of the Job.}

\item{run_id}{integer required. The ID of the Run.}
}
\value{
An empty HTTP response
}
\description{
Cancel a run
}
