% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{jobs_get_runs}
\alias{jobs_get_runs}
\title{Check status of a job}
\usage{
jobs_get_runs(id, run_id)
}
\arguments{
\item{id}{integer required. The ID of the Job.}

\item{run_id}{integer required. The ID of the Run.}
}
\value{
A list containing the following elements:
\item{id}{integer, }
\item{state}{string, }
\item{createdAt}{string, The time that the run was queued.}
\item{startedAt}{string, The time that the run started.}
\item{finishedAt}{string, The time that the run completed.}
\item{error}{string, The error message for this run, if present.}
}
\description{
Check status of a job
}
