% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{jobs_list_runs}
\alias{jobs_list_runs}
\title{List runs for the given job}
\usage{
jobs_list_runs(id, limit = NULL, page_num = NULL, order = NULL,
  order_dir = NULL)
}
\arguments{
\item{id}{integer required. The ID for this job.}

\item{limit}{integer optional. Number of results to return. Defaults to 20. Maximum allowed is 100.}

\item{page_num}{integer optional. Page number of the results to return. Defaults to the first page, 1.}

\item{order}{string optional. The field on which to order the result set. Defaults to id. Must be one of: id.}

\item{order_dir}{string optional. Direction in which to sort, either asc (ascending) or desc (descending) defaulting to desc.}
}
\value{
An array containing the following fields:
\item{id}{integer, }
\item{state}{string, }
\item{createdAt}{string, The time that the run was queued.}
\item{startedAt}{string, The time that the run started.}
\item{finishedAt}{string, The time that the run completed.}
\item{error}{string, The error message for this run, if present.}
}
\description{
List runs for the given job
}
