% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{jobs_list_runs_logs}
\alias{jobs_list_runs_logs}
\title{Get the logs for a run}
\usage{
jobs_list_runs_logs(id, run_id, last_id = NULL, limit = NULL)
}
\arguments{
\item{id}{integer required. The ID of the job job.}

\item{run_id}{integer required. The ID of the run.}

\item{last_id}{integer optional. The ID of the last log message received. Log entries with this ID value or lower will be omitted.Logs are sorted by ID if this value is provided, and are otherwise sorted by createdAt.}

\item{limit}{integer optional. The maximum number of log messages to return. Default of 10000.}
}
\value{
An array containing the following fields:
\item{id}{integer, The ID of the log.}
\item{createdAt}{string, The time the log was created.}
\item{message}{string, The log message.}
\item{level}{string, The level of the log. One of unknown,fatal,error,warn,info,debug.}
}
\description{
Get the logs for a run
}
