% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{jobs_post_runs}
\alias{jobs_post_runs}
\title{Run a job}
\usage{
jobs_post_runs(id)
}
\arguments{
\item{id}{integer required. The ID for this job.}
}
\value{
A list containing the following elements:
\item{id}{integer, }
\item{state}{string, }
\item{createdAt}{string, The time that the run was queued.}
\item{startedAt}{string, The time that the run started.}
\item{finishedAt}{string, The time that the run completed.}
\item{error}{string, The error message for this run, if present.}
}
\description{
Run a job
}
