% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{jobs_post_trigger_email}
\alias{jobs_post_trigger_email}
\title{Generate and retrieve trigger email address}
\usage{
jobs_post_trigger_email(id)
}
\arguments{
\item{id}{integer required. The ID for this job.}
}
\value{
A list containing the following elements:
\item{triggerEmail}{string, Email address which may be used to trigger this job to run.}
}
\description{
Generate and retrieve trigger email address
}
