% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{json_values_patch}
\alias{json_values_patch}
\title{Update some attributes of this JSON Value}
\usage{
json_values_patch(id, name = NULL, value_str = NULL)
}
\arguments{
\item{id}{integer required. The ID of the JSON Value.}

\item{name}{string optional. The name of the JSON Value.}

\item{value_str}{string optional. The JSON value to store. Should be a serialized JSON string. Limited to 1000000 bytes.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the JSON Value.}
\item{name}{string, The name of the JSON Value.}
\item{value}{string, The deserialized JSON value.}
}
\description{
Update some attributes of this JSON Value
}
