% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{json_values_post}
\alias{json_values_post}
\title{Create a JSON Value}
\usage{
json_values_post(value_str, name = NULL)
}
\arguments{
\item{value_str}{string required. The JSON value to store. Should be a serialized JSON string. Limited to 1000000 bytes.}

\item{name}{string optional. The name of the JSON Value.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the JSON Value.}
\item{name}{string, The name of the JSON Value.}
\item{value}{string, The deserialized JSON value.}
}
\description{
Create a JSON Value
}
