% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{match_targets_get}
\alias{match_targets_get}
\title{Show Match Target info}
\usage{
match_targets_get(id)
}
\arguments{
\item{id}{integer required. The ID of the match target}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the match target}
\item{name}{string, The name of the match target}
\item{targetFileName}{string, The name of the target file}
\item{createdAt}{string, }
\item{updatedAt}{string, }
\item{archived}{boolean, Whether the match target has been archived.}
}
\description{
Show Match Target info
}
