% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{match_targets_list}
\alias{match_targets_list}
\title{List match targets}
\usage{
match_targets_list()
}
\value{
An array containing the following fields:
\item{id}{integer, }
\item{name}{string, }
\item{table}{string, }
}
\description{
List match targets
}
