% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{match_targets_put_archive}
\alias{match_targets_put_archive}
\title{Update the archive status of this object}
\usage{
match_targets_put_archive(id, status)
}
\arguments{
\item{id}{integer required. The ID of the object.}

\item{status}{boolean required. The desired archived status of the object.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the match target}
\item{name}{string, The name of the match target}
\item{targetFileName}{string, The name of the target file}
\item{createdAt}{string, }
\item{updatedAt}{string, }
\item{archived}{boolean, Whether the match target has been archived.}
}
\description{
Update the archive status of this object
}
