% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{media_get_ratecards}
\alias{media_get_ratecards}
\title{Get a Ratecard}
\usage{
media_get_ratecards(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ratecard ID.}
\item{filename}{string, Name of the ratecard file.}
\item{startOn}{string, First day to which the ratecard applies.}
\item{endOn}{string, Last day to which the ratecard applies.}
\item{dmaNumber}{integer, Number of the DMA associated with the ratecard.}
\item{archived}{string, The archival status of the requested item(s).}
}
\description{
Get a Ratecard
}
