% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{media_list_dmas}
\alias{media_list_dmas}
\title{List all Designated Market Areas}
\usage{
media_list_dmas(name = NULL, number = NULL)
}
\arguments{
\item{name}{string optional. If specified, will be used to filter the DMAs returned. Substring matching is supported with "\%" and "*" wildcards (e.g., "name=\%region\%" will return both "region1" and "my region").}

\item{number}{integer optional. If specified, will be used to filter the DMAS by number.}
}
\value{
An array containing the following fields:
\item{name}{string, Name for the DMA region.}
\item{number}{integer, Identifier number for a DMA.}
}
\description{
List all Designated Market Areas
}
