% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{media_list_spot_orders}
\alias{media_list_spot_orders}
\title{List all spot orders}
\usage{
media_list_spot_orders(id = NULL, archived = NULL)
}
\arguments{
\item{id}{integer optional. The ID for the spot order.}

\item{archived}{string optional. The archival status of the requested object(s).}
}
\value{
An array containing the following fields:
\item{id}{integer, The ID for the spot order.}
\item{archived}{string, The archival status of the requested object(s).}
}
\description{
List all spot orders
}
