% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{media_list_targets}
\alias{media_list_targets}
\title{List all Media Targets}
\usage{
media_list_targets(name = NULL, identifier = NULL, data_source = NULL)
}
\arguments{
\item{name}{string optional. The name of the target.}

\item{identifier}{string optional. A unique identifier for this target.}

\item{data_source}{string optional. The source of viewership data for this target.}
}
\value{
An array containing the following fields:
\item{name}{string, The name of the target.}
\item{identifier}{string, A unique identifier for this target.}
\item{dataSource}{string, The source of viewership data for this target.}
}
\description{
List all Media Targets
}
