% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{media_patch_ratecards}
\alias{media_patch_ratecards}
\title{Update some attributes of this Ratecard}
\usage{
media_patch_ratecards(
  id,
  filename = NULL,
  start_on = NULL,
  end_on = NULL,
  dma_number = NULL
)
}
\arguments{
\item{id}{integer required. The ratecard ID.}

\item{filename}{string optional. Name of the ratecard file.}

\item{start_on}{string optional. First day to which the ratecard applies.}

\item{end_on}{string optional. Last day to which the ratecard applies.}

\item{dma_number}{integer optional. Number of the DMA associated with the ratecard.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ratecard ID.}
\item{filename}{string, Name of the ratecard file.}
\item{startOn}{string, First day to which the ratecard applies.}
\item{endOn}{string, Last day to which the ratecard applies.}
\item{dmaNumber}{integer, Number of the DMA associated with the ratecard.}
\item{archived}{string, The archival status of the requested item(s).}
}
\description{
Update some attributes of this Ratecard
}
