% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{media_post_spot_orders}
\alias{media_post_spot_orders}
\title{Create a spot order}
\usage{
media_post_spot_orders(body = NULL)
}
\arguments{
\item{body}{string optional. CSV body of a spot order.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID for the spot order.}
\item{archived}{string, The archival status of the requested item(s).}
\item{csvS3Uri}{string, S3 URI for the spot order CSV file.}
\item{jsonS3Uri}{string, S3 URI for the spot order JSON file.}
\item{xmlArchiveS3Uri}{string, S3 URI for the spot order XML archive.}
\item{lastTransformJobId}{integer, ID of the spot order transformation job.}
}
\description{
Create a spot order
}
