% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{media_put_optimizations_shares_groups}
\alias{media_put_optimizations_shares_groups}
\title{Set the permissions groups has on this object}
\usage{
media_put_optimizations_shares_groups(id, group_ids, permission_level)
}
\arguments{
\item{id}{integer required. ID of the resource to be shared}

\item{group_ids}{array required. An array of one or more group IDs}

\item{permission_level}{string required. Options are: "read", "write", or "manage"}
}
\value{
A list containing the following elements:
\item{readers}{object, A list containing the following elements: 
\itemize{
\item users array, 
\item groups array, 
}}
\item{writers}{object, A list containing the following elements: 
\itemize{
\item users array, 
\item groups array, 
}}
\item{owners}{object, A list containing the following elements: 
\itemize{
\item users array, 
\item groups array, 
}}
\item{totalUserShares}{integer, For owners, the number of total users shared. For writers and readers, the number of visible users shared.}
\item{totalGroupShares}{integer, For owners, the number of total groups shared. For writers and readers, the number of visible groups shared.}
}
\description{
Set the permissions groups has on this object
}
