% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{media_put_ratecards}
\alias{media_put_ratecards}
\title{Replace all attributes of this Ratecard}
\usage{
media_put_ratecards(id, filename, start_on, end_on, dma_number)
}
\arguments{
\item{id}{integer required. The ratecard ID.}

\item{filename}{string required. Name of the ratecard file.}

\item{start_on}{string required. First day to which the ratecard applies.}

\item{end_on}{string required. Last day to which the ratecard applies.}

\item{dma_number}{integer required. Number of the DMA associated with the ratecard.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ratecard ID.}
\item{filename}{string, Name of the ratecard file.}
\item{startOn}{string, First day to which the ratecard applies.}
\item{endOn}{string, Last day to which the ratecard applies.}
\item{dmaNumber}{integer, Number of the DMA associated with the ratecard.}
\item{archived}{string, The archival status of the requested object(s).}
}
\description{
Replace all attributes of this Ratecard
}
