% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{media_put_ratecards_archive}
\alias{media_put_ratecards_archive}
\title{Update the archive status of this object}
\usage{
media_put_ratecards_archive(id, status)
}
\arguments{
\item{id}{integer required. The ID of the object.}

\item{status}{boolean required. The desired archived status of the object.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ratecard ID.}
\item{filename}{string, Name of the ratecard file.}
\item{startOn}{string, First day to which the ratecard applies.}
\item{endOn}{string, Last day to which the ratecard applies.}
\item{dmaNumber}{integer, Number of the DMA associated with the ratecard.}
\item{archived}{string, The archival status of the requested item(s).}
}
\description{
Update the archive status of this object
}
