% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{media_put_spot_orders}
\alias{media_put_spot_orders}
\title{Edit the specified spot order}
\usage{
media_put_spot_orders(id, body = NULL)
}
\arguments{
\item{id}{integer required. The ID for the spot order.}

\item{body}{string optional. CSV body of a spot order.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID for the spot order.}
\item{archived}{string, The archival status of the requested object(s).}
\item{csvS3Uri}{string, S3 URI for the spot order CSV file.}
\item{jsonS3Uri}{string, S3 URI for the spot order JSON file.}
\item{xmlArchiveS3Uri}{string, S3 URI for the spot order XML archive.}
\item{lastTransformJobId}{integer, ID of the spot order transformation job.}
}
\description{
Edit the specified spot order
}
