% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{models_delete_builds}
\alias{models_delete_builds}
\title{Cancel a build}
\usage{
models_delete_builds(id, build_id)
}
\arguments{
\item{id}{integer required. The ID of the model.}

\item{build_id}{integer required. The ID of the build.}
}
\value{
An empty HTTP response
}
\description{
Cancel a build
}
