% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{models_delete_projects}
\alias{models_delete_projects}
\title{Remove a models from a project}
\usage{
models_delete_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the resource.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Remove a models from a project
}
