% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{models_delete_shares_groups}
\alias{models_delete_shares_groups}
\title{Revoke the permissions a group has on this object}
\usage{
models_delete_shares_groups(id, group_id)
}
\arguments{
\item{id}{integer required. The ID of the resource that is shared.}

\item{group_id}{integer required. The ID of the group.}
}
\value{
An empty HTTP response
}
\description{
Revoke the permissions a group has on this object
}
