% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{models_list_types}
\alias{models_list_types}
\title{List all available model types}
\usage{
models_list_types()
}
\value{
An array containing the following fields:
\item{id}{integer, The ID of the model type.}
\item{algorithm}{string, The name of the algorithm used to train the model.}
\item{dvType}{string, The type of dependent variable predicted by the model.}
\item{fintAllowed}{boolean, Whether this model type supports searching for interaction terms.}
}
\description{
List all available model types
}
