% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{models_put_projects}
\alias{models_put_projects}
\title{Add a Model to a project}
\usage{
models_put_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the Model.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Add a Model to a project
}
