% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{models_put_shares_groups}
\alias{models_put_shares_groups}
\title{Set the permissions groups has on this object}
\usage{
models_put_shares_groups(id, group_ids, permission_level,
  share_email_body = NULL, send_shared_email = NULL)
}
\arguments{
\item{id}{integer required. ID of the resource to be shared}

\item{group_ids}{array required. An array of one or more group IDs}

\item{permission_level}{string required. Options are: "read", "write", or "manage"}

\item{share_email_body}{string optional. Custom body text for e-mail sent on a share.}

\item{send_shared_email}{boolean optional. Send email to the recipients of a share.}
}
\value{
A list containing the following elements:
\item{readers}{list, A list containing the following elements: 
\itemize{
\item users array, 
\item groups array, 
}}
\item{writers}{list, A list containing the following elements: 
\itemize{
\item users array, 
\item groups array, 
}}
\item{owners}{list, A list containing the following elements: 
\itemize{
\item users array, 
\item groups array, 
}}
\item{totalUserShares}{integer, For owners, the number of total users shared. For writers and readers, the number of visible users shared.}
\item{totalGroupShares}{integer, For owners, the number of total groups shared. For writers and readers, the number of visible groups shared.}
}
\description{
Set the permissions groups has on this object
}
