% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{multipart_upload}
\alias{multipart_upload}
\title{Upload to files endpoint in parts.}
\usage{
multipart_upload(file, name = "", chunk_size = 32 * 1024, expires_at = NULL)
}
\arguments{
\item{file}{the file}

\item{name}{name of the upload, defualts to}

\item{chunk_size}{size of the chunks in bytes}

\item{expires_at}{when the file expires (default never).}
}
\description{
If a future::plan has been set, will be carried out in parallel.
}
