% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{notebooks_delete}
\alias{notebooks_delete}
\title{Archive a Notebook (deprecated, use archiving endpoints instead)}
\usage{
notebooks_delete(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
An empty HTTP response
}
\description{
Archive a Notebook (deprecated, use archiving endpoints instead)
}
