% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{notebooks_list_git_commits}
\alias{notebooks_list_git_commits}
\title{Get the git commits for this Notebook}
\usage{
notebooks_list_git_commits(id)
}
\arguments{
\item{id}{integer required. The ID of the object.}
}
\value{
A list containing the following elements:
\item{commitHash}{string, The SHA of the commit.}
\item{authorName}{string, The name of the commit's author.}
\item{date}{string, The commit's timestamp.}
\item{message}{string, The commit message.}
}
\description{
Get the git commits for this Notebook
}
