% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{notebooks_list_update_links}
\alias{notebooks_list_update_links}
\title{Get URLs to update notebook}
\usage{
notebooks_list_update_links(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
A list containing the following elements:
\item{updateUrl}{string, Time-limited URL to PUT new contents of the .ipynb file for this notebook.}
\item{updatePreviewUrl}{string, Time-limited URL to PUT new contents of the .htm preview file for this notebook.}
}
\description{
Get URLs to update notebook
}
