% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{notebooks_put_projects}
\alias{notebooks_put_projects}
\title{Add a Notebook to a project}
\usage{
notebooks_put_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the resource.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Add a Notebook to a project
}
