% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{notifications_list}
\alias{notifications_list}
\title{Receive a stream of notifications as they come in}
\usage{
notifications_list(last_event_id = NULL, r = NULL, mock = NULL)
}
\arguments{
\item{last_event_id}{string optional. allows browser to keep track of last event fired}

\item{r}{string optional. specifies retry/reconnect timeout}

\item{mock}{string optional. used for testing}
}
\value{
An empty HTTP response
}
\description{
Receive a stream of notifications as they come in
}
