% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{ontology_list}
\alias{ontology_list}
\title{List the ontology of column names Civis uses}
\usage{
ontology_list(subset = NULL)
}
\arguments{
\item{subset}{string optional. A subset of fields to return.}
}
\value{
An array containing the following fields:
\item{key}{string, }
\item{title}{string, }
\item{desc}{string, A description of this field.}
\item{aliases}{array, }
}
\description{
List the ontology of column names Civis uses
}
