% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{organizations_delete_favorites}
\alias{organizations_delete_favorites}
\title{Unfavorite an item for your organization}
\usage{
organizations_delete_favorites(id)
}
\arguments{
\item{id}{integer required. The id of the favorite.}
}
\value{
An empty HTTP response
}
\description{
Unfavorite an item for your organization
}
