% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{permission_sets_delete_resources}
\alias{permission_sets_delete_resources}
\title{Delete a resource in a permission set}
\usage{
permission_sets_delete_resources(id, name)
}
\arguments{
\item{id}{integer required. The ID for this permission set.}

\item{name}{string required. The name of this resource.}
}
\value{
An empty HTTP response
}
\description{
Delete a resource in a permission set
}
