% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{permission_sets_delete_resources_shares_groups}
\alias{permission_sets_delete_resources_shares_groups}
\title{Revoke the permissions a group has on this object}
\usage{
permission_sets_delete_resources_shares_groups(id, name, group_id)
}
\arguments{
\item{id}{integer required. The ID for this permission set.}

\item{name}{string required. The name of this resource.}

\item{group_id}{integer required. The ID of the group.}
}
\value{
An empty HTTP response
}
\description{
Revoke the permissions a group has on this object
}
