% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{permission_sets_delete_resources_shares_users}
\alias{permission_sets_delete_resources_shares_users}
\title{Revoke the permissions a user has on this object}
\usage{
permission_sets_delete_resources_shares_users(id, name, user_id)
}
\arguments{
\item{id}{integer required. The ID for this permission set.}

\item{name}{string required. The name of this resource.}

\item{user_id}{integer required. The ID of the user.}
}
\value{
An empty HTTP response
}
\description{
Revoke the permissions a user has on this object
}
