% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{permission_sets_post_resources}
\alias{permission_sets_post_resources}
\title{Create a resource in a permission set}
\usage{
permission_sets_post_resources(id, name, description = NULL)
}
\arguments{
\item{id}{integer required. The ID for this permission set.}

\item{name}{string required. The name of this resource.}

\item{description}{string optional. A description of this resource.}
}
\value{
A list containing the following elements:
\item{permissionSetId}{integer, The ID for the permission set this resource belongs to.}
\item{name}{string, The name of this resource.}
\item{description}{string, A description of this resource.}
\item{createdAt}{string, }
\item{updatedAt}{string, }
}
\description{
Create a resource in a permission set
}
