% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{permission_sets_put_archive}
\alias{permission_sets_put_archive}
\title{Update the archive status of this object}
\usage{
permission_sets_put_archive(id, status)
}
\arguments{
\item{id}{integer required. The ID of the object.}

\item{status}{boolean required. The desired archived status of the object.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID for this permission set.}
\item{name}{string, The name of this permission set.}
\item{description}{string, A description of this permission set.}
\item{author}{list, A list containing the following elements: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
\item{createdAt}{string, }
\item{updatedAt}{string, }
\item{archived}{string, The archival status of the requested item(s).}
}
\description{
Update the archive status of this object
}
