% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_ml_plot.R
\name{plot.civis_ml_classifier}
\alias{plot.civis_ml_classifier}
\title{Decile plot for classification with civis_ml}
\usage{
\method{plot}{civis_ml_classifier}(x, name = NULL, ...)
}
\arguments{
\item{x}{\code{civis_ml} object.}

\item{name}{Name of the class in a multiclass model to plot.}

\item{...}{unused.}
}
\value{
A \code{ggplot2} plot object is returned invisibly.
}
\description{
Decile plot for classification with civis_ml
}
