% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_ml_plot.R
\name{plot.civis_ml_regressor}
\alias{plot.civis_ml_regressor}
\title{Y-yhat plot for regression with civis_ml}
\usage{
\method{plot}{civis_ml_regressor}(x, ...)
}
\arguments{
\item{x}{\code{civis_ml} object}

\item{...}{unused}
}
\value{
A \code{ggplot2} plot object is returned invisibly.
}
\description{
Y-yhat plot for regression with civis_ml
}
