% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{predictions_delete_runs}
\alias{predictions_delete_runs}
\title{Cancel a run}
\usage{
predictions_delete_runs(id, run_id)
}
\arguments{
\item{id}{integer required. The ID of the prediction.}

\item{run_id}{integer required. The ID of the run.}
}
\value{
An empty HTTP response
}
\description{
Cancel a run
}
