% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{predictions_list}
\alias{predictions_list}
\title{List predictions}
\usage{
predictions_list(model_id = NULL)
}
\arguments{
\item{model_id}{integer optional. If specified, only return predictions associated with this model ID.}
}
\value{
An array containing the following fields:
\item{id}{integer, The ID of the prediction.}
\item{modelId}{integer, The ID of the model used for this prediction.}
\item{scoredTableId}{integer, The ID of the source table for this prediction.}
\item{scoredTableName}{string, The name of the source table for this prediction.}
\item{outputTableName}{string, The name of the output table for this prediction.}
\item{state}{string, The state of the last run of this prediction.}
\item{error}{string, The error, if any, of the last run of this prediction.}
\item{startedAt}{string, The start time of the last run of this prediction.}
\item{finishedAt}{string, The end time of the last run of this prediction.}
\item{lastRun}{object, A list containing the following elements: 
\itemize{
\item id integer, 
\item state string, 
\item createdAt string, The time that the run was queued.
\item startedAt string, The time that the run started.
\item finishedAt string, The time that the run completed.
\item error string, The error message for this run, if present.
}}
}
\description{
List predictions
}
