% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{predictions_post_runs}
\alias{predictions_post_runs}
\title{Start a run}
\usage{
predictions_post_runs(id)
}
\arguments{
\item{id}{integer required. The ID of the prediction.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the prediction run.}
\item{predictionId}{integer, The ID of the prediction.}
\item{state}{string, The state of the prediction run.}
\item{exception}{string, The exception, if any, returned by the prediction run.}
\item{name}{string, The name of table created by this predictions run.}
\item{createdAt}{string, The time when the table with created predictions was created.}
\item{scoreStats}{array, An array containing the following fields: 
\itemize{
\item scoreName string, The name of the score.
\item histogram array, The histogram of the distribution of scores.
\item avgScore number, The average score.
\item minScore number, The minimum score.
\item maxScore number, The maximum score.
}}
}
\description{
Start a run
}
