% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{predictions_put_schedules}
\alias{predictions_put_schedules}
\title{Schedule the prediction}
\usage{
predictions_put_schedules(id, schedule = NULL,
  score_on_model_build = NULL)
}
\arguments{
\item{id}{integer required. ID of the prediction associated with this schedule.}

\item{schedule}{list optional. A list containing the following elements: 
\itemize{
\item scheduled boolean, If the item is scheduled.
\item scheduledDays array, Day based on numeric value starting at 0 for Sunday.
\item scheduledHours array, Hours of the day it is scheduled on.
\item scheduledMinutes array, Minutes of the day it is scheduled on.
\item scheduledRunsPerHour integer, Alternative to scheduled minutes, number of times to run per hour.
}}

\item{score_on_model_build}{boolean optional. Whether the prediction will run after a rebuild of the associated model.}
}
\value{
A list containing the following elements:
\item{id}{integer, ID of the prediction associated with this schedule.}
\item{schedule}{list, A list containing the following elements: 
\itemize{
\item scheduled boolean, If the item is scheduled.
\item scheduledDays array, Day based on numeric value starting at 0 for Sunday.
\item scheduledHours array, Hours of the day it is scheduled on.
\item scheduledMinutes array, Minutes of the day it is scheduled on.
\item scheduledRunsPerHour integer, Alternative to scheduled minutes, number of times to run per hour.
}}
\item{scoreOnModelBuild}{boolean, Whether the prediction will run after a rebuild of the associated model.}
}
\description{
Schedule the prediction
}
