% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_base.R
\name{print.civis_api}
\alias{print.civis_api}
\title{Print results from a Civis API call}
\usage{
\method{print}{civis_api}(x, ...)
}
\arguments{
\item{x}{A \code{civis_api} response.}

\item{...}{Further arguments passed to \code{str}}
}
\value{
The `civis_api` object, invisibly.
}
\description{
Print results from a Civis API call
}
\examples{
\dontrun{
query_civis("SELECT * FROM schema.my_table", "database", preview_rows = 100)
csv_to_civis("file.csv", "my_database", "schema.my_table")
print(databases_list(), max = 5)
}
}
