% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{projects_delete}
\alias{projects_delete}
\title{Archive a project (deprecated, use the /archive endpoint instead)}
\usage{
projects_delete(project_id)
}
\arguments{
\item{project_id}{integer required.}
}
\value{
An empty HTTP response
}
\description{
Archive a project (deprecated, use the /archive endpoint instead)
}
