% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{projects_delete_parent_projects}
\alias{projects_delete_parent_projects}
\title{Remove an item from a Parent Project}
\usage{
projects_delete_parent_projects(id, parent_project_id)
}
\arguments{
\item{id}{integer required. The ID of the item.}

\item{parent_project_id}{integer required. The ID of the Parent Project.}
}
\value{
An empty HTTP response
}
\description{
Remove an item from a Parent Project
}
