% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{projects_put_parent_projects}
\alias{projects_put_parent_projects}
\title{Add an item to a Parent Project}
\usage{
projects_put_parent_projects(id, parent_project_id)
}
\arguments{
\item{id}{integer required. The ID of the item.}

\item{parent_project_id}{integer required. The ID of the Parent Project.}
}
\value{
An empty HTTP response
}
\description{
Add an item to a Parent Project
}
