% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{publish_addin}
\alias{publish_addin}
\title{Add in to Publish an R Markdown file to Platform Reports from RStudio}
\usage{
publish_addin()
}
\description{
Add in to Publish an R Markdown file to Platform Reports from RStudio
}
