% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{publish_html}
\alias{publish_html}
\title{Publish HTML to Platform Reports}
\usage{
publish_html(
  html_file,
  report_id = NULL,
  report_name = NULL,
  provide_api_key = NULL,
  project_id = NULL
)
}
\arguments{
\item{html_file}{string, HTML file}

\item{report_id}{integer, ID of an existing report, if provided, the
contents of the report will be replaced. If \code{report_id} is NULL, a new
report will be created.}

\item{report_name}{string, Title of report in platform}

\item{provide_api_key}{bool, Set to true to include API key in report.}

\item{project_id}{integer, Project_id that the report should be added to.}
}
\description{
Publish HTML to Platform Reports
}
\examples{
\dontrun{
# Uploads html and prints link to report in Platform
publish_html("my_beautiful_report.html")
}
}
\seealso{
\code{\link{publish_rmd}} to publish an R Markdown document
to Civis
}
