% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{queries_delete_runs}
\alias{queries_delete_runs}
\title{Cancel a run}
\usage{
queries_delete_runs(id, run_id)
}
\arguments{
\item{id}{integer required. The ID of the Query job.}

\item{run_id}{integer required. The ID of the run.}
}
\value{
An empty HTTP response
}
\description{
Cancel a run
}
