% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{queries_post_runs}
\alias{queries_post_runs}
\title{Start a run}
\usage{
queries_post_runs(id)
}
\arguments{
\item{id}{integer required. The ID of the query.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the run.}
\item{queryId}{integer, The ID of the query.}
\item{state}{string, The state of the run, one of 'queued' 'running' 'succeeded' 'failed' or 'cancelled'.}
\item{isCancelRequested}{boolean, True if run cancel requested, else false.}
\item{createdAt}{string, The time the run was created.}
\item{startedAt}{string, The time the run started at.}
\item{finishedAt}{string, The time the run completed.}
\item{error}{string, The error, if any, returned by the run.}
}
\description{
Start a run
}
