% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{refresh_table}
\alias{refresh_table}
\title{Refresh a table}
\usage{
refresh_table(tablename, database = NULL, verbose = FALSE)
}
\arguments{
\item{tablename}{string, Name of table and schema \code{"schema.tablename"}}

\item{database}{string, Name of database where data frame is to be uploaded. If no database is specified,
uses \code{options(civis.default_db)}}

\item{verbose}{bool, Set to TRUE to print intermediate progress indicators.}
}
\value{
Returns table meta-data from \code{\link{tables_get}}.
}
\description{
Refreshes a table on Redshift using \code{\link{tables_post_refresh}}, which runs the
table scanner and updates table meta-data.
}
\seealso{
Other tables: 
\code{\link{get_table_id}()},
\code{\link{transfer_table}()}
}
\concept{tables}
