% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{registrations_list}
\alias{registrations_list}
\title{List registrations}
\usage{
registrations_list(status = NULL, limit = NULL, page_num = NULL,
  order = NULL, order_dir = NULL)
}
\arguments{
\item{status}{string optional. If specified, returns registrations with one of these statuses. It accepts a comma-separated list, possible values are pending, confirmed, approved, rejected, provisioning, provisioned, notified, or decommissioned.}

\item{limit}{integer optional. Number of results to return. Defaults to 25. Maximum allowed is 100.}

\item{page_num}{integer optional. Page number of the results to return. Defaults to the first page, 1.}

\item{order}{string optional. The field on which to order the result set. Defaults to id. Must be one of: id, created_at.}

\item{order_dir}{string optional. Direction in which to sort, either asc (ascending) or desc (descending) defaulting to desc.}
}
\value{
An array containing the following fields:
\item{id}{integer, The ID of the registration.}
\item{firstName}{string, This user's first name.}
\item{lastName}{string, This user's last name.}
\item{username}{string, This user's username.}
\item{email}{string, This user's email address.}
\item{phone}{string, This user's phone.}
\item{organizationName}{string, The name of this user's organization.}
\item{status}{string, The status of the registration (pending/confirmed/approved/rejected/provisioning/provisioned/notified/decommissioned).}
\item{ipAddress}{string, The IP address used to create the registration.}
\item{createdAt}{string, The date and time when the registration was created.}
}
\description{
List registrations
}
