% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{registrations_post}
\alias{registrations_post}
\title{Register a new account}
\usage{
registrations_post(first_name, last_name, email, organization_name,
  username = NULL, phone = NULL, comment = NULL, mkto = NULL,
  utm_source = NULL, utm_medium = NULL, utm_campaign = NULL,
  utm_term = NULL, utm_content = NULL)
}
\arguments{
\item{first_name}{string required. This user's first name.}

\item{last_name}{string required. This user's last name.}

\item{email}{string required. This user's email address.}

\item{organization_name}{string required. The name of this user's organization.}

\item{username}{string optional. This user's username.}

\item{phone}{string optional. This user's phone.}

\item{comment}{array optional. Array of comments for multi-select inputs.}

\item{mkto}{string optional. The Marketo tracking cookie associated with a signup.}

\item{utm_source}{string optional. The utm_source parameter for this visit.}

\item{utm_medium}{string optional. The utm_medium parameter for this visit.}

\item{utm_campaign}{string optional. The utm_campaign parameter for this visit.}

\item{utm_term}{string optional. The utm_term parameter for this visit.}

\item{utm_content}{string optional. The utm_content parameter for this visit.}
}
\value{
An empty HTTP response
}
\description{
Register a new account
}
