% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{registrations_post_approve}
\alias{registrations_post_approve}
\title{Approve a registration}
\usage{
registrations_post_approve(id)
}
\arguments{
\item{id}{integer required. The ID of the registration.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the registration.}
\item{firstName}{string, This user's first name.}
\item{lastName}{string, This user's last name.}
\item{username}{string, This user's username.}
\item{email}{string, This user's email address.}
\item{phone}{string, This user's phone.}
\item{organizationName}{string, The name of this user's organization.}
\item{comment}{array, Array of comments for multi-select inputs.}
\item{status}{string, The status of the registration (pending/confirmed/approved/rejected/provisioning/provisioned/notified/decommissioned).}
\item{ipAddress}{string, The IP address used to create the registration.}
\item{createdAt}{string, The date and time when the registration was created.}
\item{confirmedAt}{string, The date and time when the registration was confirmed.}
\item{decommissionedAt}{string, The date and time when the registration was decommissioned.}
\item{marketoLeadId}{integer, The ID of the associated lead in Marketo, if it exists.}
\item{provisionJobId}{integer, The ID of the provisioning job, if it has been created.}
\item{organizationId}{integer, The ID of the free trial organization, if it has been created.}
\item{userId}{integer, The ID of the free trial user, if it has been created.}
}
\description{
Approve a registration
}
