% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{registrations_post_provision}
\alias{registrations_post_provision}
\title{Create an account for an approved registration}
\usage{
registrations_post_provision(id, server_id, bocce_cluster_id, org_slug = NULL,
  username = NULL)
}
\arguments{
\item{id}{integer required. The ID of the registration.}

\item{server_id}{integer required. The remote host id for the database that will be assigned to the organization}

\item{bocce_cluster_id}{integer required. The bocce cluster that will be assigned to the organization}

\item{org_slug}{string optional. The slug that will be used instead of platform's guess}

\item{username}{string optional. The username that will be used instead of what is in the registration}
}
\value{
An empty HTTP response
}
\description{
Create an account for an approved registration
}
