% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{remote_hosts_list_data_sets}
\alias{remote_hosts_list_data_sets}
\title{List data sets available from a remote host}
\usage{
remote_hosts_list_data_sets(id, credential_id = NULL, username = NULL,
  password = NULL, q = NULL, s = NULL)
}
\arguments{
\item{id}{integer required. The ID of the remote host.}

\item{credential_id}{integer optional. The credential ID.}

\item{username}{string optional. The user name for remote host.}

\item{password}{string optional. The password for remote host.}

\item{q}{string optional. The query string for data set.}

\item{s}{boolean optional. If true will only return schemas, otherwise, the results will be the full path.}
}
\value{
An array containing the following fields:
\item{name}{string, The path to a data_set.}
\item{fullPath}{boolean, Boolean that indicates whether further querying needs to be done before the table can be selected.}
}
\description{
List data sets available from a remote host
}
