% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{reports_delete_sql_projects}
\alias{reports_delete_sql_projects}
\title{Remove a SQL Report from a project}
\usage{
reports_delete_sql_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the SQL Report.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Remove a SQL Report from a project
}
