% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{reports_list_projects}
\alias{reports_list_projects}
\title{List the projects a Report belongs to}
\usage{
reports_list_projects(id, hidden = NULL)
}
\arguments{
\item{id}{integer required. The ID of the resource.}

\item{hidden}{boolean optional. If specified to be true, returns hidden items. Defaults to false, returning non-hidden items.}
}
\value{
An array containing the following fields:
\item{id}{integer, The ID for this project.}
\item{author}{list, A list containing the following elements: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
\item{name}{string, The name of this project.}
\item{description}{string, A description of the project.}
\item{users}{array, An array containing the following fields: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
\item{autoShare}{boolean, }
\item{createdAt}{string, }
\item{updatedAt}{string, }
\item{archived}{string, The archival status of the requested item(s).}
}
\description{
List the projects a Report belongs to
}
