% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{reports_patch_services}
\alias{reports_patch_services}
\title{Update some attributes of this service report}
\usage{
reports_patch_services(id, name = NULL, provide_api_key = NULL)
}
\arguments{
\item{id}{integer required. The ID of this report.}

\item{name}{string optional. The name of the service report.}

\item{provide_api_key}{boolean optional. Whether the report requests an API Key from the report viewer.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this report.}
\item{name}{string, The name of the report.}
\item{user}{list, A list containing the following elements: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
\item{createdAt}{string, The creation time for this report.}
\item{updatedAt}{string, The last updated at time for this report.}
\item{type}{string, The type of the report. One of: ReportTypes::HTML, ReportTypes::Tableau, ReportTypes::ShinyApp, ReportTypes::SQL}
\item{description}{string, The user-defined description of the report.}
\item{myPermissionLevel}{string, Your permission level on the object. One of "read", "write", or "manage".}
\item{archived}{string, The archival status of the requested item(s).}
\item{host}{string, The host for the service report}
\item{displayUrl}{string, The URL to display the service report.}
\item{serviceId}{integer, The id of the backing service}
\item{provideAPIKey}{boolean, Whether the report requests an API Key from the report viewer.}
\item{apiKey}{string, A Civis API key that can be used by this report.}
\item{apiKeyId}{integer, The ID of the API key. Can be used for auditing API use by this report.}
}
\description{
Update some attributes of this service report
}
