% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{reports_patch_solutions}
\alias{reports_patch_solutions}
\title{Update some attributes of this solutions report}
\usage{
reports_patch_solutions(id, name = NULL)
}
\arguments{
\item{id}{integer required. The ID of this report.}

\item{name}{string optional. The name of the solution report.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this report.}
\item{name}{string, The name of the report.}
\item{user}{object, A list containing the following elements: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
\item{createdAt}{string, }
\item{updatedAt}{string, }
\item{host}{string, The host for the solution}
\item{displayUrl}{string, The URL to display the solution report.}
\item{serviceId}{integer, The id of the backing service}
}
\description{
Update some attributes of this solutions report
}
