% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{reports_post_git_commits}
\alias{reports_post_git_commits}
\title{Commit and push a new version of the file}
\usage{
reports_post_git_commits(id, content, message, file_hash)
}
\arguments{
\item{id}{integer required. The ID of the item.}

\item{content}{string required. The contents to commit to the file.}

\item{message}{string required. A commit message describing the changes being made.}

\item{file_hash}{string required. The full SHA of the file being replaced.}
}
\value{
A list containing the following elements:
\item{content}{string, The file's contents.}
\item{type}{string, The file's type.}
\item{size}{integer, The file's size.}
\item{fileHash}{string, The SHA of the file.}
}
\description{
Commit and push a new version of the file
}
