% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{reports_post_refresh}
\alias{reports_post_refresh}
\title{Refresh the data in this Tableau report}
\usage{
reports_post_refresh(id)
}
\arguments{
\item{id}{integer required. The ID of this report.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this report.}
\item{organization}{list, A list containing the following elements: 
\itemize{
\item id integer, The ID of this organization.
\item tableauRefreshUsage integer, The number of tableau refreshes used this month.
\item tableauRefreshLimit integer, The number of monthly tableau refreshes permitted to this organization.
\item tableauRefreshHistory array, The number of tableau refreshes used this month.
}}
}
\description{
Refresh the data in this Tableau report
}
