% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{reports_post_solutions}
\alias{reports_post_solutions}
\title{Create a solutions report}
\usage{
reports_post_solutions(shiny_app_id)
}
\arguments{
\item{shiny_app_id}{integer required. The id of the backing Shiny app}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this report.}
\item{name}{string, The name of the report.}
\item{user}{object, A list containing the following elements: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
\item{createdAt}{string, }
\item{updatedAt}{string, }
\item{host}{string, The host for the Shiny App}
\item{displayUrl}{string, The URL to display the Shiny-based report.}
\item{shinyAppId}{integer, The id of the backing Shiny app}
}
\description{
Create a solutions report
}
