% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{reports_put_services_projects}
\alias{reports_put_services_projects}
\title{Add a Service Report to a project}
\usage{
reports_put_services_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the Service Report.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Add a Service Report to a project
}
