% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{reports_put_sql_projects}
\alias{reports_put_sql_projects}
\title{Add a SQL Report to a project}
\usage{
reports_put_sql_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the SQL Report.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Add a SQL Report to a project
}
