% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_future.R
\name{resolved}
\alias{resolved}
\title{Check whether a CivisFuture has resolved.}
\usage{
resolved(future, ...)
}
\arguments{
\item{future}{CivisFuture object.}

\item{...}{unused for CivisFuture.}
}
\description{
Check whether a CivisFuture has resolved.
}
