% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{results_delete_grants}
\alias{results_delete_grants}
\title{Revoke permission for this report to perform Civis platform API operations on your behalf}
\usage{
results_delete_grants(id)
}
\arguments{
\item{id}{integer required. The ID of this report.}
}
\value{
An empty HTTP response
}
\description{
Revoke permission for this report to perform Civis platform API operations on your behalf
}
