% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{results_delete_projects}
\alias{results_delete_projects}
\title{Remove a Report from a project}
\usage{
results_delete_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the Report.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Remove a Report from a project
}
