% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{results_delete_services_shares_users}
\alias{results_delete_services_shares_users}
\title{Revoke the permissions a user has on this object}
\usage{
results_delete_services_shares_users(id, user_id)
}
\arguments{
\item{id}{integer required. The ID of the resource that is shared.}

\item{user_id}{integer required. The ID of the user.}
}
\value{
An empty HTTP response
}
\description{
Revoke the permissions a user has on this object
}
