% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{results_get_git_commits}
\alias{results_get_git_commits}
\title{Get file contents at commit_hash}
\usage{
results_get_git_commits(id, commit_hash)
}
\arguments{
\item{id}{integer required. The ID of the file.}

\item{commit_hash}{string required. The SHA (full or shortened) of the desired git commit.}
}
\value{
A list containing the following elements:
\item{content}{string, The file's contents.}
\item{type}{string, The file's type.}
\item{size}{integer, The file's size.}
\item{fileHash}{string, The SHA of the file.}
}
\description{
Get file contents at commit_hash
}
