% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{roles_list}
\alias{roles_list}
\title{List Roles}
\usage{
roles_list(limit = NULL, page_num = NULL, order = NULL, order_dir = NULL)
}
\arguments{
\item{limit}{integer optional. Number of results to return. Defaults to 50. Maximum allowed is 1000.}

\item{page_num}{integer optional. Page number of the results to return. Defaults to the first page, 1.}

\item{order}{string optional. The field on which to order the result set. Defaults to id. Must be one of: id.}

\item{order_dir}{string optional. Direction in which to sort, either asc (ascending) or desc (descending) defaulting to asc.}
}
\value{
An array containing the following fields:
\item{id}{integer, ID of the Role.}
\item{name}{string, The name of the Role.}
\item{slug}{string, The slug.}
\item{description}{string, The description of the Role.}
}
\description{
List Roles
}
