% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{run_template}
\alias{run_template}
\title{Run a template script}
\usage{
run_template(id, arguments, ...)
}
\arguments{
\item{id}{id of the template script.}

\item{arguments}{list of arguments to the script.}

\item{...}{additional arguments to \code{scripts_post_custom}}
}
\value{
File ids of any run outputs are returned.
}
\description{
Run a template script
}
\examples{
\dontrun{
# Try a search for the template id
search_list('template name', type = 'template_script')

# Run the template
run_template(id, arguments = list(arg1 = 1, arg2 = 2), ...)
}
}
\seealso{
Other script_utils: \code{\link{civis_script}},
  \code{\link{fetch_output_file_ids}},
  \code{\link{run_civis}}
}
\concept{script_utils}
