% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_delete}
\alias{scripts_delete}
\title{Archive a script (deprecated, use archive endpoints instead)}
\usage{
scripts_delete(id)
}
\arguments{
\item{id}{integer required. The ID for the script.}
}
\value{
An empty HTTP response
}
\description{
Archive a script (deprecated, use archive endpoints instead)
}
