% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_delete_custom}
\alias{scripts_delete_custom}
\title{Archive a Custom Script (deprecated, use archiving endpoints instead)}
\usage{
scripts_delete_custom(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
An empty HTTP response
}
\description{
Archive a Custom Script (deprecated, use archiving endpoints instead)
}
