% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_delete_javascript}
\alias{scripts_delete_javascript}
\title{Archive a JavaScript Script (deprecated, use archiving endpoints instead)}
\usage{
scripts_delete_javascript(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
An empty HTTP response
}
\description{
Archive a JavaScript Script (deprecated, use archiving endpoints instead)
}
