% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_delete_python3}
\alias{scripts_delete_python3}
\title{Archive a Python Script (deprecated, use archiving endpoints instead)}
\usage{
scripts_delete_python3(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
An empty HTTP response
}
\description{
Archive a Python Script (deprecated, use archiving endpoints instead)
}
