% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_delete_r}
\alias{scripts_delete_r}
\title{Archive an R Script (deprecated, use archiving endpoints instead)}
\usage{
scripts_delete_r(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
An empty HTTP response
}
\description{
Archive an R Script (deprecated, use archiving endpoints instead)
}
