% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_delete_r_projects}
\alias{scripts_delete_r_projects}
\title{Remove an R Script from a project}
\usage{
scripts_delete_r_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the R Script.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Remove an R Script from a project
}
