% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_delete_r_runs}
\alias{scripts_delete_r_runs}
\title{Cancel a run}
\usage{
scripts_delete_r_runs(id, run_id)
}
\arguments{
\item{id}{integer required. The ID of the r.}

\item{run_id}{integer required. The ID of the run.}
}
\value{
An empty HTTP response
}
\description{
Cancel a run
}
