% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_delete_sql}
\alias{scripts_delete_sql}
\title{Archive a SQL script (deprecated, use archiving endpoints instead)}
\usage{
scripts_delete_sql(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
An empty HTTP response
}
\description{
Archive a SQL script (deprecated, use archiving endpoints instead)
}
