% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_get_custom_runs}
\alias{scripts_get_custom_runs}
\title{Check status of a run}
\usage{
scripts_get_custom_runs(id, run_id)
}
\arguments{
\item{id}{integer required. The ID of the Custom job.}

\item{run_id}{integer required. The ID of the run.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the run.}
\item{customId}{integer, The ID of the Custom job.}
\item{state}{string, The state of the run, one of 'queued' 'running' 'succeeded' 'failed' or 'cancelled'.}
\item{isCancelRequested}{boolean, True if run cancel requested, else false.}
\item{createdAt}{string, The time the run was created.}
\item{startedAt}{string, The time the run started at.}
\item{finishedAt}{string, The time the run completed.}
\item{error}{string, The error, if any, returned by the run.}
\item{maxMemoryUsage}{number, If the run has finished, the maximum amount of memory used during the run, in MB. Only available if the backing script is a Python, R, or container script.}
\item{maxCpuUsage}{number, If the run has finished, the maximum amount of cpu used during the run, in millicores. Only available if the backing script is a Python, R, or container script.}
}
\description{
Check status of a run
}
