% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_get_sql_runs}
\alias{scripts_get_sql_runs}
\title{Check status of a run}
\usage{
scripts_get_sql_runs(id, run_id)
}
\arguments{
\item{id}{integer required. The ID of the sql.}

\item{run_id}{integer required. The ID of the run.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this run.}
\item{sqlId}{integer, The ID of this sql.}
\item{state}{string, The state of this run.}
\item{isCancelRequested}{boolean, True if run cancel requested, else false.}
\item{startedAt}{string, The time the last run started.}
\item{finishedAt}{string, The time that this run finished.}
\item{error}{string, The error message for this run, if present.}
\item{output}{array, An array containing the following fields: 
\itemize{
\item outputName string, The name of the output file.
\item fileId integer, The unique ID of the output file.
\item path string, The temporary link to download this output file, valid for 36 hours.
}}
}
\description{
Check status of a run
}
