% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_list_dbt_runs_outputs}
\alias{scripts_list_dbt_runs_outputs}
\title{List the outputs for a run}
\usage{
scripts_list_dbt_runs_outputs(
  id,
  run_id,
  limit = NULL,
  page_num = NULL,
  order = NULL,
  order_dir = NULL
)
}
\arguments{
\item{id}{integer required. The ID of the dbt script.}

\item{run_id}{integer required. The ID of the run.}

\item{limit}{integer optional. Number of results to return. Defaults to its maximum of 50.}

\item{page_num}{integer optional. Page number of the results to return. Defaults to the first page, 1.}

\item{order}{string optional. The field on which to order the result set. Defaults to created_at. Must be one of: created_at, id.}

\item{order_dir}{string optional. Direction in which to sort, either asc (ascending) or desc (descending) defaulting to desc.}
}
\value{
An array containing the following fields:
\item{objectType}{string, The type of the output. Valid values are File, Table, Report, Project, Credential, or JSONValue}
\item{objectId}{integer, The ID of the output.}
\item{name}{string, The name of the output.}
\item{link}{string, The hypermedia link to the output.}
\item{value}{string, }
}
\description{
List the outputs for a run
}
