% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_list_sql_git}
\alias{scripts_list_sql_git}
\title{Get the git metadata attached to an item}
\usage{
scripts_list_sql_git(id)
}
\arguments{
\item{id}{integer required. The ID of the file.}
}
\value{
A list containing the following elements:
\item{gitRef}{string, A git reference specifying an unambiguous version of the file. Can be a branch name, tag or the full or shortened SHA of a commit.}
\item{gitBranch}{string, The git branch that the file is on.}
\item{gitPath}{string, The path of the file in the repository.}
\item{gitRepo}{list, A list containing the following elements: 
\itemize{
\item id integer, The ID for this git repository.
\item repoUrl string, The URL for this git repository.
\item createdAt string, 
\item updatedAt string, 
}}
\item{gitRefType}{string, Specifies if the file is versioned by branch or tag.}
\item{pullFromGit}{boolean, Automatically pull latest commit from git. Only works for scripts and workflows (assuming you have the feature enabled)}
}
\description{
Get the git metadata attached to an item
}
