% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_patch_containers_runs}
\alias{scripts_patch_containers_runs}
\title{Update a run}
\usage{
scripts_patch_containers_runs(id, run_id, state = NULL,
  bocce_accepted_at = NULL, bocce_started_at = NULL)
}
\arguments{
\item{id}{integer required. The ID for the script.}

\item{run_id}{integer required. The ID of the script run.}

\item{state}{string optional. The state of the script.}

\item{bocce_accepted_at}{string optional. The time when a bocce worker began processing the script.}

\item{bocce_started_at}{string optional. The time when a bocce worker began executing the script.}
}
\value{
An empty HTTP response
}
\description{
Update a run
}
