% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_patch_python3}
\alias{scripts_patch_python3}
\title{Update some attributes of this Python Script}
\usage{
scripts_patch_python3(
  id,
  name = NULL,
  parent_id = NULL,
  user_context = NULL,
  params = NULL,
  arguments = NULL,
  schedule = NULL,
  notifications = NULL,
  next_run_at = NULL,
  time_zone = NULL,
  target_project_id = NULL,
  required_resources = NULL,
  instance_type = NULL,
  cancel_timeout = NULL,
  docker_image_tag = NULL,
  partition_label = NULL,
  running_as_id = NULL,
  source = NULL
)
}
\arguments{
\item{id}{integer required. The ID for the script.}

\item{name}{string optional. The name of the script.}

\item{parent_id}{integer optional. The ID of the parent job that will trigger this script}

\item{user_context}{string optional. "runner" or "author", who to execute the script as when run as a template.}

\item{params}{array optional. An array containing the following fields: 
\itemize{
\item name string, The variable's name as used within your code.
\item label string, The label to present to users when asking them for the value.
\item description string, A short sentence or fragment describing this parameter to the end user.
\item type string, The type of parameter. Valid options: string, multi_line_string, integer, float, bool, file, table, database, credential_aws, credential_redshift, or credential_custom
\item required boolean, Whether this param is required.
\item value string, The value you would like to set this param to. Setting this value makes this parameter a fixed param.
\item default string, If an argument for this parameter is not defined, it will use this default value. Use true, True, t, y, yes, or 1 for true bool's or false, False, f, n, no, or 0 for false bool's. Cannot be used for parameters that are required or a credential type.
\item allowedValues array, The possible values this parameter can take, effectively making this an enumerable parameter. Allowed values is an array of hashes of the following format: `{label: 'Import', 'value': 'import'}`
}}

\item{arguments}{list optional. Parameter-value pairs to use when running this script. Only settable if this script has defined parameters.}

\item{schedule}{list optional. A list containing the following elements: 
\itemize{
\item scheduled boolean, If the item is scheduled.
\item scheduledDays array, Days of the week, based on numeric value starting at 0 for Sunday. Mutually exclusive with scheduledDaysOfMonth
\item scheduledHours array, Hours of the day it is scheduled on.
\item scheduledMinutes array, Minutes of the day it is scheduled on.
\item scheduledRunsPerHour integer, Deprecated in favor of scheduled minutes.
\item scheduledDaysOfMonth array, Days of the month it is scheduled on, mutually exclusive with scheduledDays.
}}

\item{notifications}{list optional. A list containing the following elements: 
\itemize{
\item urls array, URLs to receive a POST request at job completion
\item successEmailSubject string, Custom subject line for success e-mail.
\item successEmailBody string, Custom body text for success e-mail, written in Markdown.
\item successEmailAddresses array, Addresses to notify by e-mail when the job completes successfully.
\item successEmailFromName string, Name from which success emails are sent; defaults to "Civis."
\item successEmailReplyTo string, Address for replies to success emails; defaults to the author of the job.
\item failureEmailAddresses array, Addresses to notify by e-mail when the job fails.
\item stallWarningMinutes integer, Stall warning emails will be sent after this amount of minutes.
\item successOn boolean, If success email notifications are on. Defaults to user's preferences.
\item failureOn boolean, If failure email notifications are on. Defaults to user's preferences.
}}

\item{next_run_at}{string optional. The time of the next scheduled run.}

\item{time_zone}{string optional. The time zone of this script.}

\item{target_project_id}{integer optional. Target project to which script outputs will be added.}

\item{required_resources}{list optional. A list containing the following elements: 
\itemize{
\item cpu integer, The number of CPU shares to allocate for the container. Each core has 1000 shares. Must be at least 2 shares.
\item memory integer, The amount of RAM to allocate for the container (in MB). Must be at least 4 MB.
\item diskSpace number, The amount of disk space, in GB, to allocate for the container. This space will be used to hold the git repo configured for the container and anything your container writes to /tmp or /data. Fractional values (e.g. 0.25) are supported.
}}

\item{instance_type}{string optional. The EC2 instance type to deploy to. Only available for jobs running on kubernetes.}

\item{cancel_timeout}{integer optional. The amount of time (in seconds) to wait before forcibly terminating the script. When the script is cancelled, it is first sent a TERM signal. If the script is still running after the timeout, it is sent a KILL signal. Defaults to 0.}

\item{docker_image_tag}{string optional. The tag of the docker image to pull from DockerHub.}

\item{partition_label}{string optional. The partition label used to run this object.}

\item{running_as_id}{integer optional. The ID of the runner of this script.}

\item{source}{string optional. The body/text of the script.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID for the script.}
\item{name}{string, The name of the script.}
\item{type}{string, The type of the script (e.g SQL, Container, Python, R, JavaScript, dbt)}
\item{createdAt}{string, The time this script was created.}
\item{updatedAt}{string, The time the script was last updated.}
\item{author}{list, A list containing the following elements: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
\item{state}{string, The status of the script's last run.}
\item{finishedAt}{string, The time that the script's last run finished.}
\item{category}{string, The category of the script.}
\item{projects}{array, An array containing the following fields: 
\itemize{
\item id integer, The ID for the project.
\item name string, The name of the project.
}}
\item{parentId}{integer, The ID of the parent job that will trigger this script}
\item{userContext}{string, "runner" or "author", who to execute the script as when run as a template.}
\item{params}{array, An array containing the following fields: 
\itemize{
\item name string, The variable's name as used within your code.
\item label string, The label to present to users when asking them for the value.
\item description string, A short sentence or fragment describing this parameter to the end user.
\item type string, The type of parameter. Valid options: string, multi_line_string, integer, float, bool, file, table, database, credential_aws, credential_redshift, or credential_custom
\item required boolean, Whether this param is required.
\item value string, The value you would like to set this param to. Setting this value makes this parameter a fixed param.
\item default string, If an argument for this parameter is not defined, it will use this default value. Use true, True, t, y, yes, or 1 for true bool's or false, False, f, n, no, or 0 for false bool's. Cannot be used for parameters that are required or a credential type.
\item allowedValues array, The possible values this parameter can take, effectively making this an enumerable parameter. Allowed values is an array of hashes of the following format: `{label: 'Import', 'value': 'import'}`
}}
\item{arguments}{list, Parameter-value pairs to use when running this script. Only settable if this script has defined parameters.}
\item{isTemplate}{boolean, Whether others scripts use this one as a template.}
\item{publishedAsTemplateId}{integer, The ID of the template that this script is backing.}
\item{fromTemplateId}{integer, The ID of the template this script uses, if any.}
\item{templateDependentsCount}{integer, How many other scripts use this one as a template.}
\item{templateScriptName}{string, The name of the template script.}
\item{links}{list, A list containing the following elements: 
\itemize{
\item details string, The details link to get more information about the script.
\item runs string, The runs link to get the run information list for this script.
}}
\item{schedule}{list, A list containing the following elements: 
\itemize{
\item scheduled boolean, If the item is scheduled.
\item scheduledDays array, Days of the week, based on numeric value starting at 0 for Sunday. Mutually exclusive with scheduledDaysOfMonth
\item scheduledHours array, Hours of the day it is scheduled on.
\item scheduledMinutes array, Minutes of the day it is scheduled on.
\item scheduledRunsPerHour integer, Deprecated in favor of scheduled minutes.
\item scheduledDaysOfMonth array, Days of the month it is scheduled on, mutually exclusive with scheduledDays.
}}
\item{notifications}{list, A list containing the following elements: 
\itemize{
\item urls array, URLs to receive a POST request at job completion
\item successEmailSubject string, Custom subject line for success e-mail.
\item successEmailBody string, Custom body text for success e-mail, written in Markdown.
\item successEmailAddresses array, Addresses to notify by e-mail when the job completes successfully.
\item successEmailFromName string, Name from which success emails are sent; defaults to "Civis."
\item successEmailReplyTo string, Address for replies to success emails; defaults to the author of the job.
\item failureEmailAddresses array, Addresses to notify by e-mail when the job fails.
\item stallWarningMinutes integer, Stall warning emails will be sent after this amount of minutes.
\item successOn boolean, If success email notifications are on. Defaults to user's preferences.
\item failureOn boolean, If failure email notifications are on. Defaults to user's preferences.
}}
\item{runningAs}{list, A list containing the following elements: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
\item{nextRunAt}{string, The time of the next scheduled run.}
\item{timeZone}{string, The time zone of this script.}
\item{lastRun}{list, A list containing the following elements: 
\itemize{
\item id integer, 
\item state string, 
\item createdAt string, The time that the run was queued.
\item startedAt string, The time that the run started.
\item finishedAt string, The time that the run completed.
\item error string, The error message for this run, if present.
}}
\item{myPermissionLevel}{string, Your permission level on the object. One of "read", "write", or "manage".}
\item{hidden}{boolean, The hidden status of the item.}
\item{targetProjectId}{integer, Target project to which script outputs will be added.}
\item{archived}{string, The archival status of the requested item(s).}
\item{requiredResources}{list, A list containing the following elements: 
\itemize{
\item cpu integer, The number of CPU shares to allocate for the container. Each core has 1000 shares. Must be at least 2 shares.
\item memory integer, The amount of RAM to allocate for the container (in MB). Must be at least 4 MB.
\item diskSpace number, The amount of disk space, in GB, to allocate for the container. This space will be used to hold the git repo configured for the container and anything your container writes to /tmp or /data. Fractional values (e.g. 0.25) are supported.
}}
\item{instanceType}{string, The EC2 instance type to deploy to. Only available for jobs running on kubernetes.}
\item{cancelTimeout}{integer, The amount of time (in seconds) to wait before forcibly terminating the script. When the script is cancelled, it is first sent a TERM signal. If the script is still running after the timeout, it is sent a KILL signal. Defaults to 0.}
\item{dockerImageTag}{string, The tag of the docker image to pull from DockerHub.}
\item{partitionLabel}{string, The partition label used to run this object. }
\item{runningAsId}{integer, The ID of the runner of this script.}
\item{source}{string, The body/text of the script.}
}
\description{
Update some attributes of this Python Script
}
