% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_patch_sql_runs}
\alias{scripts_patch_sql_runs}
\title{Update the given run}
\usage{
scripts_patch_sql_runs(id, run_id, error = NULL)
}
\arguments{
\item{id}{integer required. ID of the Job}

\item{run_id}{integer required. ID of the Run}

\item{error}{string optional. The error message to update}
}
\value{
An empty HTTP response
}
\description{
Update the given run
}
