% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_post_containers_runs_heartbeats}
\alias{scripts_post_containers_runs_heartbeats}
\title{Indicate that the given run is being handled}
\usage{
scripts_post_containers_runs_heartbeats(id, run_id)
}
\arguments{
\item{id}{integer required. The ID of the container.}

\item{run_id}{integer required. The ID of the run.}
}
\value{
An empty HTTP response
}
\description{
Indicate that the given run is being handled
}
