% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_post_containers_runs_logs}
\alias{scripts_post_containers_runs_logs}
\title{Add log messages}
\usage{
scripts_post_containers_runs_logs(id, run_id, message = NULL,
  level = NULL, messages = NULL, child_job_id = NULL)
}
\arguments{
\item{id}{integer required. The ID of the script.}

\item{run_id}{integer required. The ID of the script run.}

\item{message}{string optional. The log message to store.}

\item{level}{string optional. The log level of this message [default: info]}

\item{messages}{array optional. An array containing the following fields: 
\itemize{
\item message string, The log message to store.
\item level string, The log level of this message [default: info]
\item createdAt string, 
}}

\item{child_job_id}{integer optional. The ID of the child job the message came from.}
}
\value{
An empty HTTP response
}
\description{
Add log messages
}
