% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_post_dbt_runs_outputs}
\alias{scripts_post_dbt_runs_outputs}
\title{Add an output for a run}
\usage{
scripts_post_dbt_runs_outputs(id, run_id, object_type, object_id)
}
\arguments{
\item{id}{integer required. The ID of the dbt script.}

\item{run_id}{integer required. The ID of the run.}

\item{object_type}{string required. The type of the output. Valid values are File, Table, Report, Project, Credential, or JSONValue}

\item{object_id}{integer required. The ID of the output.}
}
\value{
A list containing the following elements:
\item{objectType}{string, The type of the output. Valid values are File, Table, Report, Project, Credential, or JSONValue}
\item{objectId}{integer, The ID of the output.}
\item{name}{string, The name of the output.}
\item{link}{string, The hypermedia link to the output.}
\item{value}{string, }
}
\description{
Add an output for a run
}
