% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_post_python3_git_checkout_latest}
\alias{scripts_post_python3_git_checkout_latest}
\title{Checkout latest commit on the current branch of a script or workflow}
\usage{
scripts_post_python3_git_checkout_latest(id)
}
\arguments{
\item{id}{integer required. The ID of the item.}
}
\value{
A list containing the following elements:
\item{content}{string, The file's contents.}
\item{type}{string, The file's type.}
\item{size}{integer, The file's size.}
\item{fileHash}{string, The SHA of the file.}
}
\description{
Checkout latest commit on the current branch of a script or workflow
}
