% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_post_run}
\alias{scripts_post_run}
\title{Run a SQL script (legacy)}
\usage{
scripts_post_run(id)
}
\arguments{
\item{id}{integer required. The ID for the script.}
}
\value{
An empty HTTP response
}
\description{
Run a SQL script (legacy)
}
