% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_post_sql_git_checkout}
\alias{scripts_post_sql_git_checkout}
\title{Checkout content that the existing git_ref points to and save to the object}
\usage{
scripts_post_sql_git_checkout(id)
}
\arguments{
\item{id}{integer required. The ID of the file.}
}
\value{
A list containing the following elements:
\item{content}{string, The file's contents.}
\item{type}{string, The file's type.}
\item{size}{integer, The file's size.}
\item{fileHash}{string, The SHA of the file.}
}
\description{
Checkout content that the existing git_ref points to and save to the object
}
