% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_put_r_git}
\alias{scripts_put_r_git}
\title{Attach this r docker to a git repo/file}
\usage{
scripts_put_r_git(id, git_ref = NULL, git_branch = NULL, git_path = NULL,
  git_repo_url = NULL)
}
\arguments{
\item{id}{integer required. The ID of the object.}

\item{git_ref}{string optional. A git reference specifying an unambiguous version of the file. Can be a branch name, or the full or shortened SHA of a commit.}

\item{git_branch}{string optional. The git branch that the file is on.}

\item{git_path}{string optional. The path of the file in the repository.}

\item{git_repo_url}{string optional. The URL of the git repository.}
}
\value{
A list containing the following elements:
\item{gitRef}{string, A git reference specifying an unambiguous version of the file. Can be a branch name, or the full or shortened SHA of a commit.}
\item{gitBranch}{string, The git branch that the file is on.}
\item{gitPath}{string, The path of the file in the repository.}
\item{gitRepo}{list, A list containing the following elements: 
\itemize{
\item id integer, The ID for this git repository.
\item repoUrl string, The URL for this git repository.
\item createdAt string, 
\item updatedAt string, 
}}
}
\description{
Attach this r docker to a git repo/file
}
