% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{search_list_types}
\alias{search_list_types}
\title{List available search types}
\usage{
search_list_types()
}
\value{
An array containing the following fields:
\item{type}{string, The name of the item type.}
}
\description{
List available search types
}
