% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{services_delete}
\alias{services_delete}
\title{Archive a Service (deprecated, use archiving endpoints instead)}
\usage{
services_delete(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
An empty HTTP response
}
\description{
Archive a Service (deprecated, use archiving endpoints instead)
}
