% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{services_delete_projects}
\alias{services_delete_projects}
\title{Remove a ShinyApp from a project}
\usage{
services_delete_projects(id, project_id)
}
\arguments{
\item{id}{integer required. ID of the resource}

\item{project_id}{integer required. The ID of the project}
}
\value{
An empty HTTP response
}
\description{
Remove a ShinyApp from a project
}
