% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{services_delete_tokens}
\alias{services_delete_tokens}
\title{Revoke a token by id}
\usage{
services_delete_tokens(id, token_id)
}
\arguments{
\item{id}{integer required. The ID of the service.}

\item{token_id}{integer required. The ID of the token.}
}
\value{
An empty HTTP response
}
\description{
Revoke a token by id
}
