% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{services_list_tokens}
\alias{services_list_tokens}
\title{List tokens}
\usage{
services_list_tokens(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
An array containing the following fields:
\item{id}{integer, The ID of token.}
\item{name}{string, The name of token.}
\item{createdAt}{string, The date and time when the token was created.}
}
\description{
List tokens
}
