% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{services_post_tokens}
\alias{services_post_tokens}
\title{Create a new long-lived service token}
\usage{
services_post_tokens(id, name)
}
\arguments{
\item{id}{integer required. The ID of token.}

\item{name}{string required. The name of token.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of token.}
\item{name}{string, The name of token.}
\item{createdAt}{string, The date and time when the token was created.}
\item{token}{string, The value for a long-lived arproxy service token.}
}
\description{
Create a new long-lived service token
}
