% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{services_put_projects}
\alias{services_put_projects}
\title{Add a Service to a project}
\usage{
services_put_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the Service.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Add a Service to a project
}
