% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{shiny_apps_delete}
\alias{shiny_apps_delete}
\title{Archive a ShinyApp (deprecated, use archiving endpoints instead)}
\usage{
shiny_apps_delete(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
An empty HTTP response
}
\description{
Archive a ShinyApp (deprecated, use archiving endpoints instead)
}
