% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{shiny_apps_delete_deployments}
\alias{shiny_apps_delete_deployments}
\title{Delete a ShinyApp deployment}
\usage{
shiny_apps_delete_deployments(shiny_app_id, deployment_id)
}
\arguments{
\item{shiny_app_id}{integer required. The ID of the owning ShinyApp}

\item{deployment_id}{integer required. The ID for this deployment}
}
\value{
An empty HTTP response
}
\description{
Delete a ShinyApp deployment
}
