% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{shiny_apps_patch}
\alias{shiny_apps_patch}
\title{Update some attributes of this ShinyApp}
\usage{
shiny_apps_patch(id, name = NULL, docker_image_name = NULL,
  docker_image_tag = NULL, always_on = NULL, replicas = NULL,
  memory = NULL, cpu = NULL, credentials = NULL,
  running_as_username = NULL, api_key_id = NULL, git_repo_url = NULL,
  git_repo_ref = NULL, environment_variables = NULL)
}
\arguments{
\item{id}{integer required. The ID for this Shiny app.}

\item{name}{string optional. The name of this Shiny app.}

\item{docker_image_name}{string optional. The name of the docker image to pull from DockerHub.}

\item{docker_image_tag}{string optional. The tag of the docker image to pull from DockerHub (default: latest).}

\item{always_on}{boolean optional.}

\item{replicas}{integer optional. The number of Shiny app replicas to deploy.}

\item{memory}{integer optional. The amount of memory allocated to each replica of the Shiny app.}

\item{cpu}{integer optional. The amount of cpu allocated to each replica of the the Shiny app.}

\item{credentials}{array optional. A list of credential IDs to pass to the shiny app.}

\item{running_as_username}{string optional. Username of user running the shiny app.}

\item{api_key_id}{integer optional. API key id of user}

\item{git_repo_url}{string optional. The url for the git repo where the Shiny app code lives.}

\item{git_repo_ref}{string optional. The git reference to use when pulling code from the repo.}

\item{environment_variables}{object optional. Environment Variables to be passed into the Shiny App.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID for this Shiny app.}
\item{name}{string, The name of this Shiny app.}
\item{user}{object, A list containing the following elements: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
\item{dockerImageName}{string, The name of the docker image to pull from DockerHub.}
\item{dockerImageTag}{string, The tag of the docker image to pull from DockerHub (default: latest).}
\item{alwaysOn}{boolean, }
\item{replicas}{integer, The number of Shiny app replicas to deploy.}
\item{memory}{integer, The amount of memory allocated to each replica of the Shiny app.}
\item{cpu}{integer, The amount of cpu allocated to each replica of the the Shiny app.}
\item{createdAt}{string, }
\item{updatedAt}{string, }
\item{credentials}{array, A list of credential IDs to pass to the shiny app.}
\item{runningAsUsername}{string, Username of user running the shiny app.}
\item{apiKeyId}{integer, API key id of user}
\item{gitRepoUrl}{string, The url for the git repo where the Shiny app code lives.}
\item{gitRepoRef}{string, The git reference to use when pulling code from the repo.}
\item{reportId}{integer, The ID of the associated report.}
\item{currentDeployment}{object, A list containing the following elements: 
\itemize{
\item deploymentId integer, The ID for this deployment.
\item userId integer, The ID of the owner.
\item host string, Domain of the deployment.
\item name string, Name of the deployment.
\item dockerImageName string, The name of the docker image to pull from DockerHub.
\item dockerImageTag string, The tag of the docker image to pull from DockerHub (default: latest).
\item displayUrl string, URL that displays the deployed platform object and includes auth token.
\item memory integer, The memory allocated to the deployment.
\item cpu integer, The cpu allocated to the deployment.
\item state string, The state of the deployment.
\item stateMessage string, A detailed description of the state.
\item createdAt string, 
\item updatedAt string, 
\item published boolean, 
\item shinyAppId integer, The ID of owning Shiny App
}}
\item{previewDeployment}{object, A list containing the following elements: 
\itemize{
\item deploymentId integer, The ID for this deployment.
\item userId integer, The ID of the owner.
\item host string, Domain of the deployment.
\item name string, Name of the deployment.
\item dockerImageName string, The name of the docker image to pull from DockerHub.
\item dockerImageTag string, The tag of the docker image to pull from DockerHub (default: latest).
\item displayUrl string, URL that displays the deployed platform object and includes auth token.
\item memory integer, The memory allocated to the deployment.
\item cpu integer, The cpu allocated to the deployment.
\item state string, The state of the deployment.
\item stateMessage string, A detailed description of the state.
\item createdAt string, 
\item updatedAt string, 
\item published boolean, 
\item shinyAppId integer, The ID of owning Shiny App
}}
\item{environmentVariables}{object, Environment Variables to be passed into the Shiny App.}
\item{archived}{string, The archival status of the requested object(s).}
\item{hidden}{boolean, The hidden status of the object.}
}
\description{
Update some attributes of this ShinyApp
}
