% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{shiny_apps_patch_deployments}
\alias{shiny_apps_patch_deployments}
\title{Patch a ShinyApp deployment}
\usage{
shiny_apps_patch_deployments(shiny_app_id, deployment_id, state = NULL,
  state_message = NULL)
}
\arguments{
\item{shiny_app_id}{integer required. The ID of the owning ShinyApp}

\item{deployment_id}{integer required. The ID for this deployment}

\item{state}{string optional. The state of the deployment}

\item{state_message}{string optional. A detailed description of the state}
}
\value{
A list containing the following elements:
\item{deploymentId}{integer, The ID for this deployment.}
\item{shinyAppId}{integer, The ID of owning Shiny App}
}
\description{
Patch a ShinyApp deployment
}
