% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{shiny_apps_put_projects}
\alias{shiny_apps_put_projects}
\title{Add a ShinyApp to a project}
\usage{
shiny_apps_put_projects(id, project_id)
}
\arguments{
\item{id}{integer required. ID of the resource}

\item{project_id}{integer required. The ID of the project}
}
\value{
An empty HTTP response
}
\description{
Add a ShinyApp to a project
}
