% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{sql_escape_ident.CivisConnection}
\alias{sql_escape_ident.CivisConnection}
\title{Escape SQL identifiers.}
\usage{
\method{sql_escape_ident}{CivisConnection}(con, x)
}
\description{
This function is used to properly escape SQL identifiers such as table
or column names.
}
\keyword{internal}
