% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{sql_translate_env.CivisConnection}
\alias{sql_translate_env.CivisConnection}
\title{Translate R functions to their SQL (Redshift flavor) equivalent.}
\usage{
\method{sql_translate_env}{CivisConnection}(x)
}
\description{
Translate R functions to their SQL (Redshift flavor) equivalent.
}
\keyword{internal}
