% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{start_import_job}
\alias{start_import_job}
\title{Kick off a job to send data to the civis platform}
\usage{
start_import_job(database, tablename, if_exists, distkey, sortkey1,
  sortkey2, max_errors, hidden, diststyle, header, credential_id, ...)
}
\arguments{
\item{...}{other args to imports_post_files}
}
\description{
Kick off a job to send data to the civis platform
}
