% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{storage_hosts_delete}
\alias{storage_hosts_delete}
\title{Delete a storage host (deprecated)}
\usage{
storage_hosts_delete(id)
}
\arguments{
\item{id}{integer required. The ID of the storage host.}
}
\value{
An empty HTTP response
}
\description{
Delete a storage host (deprecated)
}
