% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{storage_hosts_get}
\alias{storage_hosts_get}
\title{Get a storage host}
\usage{
storage_hosts_get(id)
}
\arguments{
\item{id}{integer required. The ID of the storage host.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the storage host.}
\item{owner}{list, A list containing the following elements: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
\item{name}{string, The human readable name for the storage host.}
\item{provider}{string, The storage provider.One of: s3.}
\item{bucket}{string, The bucket for this storage host.}
\item{s3Options}{list, A list containing the following elements: 
\itemize{
\item region string, The region for this storage host (ex. "us-east-1")
}}
}
\description{
Get a storage host
}
