% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{storage_hosts_patch}
\alias{storage_hosts_patch}
\title{Update some attributes of this storage host}
\usage{
storage_hosts_patch(
  id,
  name = NULL,
  provider = NULL,
  bucket = NULL,
  s3_options = NULL
)
}
\arguments{
\item{id}{integer required. The ID of the storage host.}

\item{name}{string optional. The human readable name for the storage host.}

\item{provider}{string optional. The storage provider.One of: s3.}

\item{bucket}{string optional. The bucket for this storage host.}

\item{s3_options}{list optional. A list containing the following elements: 
\itemize{
\item region string, The region for this storage host (ex. "us-east-1")
}}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the storage host.}
\item{name}{string, The human readable name for the storage host.}
\item{provider}{string, The storage provider.One of: s3.}
\item{bucket}{string, The bucket for this storage host.}
\item{s3Options}{list, A list containing the following elements: 
\itemize{
\item region string, The region for this storage host (ex. "us-east-1")
}}
}
\description{
Update some attributes of this storage host
}
