% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{studios_delete_deployments}
\alias{studios_delete_deployments}
\title{Delete a Studio deployment}
\usage{
studios_delete_deployments(studio_id, deployment_id)
}
\arguments{
\item{studio_id}{integer required. The ID of the owning Studio}

\item{deployment_id}{integer required. The ID for this deployment}
}
\value{
An empty HTTP response
}
\description{
Delete a Studio deployment
}
