% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{studios_get}
\alias{studios_get}
\title{Get a Studio}
\usage{
studios_get(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the studio.}
\item{author}{list, A list containing the following elements: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
\item{name}{string, The name of the studio.}
\item{dockerImageName}{string, The name of the docker image to pull from DockerHub.}
\item{dockerImageTag}{string, The tag of the docker image to pull from DockerHub (default: latest).}
\item{instanceType}{string, The EC2 instance type to deploy to.}
\item{requiredResources}{list, A list containing the following elements: 
\itemize{
\item memory integer, The amount of memory allocated to the studio.
\item diskSpace number, The amount of disk space, in GB, to allocate for the studio. Fractional values (e.g. 0.25) are supported.
\item cpu integer, The amount of cpu allocated to the studio.
}}
\item{createdAt}{string, }
\item{updatedAt}{string, }
\item{mostRecentDeployment}{list, A list containing the following elements: 
\itemize{
\item deploymentId integer, The ID for this deployment.
\item userId integer, The ID of the owner.
\item host string, Domain of the deployment.
\item name string, Name of the deployment.
\item dockerImageName string, The name of the docker image to pull from DockerHub.
\item dockerImageTag string, The tag of the docker image to pull from DockerHub (default: latest).
\item displayUrl string, A signed URL for viewing the deployed item.
\item instanceType string, The EC2 instance type requested for the deployment.
\item memory integer, The memory allocated to the deployment, in MB.
\item cpu integer, The cpu allocated to the deployment, in millicores.
\item state string, The state of the deployment.
\item stateMessage string, A detailed description of the state.
\item maxMemoryUsage number, If the deployment has finished, the maximum amount of memory used during the deployment, in MB.
\item maxCpuUsage number, If the deployment has finished, the maximum amount of cpu used during the deployment, in millicores.
\item createdAt string, 
\item updatedAt string, 
\item studioId integer, The ID of the owning Studio
}}
\item{gitCredentialId}{integer, The id of the git credential to be used when checking out the specified git repo. If not supplied, the first git credential you've submitted will be used. }
\item{params}{array, An array containing the following fields: 
\itemize{
\item name string, The variable's name as used within your code.
\item description string, A short sentence or fragment describing this parameter.
\item type string, The type of parameter. Valid options: string, multi_line_string, integer, float, bool, file, table, database, credential_aws, credential_redshift, or credential_custom
\item value string, The value you would like to set this param to.
}}
\item{gitRepoId}{integer, The ID of the git repository.}
\item{gitRepoUrl}{string, The URL of the git repository (e.g., https://github.com/organization/repo_name.git).}
\item{gitRef}{string, The git reference if git repo is specified}
\item{partitionLabel}{string, The partition label used to run this object.}
\item{myPermissionLevel}{string, Your permission level on the object. One of "read", "write", or "manage".}
\item{archived}{string, The archival status of the requested item(s).}
}
\description{
Get a Studio
}
