% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{surveys_delete_samples}
\alias{surveys_delete_samples}
\title{Delete a sample}
\usage{
surveys_delete_samples(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
An empty HTTP response
}
\description{
Delete a sample
}
