% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{surveys_get}
\alias{surveys_get}
\title{Get a survey}
\usage{
surveys_get(id)
}
\arguments{
\item{id}{integer required. Survey ID}
}
\value{
A list containing the following elements:
\item{id}{integer, Survey ID}
\item{scriptVersion}{object, A list containing the following elements: 
\itemize{
\item id integer, 
\item script object, 
\item questions array, 
}}
\item{vendor}{object, A list containing the following elements: 
\itemize{
\item id integer, 
\item name string, 
}}
\item{sample}{object, A list containing the following elements: 
\itemize{
\item id integer, 
\item sampleTable object, 
\item vendorSampleTable object, 
\item strataTabsTable object, 
\item uniqueId string, 
\item householdColumn integer, 
\item basefile object, 
\item desiredCompletes integer, 
\item oversamplePercent integer, 
\item buckets array, 
\item contactMethods array, 
\item targetAudience string, 
\item populationLimitingSQL string, 
\item frameLimitingSQL string, 
\item parentId integer, 
\item metadata object, 
}}
\item{sampleSchemaTable}{string, }
\item{returns}{array, An array containing the following fields: 
\itemize{
\item id integer, 
\item dials integer, 
\item completes integer, 
\item matches integer, 
\item refusals integer, 
\item matchedCompletes integer, 
\item earlyTermination integer, 
\item other integer, 
\item status string, 
}}
\item{rollupJobId}{integer, }
\item{metadata}{object, Metadata}
}
\description{
Get a survey
}
