% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{surveys_get_projects}
\alias{surveys_get_projects}
\title{Get a summary of the metadata for a single project}
\usage{
surveys_get_projects(id)
}
\arguments{
\item{id}{integer required. The ID for this project.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID for this project.}
\item{name}{string, The name of this project.}
\item{projectType}{string, }
\item{status}{string, }
\item{client}{string, }
\item{notes}{string, }
\item{createdAt}{string, }
\item{updatedAt}{string, }
\item{fieldings}{array, An array containing the following fields: 
\itemize{
\item id integer, 
\item scriptVersion object, 
\item vendor object, 
\item sample object, 
\item sampleStatus string, 
\item scriptStatus string, 
\item scheduledStartDate string, 
\item scheduledEndDate string, 
\item completesNeeded integer, 
\item completes integer, 
\item mode string, 
\item costPerComplete number, 
\item callCenterLeadId integer, 
\item callIngestionLeadId integer, 
\item voterFileSource string, 
\item notes string, 
\item sampleSchema string, 
\item sampleTable string, 
\item matchedCompletes integer, 
\item returnDataId integer, 
\item returns array, 
\item rollupJobId integer, 
}}
\item{rollupJobId}{integer, }
\item{customImportTemplateId}{integer, }
}
\description{
Get a summary of the metadata for a single project
}
