% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{surveys_get_returns}
\alias{surveys_get_returns}
\title{Get a survey fielding return}
\usage{
surveys_get_returns(id, survey_id)
}
\arguments{
\item{id}{integer required. The ID of the return.}

\item{survey_id}{integer required. The ID of the survey.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the return.}
\item{surveyId}{integer, The ID of the survey.}
\item{fileId}{integer, The ID of the file object.}
\item{s3Uri}{string, URI of the S3 file}
\item{customImportJobId}{integer, }
\item{databaseTableId}{integer, Table containing imported survey return}
\item{metadata}{object, }
\item{status}{string, }
}
\description{
Get a survey fielding return
}
