% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{surveys_list}
\alias{surveys_list}
\title{List of surveys}
\usage{
surveys_list()
}
\value{
An array containing the following fields:
\item{id}{integer, Survey ID}
\item{vendor}{object, A list containing the following elements: 
\itemize{
\item id integer, 
\item name string, 
}}
\item{sampleSchemaTable}{string, }
}
\description{
List of surveys
}
