% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{surveys_list_projects}
\alias{surveys_list_projects}
\title{List survey projects}
\usage{
surveys_list_projects()
}
\value{
An array containing the following fields:
\item{id}{integer, The ID for this project.}
}
\description{
List survey projects
}
