% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{surveys_list_returns}
\alias{surveys_list_returns}
\title{List survey fielding returns}
\usage{
surveys_list_returns(survey_id)
}
\arguments{
\item{survey_id}{integer required. The ID of the survey.}
}
\value{
An array containing the following fields:
\item{id}{integer, The ID of the return.}
\item{status}{string, }
}
\description{
List survey fielding returns
}
