% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{surveys_list_samples}
\alias{surveys_list_samples}
\title{List samples}
\usage{
surveys_list_samples()
}
\value{
An array containing the following fields:
\item{id}{integer, }
\item{sampleTable}{object, A list containing the following elements: 
\itemize{
\item schema string, 
\item name string, 
}}
}
\description{
List samples
}
