% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{surveys_put_returns}
\alias{surveys_put_returns}
\title{Update a survey fielding return}
\usage{
surveys_put_returns(id, survey_id, database_table_id = NULL,
  server_name = NULL, schema = NULL, table_name = NULL, metadata = NULL,
  status = NULL)
}
\arguments{
\item{id}{integer required. The ID of the return.}

\item{survey_id}{integer required. The ID of the survey.}

\item{database_table_id}{integer optional. Table containing imported survey return}

\item{server_name}{string optional.}

\item{schema}{string optional.}

\item{table_name}{string optional.}

\item{metadata}{object optional.}

\item{status}{string optional.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the return.}
\item{surveyId}{integer, The ID of the survey.}
\item{fileId}{integer, The ID of the file object.}
\item{s3Uri}{string, URI of the S3 file}
\item{customImportJobId}{integer, }
\item{databaseTableId}{integer, Table containing imported survey return}
\item{metadata}{object, }
\item{status}{string, }
}
\description{
Update a survey fielding return
}
