% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{surveys_put_samples}
\alias{surveys_put_samples}
\title{Update a sample}
\usage{
surveys_put_samples(id, sample_table_id = NULL, server_name = NULL,
  schema = NULL, table_name = NULL, unique_id = NULL, metadata = NULL)
}
\arguments{
\item{id}{integer required.}

\item{sample_table_id}{integer optional.}

\item{server_name}{string optional.}

\item{schema}{string optional.}

\item{table_name}{string optional.}

\item{unique_id}{string optional.}

\item{metadata}{object optional.}
}
\value{
A list containing the following elements:
\item{id}{integer, }
\item{sampleTable}{object, A list containing the following elements: 
\itemize{
\item schema string, 
\item name string, 
}}
\item{vendorSampleTable}{object, A list containing the following elements: 
\itemize{
\item schema string, 
\item name string, 
}}
\item{strataTabsTable}{object, A list containing the following elements: 
\itemize{
\item schema string, 
\item name string, 
}}
\item{uniqueId}{string, }
\item{householdColumn}{integer, }
\item{basefile}{object, A list containing the following elements: 
\itemize{
\item id integer, 
\item name string, 
\item sourceTable object, 
\item remoteHost object, 
\item uniqueId string, 
\item bucketColumns array, 
\item contactabilityColumns array, 
\item audienceColumns array, 
\item createdAt string, 
\item updatedAt string, 
}}
\item{desiredCompletes}{integer, }
\item{oversamplePercent}{integer, }
\item{buckets}{array, An array containing the following fields: 
\itemize{
\item name string, 
\item column string, 
}}
\item{contactMethods}{array, }
\item{targetAudience}{string, }
\item{populationLimitingSQL}{string, }
\item{frameLimitingSQL}{string, }
\item{parentId}{integer, }
\item{metadata}{object, }
}
\description{
Update a sample
}
