% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{table_tags_delete}
\alias{table_tags_delete}
\title{Delete a Table Tag}
\usage{
table_tags_delete(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
An empty HTTP response
}
\description{
Delete a Table Tag
}
