% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{table_tags_get}
\alias{table_tags_get}
\title{Get a Table Tag}
\usage{
table_tags_get(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
A list containing the following elements:
\item{id}{integer, Table Tag ID}
\item{name}{string, Table Tag Name}
\item{createdAt}{string, The date the tag was created.}
\item{updatedAt}{string, The date the tag was recently updated on.}
\item{tableCount}{integer, The total number of tables associated with the tag.}
\item{user}{list, A list containing the following elements: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
}
\description{
Get a Table Tag
}
