% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{tables_delete_projects}
\alias{tables_delete_projects}
\title{Remove a Table from a project}
\usage{
tables_delete_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the Table.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Remove a Table from a project
}
