% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{tables_delete_tags}
\alias{tables_delete_tags}
\title{Add a tag to a table}
\usage{
tables_delete_tags(id, table_tag_id)
}
\arguments{
\item{id}{integer required. The ID of the table.}

\item{table_tag_id}{integer required. The ID of the tag.}
}
\value{
An empty HTTP response
}
\description{
Add a tag to a table
}
