% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{tables_get_enhancements_geocodings}
\alias{tables_get_enhancements_geocodings}
\title{View the status of a geocoding table enhancement}
\usage{
tables_get_enhancements_geocodings(id, source_table_id)
}
\arguments{
\item{id}{integer required. The ID of the enhancement.}

\item{source_table_id}{integer required. The ID of the table that was enhanced.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the enhancement.}
\item{sourceTableId}{integer, The ID of the table that was enhanced.}
\item{state}{string, The state of the enhancement, one of 'queued' 'running' 'succeeded' 'failed' or 'cancelled'.}
\item{enhancedTableSchema}{string, The schema name of the table created by the enhancement.}
\item{enhancedTableName}{string, The name of the table created by the enhancement.}
}
\description{
View the status of a geocoding table enhancement
}
