% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{tables_post}
\alias{tables_post}
\title{Import a file into a table}
\usage{
tables_post(database_id, schema, name, data)
}
\arguments{
\item{database_id}{integer required. The ID of the destination database.}

\item{schema}{string required. The destination schema name.}

\item{name}{string required. The destination table name, without the schema prefix.}

\item{data}{string required. The file to import, uploaded using HTTP multipart.}
}
\value{
A list containing the following elements:
\item{databaseId}{integer, The ID of the destination database.}
\item{schema}{string, The destination schema name.}
\item{name}{string, The destination table name, without the schema prefix.}
\item{state}{string, The state of the last run.}
\item{startedAt}{string, The start time of the last run.}
\item{finishedAt}{string, The end time of the last run.}
}
\description{
Import a file into a table
}
