% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{tables_post_enhance}
\alias{tables_post_enhance}
\title{Run an enhancement on a table, deprecated use "POST /tables/:source_table_id/enhancements/<enhancement-type>" instead}
\usage{
tables_post_enhance(id, job_type, metadata = NULL)
}
\arguments{
\item{id}{integer required. ID of table to enhance}

\item{job_type}{string required. Job type}

\item{metadata}{string optional.}
}
\value{
An empty HTTP response
}
\description{
Run an enhancement on a table, deprecated use "POST /tables/:source_table_id/enhancements/<enhancement-type>" instead
}
