% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{tables_post_scan}
\alias{tables_post_scan}
\title{Creates and enqueues a single table scanner job on a new table}
\usage{
tables_post_scan(database_id, schema, table_name, queue_stats = NULL)
}
\arguments{
\item{database_id}{integer required. The ID of the database.}

\item{schema}{string required. The name of the schema containing the table.}

\item{table_name}{string required. The name of the table.}

\item{queue_stats}{boolean optional. Whether or not the job should queue on syncing table stats}
}
\value{
A list containing the following elements:
\item{jobId}{integer, The ID of the job created.}
\item{runId}{integer, The ID of the run created.}
}
\description{
Creates and enqueues a single table scanner job on a new table
}
