% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{tables_put_projects}
\alias{tables_put_projects}
\title{Add a Table to a project}
\usage{
tables_put_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the Table.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Add a Table to a project
}
