% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{tables_put_tags}
\alias{tables_put_tags}
\title{Add a tag to a table}
\usage{
tables_put_tags(id, table_tag_id)
}
\arguments{
\item{id}{integer required. The ID of the table.}

\item{table_tag_id}{integer required. The ID of the tag.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the table.}
\item{tableTagId}{integer, The ID of the tag.}
}
\description{
Add a tag to a table
}
