% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{templates_delete_scripts}
\alias{templates_delete_scripts}
\title{Archive a Script Template (deprecated, use archiving endpoints instead)}
\usage{
templates_delete_scripts(id)
}
\arguments{
\item{id}{integer required.}
}
\value{
An empty HTTP response
}
\description{
Archive a Script Template (deprecated, use archiving endpoints instead)
}
