% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{templates_delete_scripts_shares_groups}
\alias{templates_delete_scripts_shares_groups}
\title{Revoke the permissions a group has on this object}
\usage{
templates_delete_scripts_shares_groups(id, group_id)
}
\arguments{
\item{id}{integer required. ID of the resource to be revoked}

\item{group_id}{integer required. ID of the group}
}
\value{
An empty HTTP response
}
\description{
Revoke the permissions a group has on this object
}
