% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{templates_list_scripts_shares}
\alias{templates_list_scripts_shares}
\title{List users and groups permissioned on this object}
\usage{
templates_list_scripts_shares(id)
}
\arguments{
\item{id}{integer required. The ID of the object.}
}
\value{
An array containing the following fields:
\item{readers}{object, A list containing the following elements: 
\itemize{
\item users array, 
\item groups array, 
}}
\item{writers}{object, A list containing the following elements: 
\itemize{
\item users array, 
\item groups array, 
}}
\item{owners}{object, A list containing the following elements: 
\itemize{
\item users array, 
\item groups array, 
}}
\item{totalUserShares}{integer, For owners, the number of total users shared. For writers and readers, the number of visible users shared.}
\item{totalGroupShares}{integer, For owners, the number of total groups shared. For writers and readers, the number of visible groups shared.}
}
\description{
List users and groups permissioned on this object
}
