% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{templates_patch_reports}
\alias{templates_patch_reports}
\title{Update some attributes of this Report Template}
\usage{
templates_patch_reports(id, name = NULL, category = NULL, archived = NULL,
  code_body = NULL, provide_api_key = NULL)
}
\arguments{
\item{id}{integer required.}

\item{name}{string optional. The name of the template.}

\item{category}{string optional. The category of this report template.  Can be left blank.  Acceptable values are: dataset-viz}

\item{archived}{boolean optional. Whether the template has been archived.}

\item{code_body}{string optional. The code for the Template body.}

\item{provide_api_key}{boolean optional. Whether reports based on this template request an API Key from the report viewer.}
}
\value{
A list containing the following elements:
\item{id}{integer, }
\item{name}{string, The name of the template.}
\item{category}{string, The category of this report template.  Can be left blank.  Acceptable values are: dataset-viz}
\item{createdAt}{string, }
\item{updatedAt}{string, }
\item{useCount}{integer, The number of uses of this template.}
\item{archived}{boolean, Whether the template has been archived.}
\item{author}{object, A list containing the following elements: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
\item{techReviewed}{boolean, Whether this template has been audited by Civis for security vulnerability and correctness.}
\item{authCodeUrl}{string, A URL to the template's stored code body.}
\item{provideAPIKey}{boolean, Whether reports based on this template request an API Key from the report viewer.}
\item{hidden}{boolean, The hidden status of the object.}
}
\description{
Update some attributes of this Report Template
}
