% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{templates_put_scripts}
\alias{templates_put_scripts}
\title{Replace all attributes of this Script Template}
\usage{
templates_put_scripts(id, name, note = NULL, archived = NULL)
}
\arguments{
\item{id}{integer required.}

\item{name}{string required. The name of the template.}

\item{note}{string optional. A note describing what this template is used for; custom scripts created off this template will display this description.}

\item{archived}{boolean optional. Whether the template has been archived.}
}
\value{
A list containing the following elements:
\item{id}{integer, }
\item{scriptId}{integer, The id of the script that this template uses.}
\item{userContext}{string, The user context of the script that this template uses.}
\item{name}{string, The name of the template.}
\item{category}{string, The category of this template.}
\item{note}{string, A note describing what this template is used for; custom scripts created off this template will display this description.}
\item{createdAt}{string, }
\item{updatedAt}{string, }
\item{useCount}{integer, The number of uses of this template.}
\item{techReviewed}{boolean, Whether this template has been audited by Civis for security vulnerability and correctness.}
\item{archived}{boolean, Whether the template has been archived.}
\item{hidden}{boolean, The hidden status of the object. Setting this to true hides it from most API endpoints. The object can still be queried directly by ID}
}
\description{
Replace all attributes of this Script Template
}
