% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{usage_get_llm}
\alias{usage_get_llm}
\title{Get an individual usage statistic for a given organization}
\usage{
usage_get_llm(id)
}
\arguments{
\item{id}{integer required. The ID of the usage statistic to get.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the usage statistic to get.}
\item{runId}{integer, The ID of the run which contributed this usage.}
\item{jobId}{integer, The ID of the job which contributed this usage.}
\item{userId}{integer, The ID of the user who contributed this usage.}
\item{organizationId}{integer, The organization of the user who contributed this usage.}
\item{runCreatedAt}{string, When the run was created at.}
\item{runTime}{integer, The duration of the run in seconds.}
\item{credits}{number, The number of credits used.}
\item{inputTokens}{integer, The number of tokens input to the run.}
\item{outputTokens}{integer, The number of tokens output from the run.}
\item{modelId}{string, The ID of the LLM model used.}
}
\description{
Get an individual usage statistic for a given organization
}
