% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{usage_limits_get_llm}
\alias{usage_limits_get_llm}
\title{Get a LLM Usage Limit}
\usage{
usage_limits_get_llm(id)
}
\arguments{
\item{id}{integer required. The ID for the limit.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID for the limit.}
\item{organizationId}{integer, The ID of the organization to which this limit belongs.}
\item{createdAt}{string, The time this limit was created.}
\item{updatedAt}{string, The time the limit was last updated.}
\item{hardLimit}{integer, The limit value. One of 1000, 10000, 50000, and 100000.}
}
\description{
Get a LLM Usage Limit
}
