% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{usage_limits_get_matching}
\alias{usage_limits_get_matching}
\title{Get a Matching Usage Limit}
\usage{
usage_limits_get_matching(id)
}
\arguments{
\item{id}{integer required. The ID for the limit.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID for the limit.}
\item{organizationId}{integer, The ID of the organization to which this limit belongs.}
\item{createdAt}{string, The time this limit was created.}
\item{updatedAt}{string, The time the limit was last updated.}
\item{hardLimit}{integer, The limit value. One of 50000000, 200000000, 500000000, 1000000000, and 2000000000.}
\item{task}{string, The category of this limit. One of 'IDR' or 'CDM'.}
\item{notificationEmails}{array, Addresses to notify by e-mail when the limit is reached.}
}
\description{
Get a Matching Usage Limit
}
