% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{usage_limits_list_llm}
\alias{usage_limits_list_llm}
\title{List LLM Usage Limits}
\usage{
usage_limits_list_llm(organization_id = NULL)
}
\arguments{
\item{organization_id}{integer optional. If specified, return limits for this organization only.}
}
\value{
An array containing the following fields:
\item{id}{integer, The ID for the limit.}
\item{organizationId}{integer, The ID of the organization to which this limit belongs.}
\item{createdAt}{string, The time this limit was created.}
\item{updatedAt}{string, The time the limit was last updated.}
\item{hardLimit}{integer, The limit value. One of 1000, 10000, 50000, and 100000.}
}
\description{
List LLM Usage Limits
}
