% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{usage_limits_list_matching}
\alias{usage_limits_list_matching}
\title{List Matching Usage Limits}
\usage{
usage_limits_list_matching(task = NULL)
}
\arguments{
\item{task}{string optional. If specified, return limits for this task type only. One of 'IDR' or 'CDM'.}
}
\value{
An array containing the following fields:
\item{id}{integer, The ID for the limit.}
\item{organizationId}{integer, The ID of the organization to which this limit belongs.}
\item{createdAt}{string, The time this limit was created.}
\item{updatedAt}{string, The time the limit was last updated.}
\item{hardLimit}{integer, The limit value. One of 50000000, 200000000, 500000000, 1000000000, and 2000000000.}
\item{task}{string, The category of this limit. One of 'IDR' or 'CDM'.}
\item{notificationEmails}{array, Addresses to notify by e-mail when the limit is reached.}
}
\description{
List Matching Usage Limits
}
