% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{usage_list_llm_organization_summary}
\alias{usage_list_llm_organization_summary}
\title{Get summarized usage statistics for a given organization}
\usage{
usage_list_llm_organization_summary(org_id, start_date = NULL, end_date = NULL)
}
\arguments{
\item{org_id}{integer required. The ID of the organization to get usage statistics for.}

\item{start_date}{string optional. The start date of the range to get usage statistics for."\ "Defaults to the start of the current month if neither start_date nor end_date is specified.}

\item{end_date}{string optional. The end date of the range to get usage statistics for."\ "Defaults to the end of the current day if neither start_date nor end_date is specified.}
}
\value{
A list containing the following elements:
\item{credits}{number, The number of credits used.}
\item{organizationId}{integer, The organization for which LLM usage statistics are summarized.}
}
\description{
Get summarized usage statistics for a given organization
}
