% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{users_delete_me_favorites}
\alias{users_delete_me_favorites}
\title{Unfavorite an item}
\usage{
users_delete_me_favorites(id)
}
\arguments{
\item{id}{integer required. The id of the favorite.}
}
\value{
An empty HTTP response
}
\description{
Unfavorite an item
}
