% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{users_get_me_themes}
\alias{users_get_me_themes}
\title{Show a theme}
\usage{
users_get_me_themes(id)
}
\arguments{
\item{id}{integer required. The ID of this theme.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this theme.}
\item{name}{string, The name of this theme.}
\item{organizationIds}{array, List of organization ID's allowed to use this theme.}
\item{settings}{string, The theme configuration object.}
\item{logoFile}{list, A list containing the following elements: 
\itemize{
\item id integer, The ID of the logo image file.
\item downloadUrl string, The URL of the logo image file.
}}
\item{createdAt}{string, }
\item{updatedAt}{string, }
}
\description{
Show a theme
}
