% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{users_list_me}
\alias{users_list_me}
\title{Show info about the logged-in user}
\usage{
users_list_me()
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this user.}
\item{name}{string, This user's name.}
\item{email}{string, This user's email address.}
\item{username}{string, This user's username.}
\item{initials}{string, This user's initials.}
\item{lastCheckedAnnouncements}{string, The date and time at which the user last checked their announcements.}
\item{featureFlags}{object, The feature flag settings for this user.}
\item{roles}{array, The roles this user has, listed by slug.}
\item{preferences}{object, This user's preferences.}
\item{customBranding}{string, The branding of Platform for this user.}
\item{groups}{array, An array containing the following fields: 
\itemize{
\item id integer, The ID of this group.
\item name string, The name of this group.
\item organizationId integer, The organization associated with this group.
}}
\item{organizationName}{string, The name of the organization the user belongs to.}
\item{createdAt}{string, The date and time when the user was created.}
\item{signInCount}{integer, The number of times the user has signed in.}
}
\description{
Show info about the logged-in user
}
