% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{users_list_me_organization_admins}
\alias{users_list_me_organization_admins}
\title{Get list of organization admins for logged-in user}
\usage{
users_list_me_organization_admins()
}
\value{
An array containing the following fields:
\item{id}{integer, The ID of this user.}
\item{name}{string, This user's name.}
\item{username}{string, This user's username.}
\item{initials}{string, This user's initials.}
\item{online}{boolean, Whether this user is online.}
\item{email}{string, This user's email address.}
}
\description{
Get list of organization admins for logged-in user
}
