% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{users_list_me_themes}
\alias{users_list_me_themes}
\title{List themes}
\usage{
users_list_me_themes()
}
\value{
An array containing the following fields:
\item{id}{integer, The ID of this theme.}
\item{name}{string, The name of this theme.}
\item{createdAt}{string, }
\item{updatedAt}{string, }
}
\description{
List themes
}
