% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{users_list_me_ui}
\alias{users_list_me_ui}
\title{UI configuration for logged-in user}
\usage{
users_list_me_ui()
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this user.}
\item{navigationMenus}{list, Navigation menus visible to this user.}
\item{userMenus}{list, User profile menu items available to this user.}
\item{userType}{list, A list containing the following elements: 
\itemize{
\item vendor boolean, True if this user is a member of any groups with a vendor ID.
\item media boolean, True if user has access to the Media Optimizer job type.
\item mainApp string, The slug for the main app for an app-only user account.
\item appCount integer, Number of apps this user has access to.
\item reportsOnly boolean, True if user is a reports-only user.
\item reportsCreator boolean, True if this user is allowed to create HTML reports.
}}
\item{zendeskToken}{string, JSON web token for this user's Zendesk widget.}
}
\description{
UI configuration for logged-in user
}
