% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{users_patch_me_favorites_ranking_higher}
\alias{users_patch_me_favorites_ranking_higher}
\title{Move a favorite one position closer to the top of the list}
\usage{
users_patch_me_favorites_ranking_higher(id)
}
\arguments{
\item{id}{integer required. The id of the favorite.}
}
\value{
An empty HTTP response
}
\description{
Move a favorite one position closer to the top of the list
}
