% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{users_post_access_email}
\alias{users_post_access_email}
\title{Sends the target user a 'Reset Password' or 'Welcome to Platform' email depending on the their status - Only available to Org and Team Admins}
\usage{
users_post_access_email(id)
}
\arguments{
\item{id}{integer required. The ID of this user.}
}
\value{
An empty HTTP response
}
\description{
Sends the target user a 'Reset Password' or 'Welcome to Platform' email depending on the their status - Only available to Org and Team Admins
}
