% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{users_post_me_favorites}
\alias{users_post_me_favorites}
\title{Favorite an item}
\usage{
users_post_me_favorites(object_id, object_type)
}
\arguments{
\item{object_id}{integer required. The id of the object. If specified as a query parameter, must also specify object_type parameter.}

\item{object_type}{string required. The type of the object that is favorited. Valid options: Project}
}
\value{
A list containing the following elements:
\item{id}{integer, The id of the favorite.}
\item{objectId}{integer, The id of the object. If specified as a query parameter, must also specify object_type parameter.}
\item{objectType}{string, The type of the object that is favorited. Valid options: Project}
\item{objectName}{string, The name of the object that is favorited.}
\item{createdAt}{string, The time this favorite was created.}
}
\description{
Favorite an item
}
