% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{users_post_unsuspend}
\alias{users_post_unsuspend}
\title{Unsuspends user}
\usage{
users_post_unsuspend(id)
}
\arguments{
\item{id}{integer required. The ID of this user.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this user.}
\item{user}{string, The username of this user.}
\item{unlockedAt}{string, The time the user's account was unsuspended}
}
\description{
Unsuspends user
}
