% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{with_tempfile}
\alias{with_tempfile}
\title{Call a function with a temporary file.}
\usage{
with_tempfile(fn, ...)
}
\arguments{
\item{fn}{a function that takes a filename as the first argument.}

\item{...}{further args passed to fn.}
}
\value{
object the return value of fn
}
\description{
Call a function with a temporary file.
}
\examples{
\dontrun{
data(iris)
with_tempfile(function(file_name) {
 write.csv(iris, file_name)
 # add'l operations
})
}

}
\keyword{internal}
