% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{workflows_delete_projects}
\alias{workflows_delete_projects}
\title{Remove a Workflow from a project}
\usage{
workflows_delete_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the Workflow.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Remove a Workflow from a project
}
