% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_chunks}
\alias{write_chunks}
\title{Split a file into chunks of a given chunk size, returning a list of file names.}
\usage{
write_chunks(file, chunk_size)
}
\arguments{
\item{file}{name of the file}

\item{chunk_size}{size of the chunk in bytes.}
}
\description{
Split a file into chunks of a given chunk size, returning a list of file names.
}
