% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{write_job_output}
\alias{write_job_output}
\title{Add a file as a run output if called from a container job}
\usage{
write_job_output(filename)
}
\arguments{
\item{filename}{string, name of the file to add as a run output
\code{civis::scripts_post_*_runs_outputs}.}
}
\value{
Returns the filename if not running on platform.
}
\description{
Add a file as a run output if called from a container job
}
\details{
Only posts if running on Civis Platform.
}
