% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{apps_delete_releases_shares_users}
\alias{apps_delete_releases_shares_users}
\title{Revoke the permissions a user has on this object}
\usage{
apps_delete_releases_shares_users(id, user_id, slug)
}
\arguments{
\item{id}{integer required. The unique id of the release.}

\item{user_id}{integer required. ID of the user}

\item{slug}{string required. The slug for the application.}
}
\value{
An empty HTTP response
}
\description{
Revoke the permissions a user has on this object
}
