% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_future.R
\name{civis_platform}
\alias{civis_platform}
\title{Evaluate an expression in Civis Platform}
\usage{
civis_platform(...)
}
\arguments{
\item{...}{Arguments to \code{\link{CivisFuture}} and then \code{\link{scripts_post_containers}}}
}
\value{
The result of evaluating \code{expr}.
}
\description{
This is used as with the \code{\link{future}} API as an argument to \code{\link{plan}}.
}
\examples{
\dontrun{
 plan(civis_platform)
 fut <- future({2 + 2}, required_resources = list(cpu = 1024, memory = 2048))
 value(fut)
}

}
