% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{clusters_get_kubernetes}
\alias{clusters_get_kubernetes}
\title{Describe a Kubernetes Cluster}
\usage{
clusters_get_kubernetes(id)
}
\arguments{
\item{id}{integer required. The ID of this cluster.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this cluster.}
\item{instanceType}{string, The EC2 instance types in this cluster.}
\item{minInstances}{integer, The minimum number of instances in this cluster.}
\item{maxInstances}{integer, The maximum number of instances in this cluster.}
\item{instanceMaxMemory}{integer, The amount of memory available to a single instance.}
\item{instanceMaxCpu}{integer, The number of processor shares available to a single instance.}
\item{organizationId}{string, The id of this cluster's organization.}
\item{organizationSlug}{string, The slug of this cluster's organization.}
}
\description{
Describe a Kubernetes Cluster
}
