% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{clusters_get_kubernetes}
\alias{clusters_get_kubernetes}
\title{Describe a Kubernetes Cluster}
\usage{
clusters_get_kubernetes(id, include_usage_stats = NULL)
}
\arguments{
\item{id}{integer required.}

\item{include_usage_stats}{boolean optional. When true, usage stats are returned in instance config objects. Defaults to false.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this cluster.}
\item{organizationId}{string, The id of this cluster's organization.}
\item{organizationName}{string, The name of this cluster's organization.}
\item{organizationSlug}{string, The slug of this cluster's organization.}
\item{clusterPartitions}{array, An array containing the following fields: 
\itemize{
\item clusterPartitionId integer, The ID of this cluster partition.
\item name string, The name of the cluster partition.
\item labels array, Labels associated with this partition.
\item instanceConfigs array, The instances configured for this cluster partition.
\item defaultInstanceConfigId integer, The id of the InstanceConfig that is the default for this partition.
}}
\item{isNatEnabled}{boolean, Whether this cluster needs a NAT gateway or not.}
\item{hours}{number, The number of hours used this month for this cluster }
}
\description{
Describe a Kubernetes Cluster
}
