% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_ml_utils.R
\name{coef.civis_ml}
\alias{coef.civis_ml}
\title{Function to get coefficients from a civis_ml_classifier}
\usage{
\method{coef}{civis_ml}(object, complete = TRUE, ...)
}
\arguments{
\item{object}{civis_ml_classifier}

\item{complete}{see documentation for generic \code{coef()}}

\item{...}{other arguments}
}
\value{
coefs a vector of coefficients or `NULL` if the classifier does not have coefficients
}
\description{
Function to get coefficients from a civis_ml_classifier
}
\details{
Outputs coefficients with names in the style of `stats::coef`
}
