% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{databases_list}
\alias{databases_list}
\title{List databases}
\usage{
databases_list()
}
\value{
An array containing the following fields:
\item{id}{integer, The ID for the database.}
\item{name}{string, The name of the database.}
\item{adapter}{string, The type of the database.}
}
\description{
List databases
}
