% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{databases_list_schemas}
\alias{databases_list_schemas}
\title{List schemas in this database}
\usage{
databases_list_schemas(id, name = NULL, credential_id = NULL)
}
\arguments{
\item{id}{integer required. The ID of the database.}

\item{name}{string optional. If specified, will be used to filter the schemas returned.Substring matching is supported (e.g., "name=schema" will return both "schema1" and "schema2"). Does not apply to BigQuery databases.}

\item{credential_id}{integer optional. If provided, schemas will be filtered based on the given credential.}
}
\value{
An array containing the following fields:
\item{schema}{string, The name of a schema.}
}
\description{
List schemas in this database
}
