% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{enhancements_put_cass_ncoa_projects}
\alias{enhancements_put_cass_ncoa_projects}
\title{Add a JobTypes::CassNcoa to a project}
\usage{
enhancements_put_cass_ncoa_projects(id, project_id)
}
\arguments{
\item{id}{integer required. ID of the resource}

\item{project_id}{integer required. The ID of the project}
}
\value{
An empty HTTP response
}
\description{
Add a JobTypes::CassNcoa to a project
}
