% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_client.R
\name{fetch_and_generate_client}
\alias{fetch_and_generate_client}
\title{Fetches and generates the client in generated_client.R}
\usage{
fetch_and_generate_client()
}
\description{
Fetches and generates the client in generated_client.R
}
\details{
Skips autogeneration on windows with R < 3.4.0 and if R_CLIENT_DEV == "TRUE".
Skips downloading a fresh spec if CIVIS_API_KEY is not set.
}
