% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_client.R
\name{fetch_and_generate_client}
\alias{fetch_and_generate_client}
\title{Fetches and generates the client.}
\usage{
fetch_and_generate_client()
}
\description{
Fetches and generates the client.
}
\details{
Skips autogeneration if:

\itemize{
\item 1. If the environment variable R_CLIENT_DEV == "TRUE"
\item 2. If CIVIS_API_KEY not present
\item 3. On windows with R < 3.4.0.
}
}
