% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civis_ml.R, R/logs.R
\name{fetch_logs.civis_ml}
\alias{fetch_logs.civis_ml}
\alias{fetch_logs}
\title{Fetch job logs from the Civis Platform}
\usage{
\method{fetch_logs}{civis_ml}(object, limit = 100, ...)

fetch_logs(object, limit = 100, ...)
}
\arguments{
\item{object}{An api response, CivisML model, or other object from the Civis
Platform.}

\item{limit}{The number of log lines to fetch, the default is 100.}

\item{\dots}{Extra parameters passed to methods, currently unused.}
}
\value{
A list of log messages with class \code{civis_logs}.
}
\description{
Fetch job logs from the Civis Platform
}
\examples{
\dontrun{
m <- civis_ml_fetch_existing(123)
fetch_logs(m)
}

}
