% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_client.R
\name{generate_client}
\alias{generate_client}
\title{Generate a client}
\usage{
generate_client(spec)
}
\arguments{
\item{spec}{usually from \code{get_spec}}
}
\value{
A string containing one documented function for each verb at each endpoint.
}
\description{
Generate a client
}
