% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{groups_patch}
\alias{groups_patch}
\title{Update some attributes of this Group}
\usage{
groups_patch(id, name = NULL, description = NULL)
}
\arguments{
\item{id}{integer required. The ID of this group.}

\item{name}{string optional. This group's name.}

\item{description}{string optional. The description of the group.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this group.}
\item{name}{string, This group's name.}
\item{createdAt}{string, The date and time when this group was created.}
\item{description}{string, The description of the group.}
\item{slug}{string, The slug for this group.}
\item{organizationId}{integer, The organization associated with this group.}
\item{members}{array, An array containing the following fields: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
}
\description{
Update some attributes of this Group
}
