% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{match_targets_list}
\alias{match_targets_list}
\title{List match targets}
\usage{
match_targets_list()
}
\value{
An array containing the following fields:
\item{id}{integer, The ID of the match target}
\item{name}{string, The name of the match target}
\item{table}{string, The name of the DynamoDB table}
\item{targetFileName}{string, The name of the target file}
\item{encryptionKeyCredentialId}{integer, The ID of the encryption key credential.}
\item{createdAt}{string, }
\item{updatedAt}{string, }
\item{archived}{boolean, Whether the match target has been archived.}
}
\description{
List match targets
}
