% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{match_targets_patch}
\alias{match_targets_patch}
\title{Update a match target}
\usage{
match_targets_patch(id, name = NULL, table = NULL, region = NULL)
}
\arguments{
\item{id}{integer required. The ID of the match target}

\item{name}{string optional. The name of the match target}

\item{table}{string optional. The name of the DynamoDB table}

\item{region}{string optional. The AWS region of the DynamoDB table. By default, us-east-1.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the match target}
\item{name}{string, The name of the match target}
\item{table}{string, The name of the DynamoDB table}
\item{region}{string, The AWS region of the DynamoDB table. By default, us-east-1.}
}
\description{
Update a match target
}
