% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{match_targets_patch}
\alias{match_targets_patch}
\title{Update a match target}
\usage{
match_targets_patch(id, name = NULL, target_file_name = NULL, archived = NULL)
}
\arguments{
\item{id}{integer required. The ID of the match target}

\item{name}{string optional. The name of the match target}

\item{target_file_name}{string optional. The name of the target file}

\item{archived}{boolean optional. Whether the match target has been archived.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the match target}
\item{name}{string, The name of the match target}
\item{targetFileName}{string, The name of the target file}
\item{createdAt}{string, }
\item{updatedAt}{string, }
\item{archived}{boolean, Whether the match target has been archived.}
}
\description{
Update a match target
}
