% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{notebooks_delete_deployments}
\alias{notebooks_delete_deployments}
\title{Delete a Notebook deployment}
\usage{
notebooks_delete_deployments(deployment_id, notebook_id)
}
\arguments{
\item{deployment_id}{integer required. The ID for this deployment}

\item{notebook_id}{integer required. The ID of the owning Notebook}
}
\value{
An empty HTTP response
}
\description{
Delete a Notebook deployment
}
