% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{predictions_list_schedules}
\alias{predictions_list_schedules}
\title{Show the prediction schedule}
\usage{
predictions_list_schedules(id)
}
\arguments{
\item{id}{integer required. ID of the prediction associated with this schedule.}
}
\value{
A list containing the following elements:
\item{id}{integer, ID of the prediction associated with this schedule.}
\item{schedule}{list, A list containing the following elements: 
\itemize{
\item scheduled boolean, If the item is scheduled.
\item scheduledDays array, Day based on numeric value starting at 0 for Sunday.
\item scheduledHours array, Hours of the day it is scheduled on.
\item scheduledMinutes array, Minutes of the day it is scheduled on.
\item scheduledRunsPerHour integer, Alternative to scheduled minutes, number of times to run per hour.
}}
\item{scoreOnModelBuild}{boolean, Whether the prediction will run after a rebuild of the associated model.}
}
\description{
Show the prediction schedule
}
