% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{publish_addin}
\alias{publish_addin}
\title{Add in to Publish an R Markdown file to Platform Reports from RStudio}
\usage{
publish_addin()
}
\description{
This function is called when the "Publish To Civis" button is pressed in
the RStudio IDE AddIn menu. It will attempt to save the current changes
of the active document and push it to the appropriate location in Platform.
RMarkdown files will be rendered to html with \code{\link{publish_rmd}}
and published as a report. Currently only RMarkdown files are supported.
}
\seealso{
\code{\link{publish_rmd}} for publishing RMarkdown files
}
