% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_civis}
\alias{read_civis}
\alias{read_civis.numeric}
\alias{read_civis.character}
\alias{read_civis.sql}
\alias{read_civis.civis_script}
\title{Read tables and files from Civis Platform}
\usage{
read_civis(x, ...)

\method{read_civis}{numeric}(x, using = readRDS, verbose = FALSE, ...)

\method{read_civis}{character}(x, database = NULL, ...)

\method{read_civis}{sql}(x, database = NULL, using = utils::read.csv,
  job_name = NULL, hidden = TRUE, verbose = FALSE, ...)

\method{read_civis}{civis_script}(x, regex = NULL, using = NULL, ...)
}
\arguments{
\item{x}{\code{"schema.table"}, \code{sql("query")}, or a file id.}

\item{...}{arguments passed to \code{using}.}

\item{using}{function, Function to convert the file to a data frame or to unserialize.
the file (e.g. \code{read.csv} or \code{readRDS}).}

\item{verbose}{bool, Set to TRUE to print intermediate progress indicators.}

\item{database}{string, Name of database where data frame is to be uploaded.
If no database is specified, uses \code{options(civis.default_db)}.}

\item{job_name}{string, Name of the job (default: \code{"Civis Export Via R Client"}).}

\item{hidden}{bool, Whether the job is hidden.}

\item{regex}{Regex of matching run output names.}
}
\description{
\code{read_civis} loads a table from Redshift as a data frame if
given a \code{"schema.table"} or \code{sql("query")} as the first argument. A
file from Amazon S3 (the files endpoint) if a file id is given. Run output ids
or values from any Civis platform script
are returned if a \code{\link{civis_script}} is given.

A default database can be set using \code{options(civis.default_db = "my_database")}.
If there is only one database available,
this database will automatically be used as the default.
}
\details{
By default, \code{read_civis.numeric} assumes the file has been serialized using
\code{saveRDS}, as in \code{write_civis_file} and uses \code{using = readRDS} by default. For reading
an uncompressed text or csv from the files endpoint, set \code{using = read.csv} for example.

If \code{using = NULL}, \code{read_civis.civis_script} will return run output
file ids, including ids of JsonValues.
Otherwise all run outputs matching \code{regex} will be read into memory
with \code{using}. JsonValues are always returned as values if \code{using} is not \code{NULL}.
Results are always a list. If the script has no outputs, the results are a list of length is 0.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: Return a file as a data frame

\item \code{character}: Return all columns from a table as a data frame.

\item \code{sql}: Return a SQL query as a data frame.

\item \code{civis_script}: Return run output values or file ids of a \code{civis_script} as a list.
}}

\examples{
\dontrun{
# Read all columns in a single table
df <- read_civis("schema.my_table", database = "my_database")

# Read data from a SQL select statement
query <- sql("SELECT * FROM table JOIN other_table USING id WHERE var1 < 23")
df <- read_civis(query, database = "my_database")

# Read an R object from the files endpoint.
id <- write_civis_file(df)
df <- read_civis(id)

# Read a text file or csv from the files endpoint.
id <- write_civis_file("my_csv.csv")
df <- read_civis(id, using = read.csv)

# Read and download all run outputs from a civis script.
ids <- read_civis(civis_script(1234))
mapply(download_civis, ids, names(ids))

# Read run output objects into memory
df <- read_civis(civis_script(1234), regex = '.csv', using = read.csv)
obj <- read_civis(civis_script(1234), regex = '.rds', using = readRDS)

# Gracefully handle when read_civis.sql returns no rows
query <- sql("SELECT * FROM table WHERE 1 = 2")
mean_x <- tryCatch({
  df <- read_civis(query, database = "my_database")
  mean(df$x)
}, empty_result_error = function(e) {
   NA
})
}
}
\seealso{
Other io: \code{\link{download_civis}},
  \code{\link{query_civis_file}},
  \code{\link{query_civis}},
  \code{\link{write_civis_file}}, \code{\link{write_civis}}
}
\concept{io}
