% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{remote_hosts_list}
\alias{remote_hosts_list}
\title{List the remote hosts}
\usage{
remote_hosts_list(type = NULL)
}
\arguments{
\item{type}{string optional. The type of remote host. One of:}
}
\value{
An array containing the following fields:
\item{id}{integer, The ID of the remote host.}
\item{name}{string, The name of the remote host.}
\item{type}{string, The type of remote host. One of: }
\item{url}{string, The URL for remote host.}
}
\description{
List the remote hosts
}
