% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{remote_hosts_post}
\alias{remote_hosts_post}
\title{Create a new remote host}
\usage{
remote_hosts_post(name, url, type)
}
\arguments{
\item{name}{string required. The human readable name for the remote host.}

\item{url}{string required. The URL to your host.}

\item{type}{string required. The type of remote host. One of: RemoteHostTypes::Bitbucket, RemoteHostTypes::GitSSH, RemoteHostTypes::Github, RemoteHostTypes::GoogleDoc, RemoteHostTypes::JDBC, RemoteHostTypes::Postgres, RemoteHostTypes::Redshift, RemoteHostTypes::S3Storage, and RemoteHostTypes::Salesforce}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of the remote host.}
\item{name}{string, The name of the remote host.}
\item{type}{string, The type of remote host. One of: RemoteHostTypes::Bitbucket, RemoteHostTypes::GitSSH, RemoteHostTypes::Github, RemoteHostTypes::GoogleDoc, RemoteHostTypes::JDBC, RemoteHostTypes::Postgres, RemoteHostTypes::Redshift, RemoteHostTypes::S3Storage, and RemoteHostTypes::Salesforce}
\item{url}{string, The URL for remote host.}
}
\description{
Create a new remote host
}
