% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{reports_get_services}
\alias{reports_get_services}
\title{Show a single service report}
\usage{
reports_get_services(id)
}
\arguments{
\item{id}{integer required. The ID of this report.}
}
\value{
A list containing the following elements:
\item{id}{integer, The ID of this report.}
\item{name}{string, The name of the report.}
\item{user}{list, A list containing the following elements: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
\item{createdAt}{string, }
\item{updatedAt}{string, }
\item{myPermissionLevel}{string, Your permission level on the object. One of "read", "write", or "manage".}
\item{host}{string, The host for the service report}
\item{displayUrl}{string, The URL to display the service report.}
\item{serviceId}{integer, The id of the backing service}
\item{provideAPIKey}{boolean, Whether the report requests an API Key from the report viewer.}
\item{apiKey}{string, A Civis API key that can be used by this report.}
\item{apiKeyId}{integer, The ID of the API key. Can be used for auditing API use by this report.}
\item{archived}{string, The archival status of the requested item(s).}
}
\description{
Show a single service report
}
