% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_delete_containers_projects}
\alias{scripts_delete_containers_projects}
\title{Remove a container docker from a project}
\usage{
scripts_delete_containers_projects(id, project_id)
}
\arguments{
\item{id}{integer required. ID of the resource}

\item{project_id}{integer required. The ID of the project}
}
\value{
An empty HTTP response
}
\description{
Remove a container docker from a project
}
