% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{scripts_put_sql_projects}
\alias{scripts_put_sql_projects}
\title{Add a scripts to a project}
\usage{
scripts_put_sql_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the resource.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Add a scripts to a project
}
