% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{services_delete_deployments}
\alias{services_delete_deployments}
\title{Delete a ShinyApp deployment}
\usage{
services_delete_deployments(service_id, deployment_id)
}
\arguments{
\item{service_id}{integer required. The ID of the owning ShinyApp}

\item{deployment_id}{integer required. The ID for this deployment}
}
\value{
An empty HTTP response
}
\description{
Delete a ShinyApp deployment
}
