% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{services_list_tokens}
\alias{services_list_tokens}
\title{List tokens}
\usage{
services_list_tokens(id)
}
\arguments{
\item{id}{integer required. The ID of the service.}
}
\value{
An array containing the following fields:
\item{id}{integer, The ID of the token.}
\item{name}{string, The name of the token.}
\item{user}{list, A list containing the following elements: 
\itemize{
\item id integer, The ID of this user.
\item name string, This user's name.
\item username string, This user's username.
\item initials string, This user's initials.
\item online boolean, Whether this user is online.
}}
\item{machineToken}{boolean, If true, this token is not tied to a particular user.}
\item{expiresAt}{string, The date and time when the token expires.}
\item{createdAt}{string, The date and time when the token was created.}
}
\description{
List tokens
}
