% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{templates_patch_scripts}
\alias{templates_patch_scripts}
\title{Update some attributes of this Script Template}
\usage{
templates_patch_scripts(
  id,
  name = NULL,
  note = NULL,
  ui_report_id = NULL,
  archived = NULL
)
}
\arguments{
\item{id}{integer required.}

\item{name}{string optional. The name of the template.}

\item{note}{string optional. A note describing what this template is used for; custom scripts created off this template will display this description.}

\item{ui_report_id}{integer optional. The id of the report that this template uses.}

\item{archived}{boolean optional. Whether the template has been archived.}
}
\value{
A list containing the following elements:
\item{id}{integer, }
\item{public}{boolean, If the template is public or not.}
\item{scriptId}{integer, The id of the script that this template uses.}
\item{scriptType}{string, The type of the template's backing script (e.g SQL, Container, Python, R, JavaScript)}
\item{userContext}{string, The user context of the script that this template uses.}
\item{name}{string, The name of the template.}
\item{category}{string, The category of this template.}
\item{note}{string, A note describing what this template is used for; custom scripts created off this template will display this description.}
\item{createdAt}{string, }
\item{updatedAt}{string, }
\item{useCount}{integer, The number of uses of this template.}
\item{uiReportId}{integer, The id of the report that this template uses.}
\item{techReviewed}{boolean, Whether this template has been audited by Civis for security vulnerability and correctness.}
\item{archived}{boolean, Whether the template has been archived.}
\item{hidden}{boolean, The hidden status of the item.}
}
\description{
Update some attributes of this Script Template
}
