% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_client.R
\name{templates_put_scripts_projects}
\alias{templates_put_scripts_projects}
\title{Add a Script Template to a project}
\usage{
templates_put_scripts_projects(id, project_id)
}
\arguments{
\item{id}{integer required. The ID of the Script Template.}

\item{project_id}{integer required. The ID of the project.}
}
\value{
An empty HTTP response
}
\description{
Add a Script Template to a project
}
